Infections and Immunity

Overview Structs

GEMS.AgeDependentTransmissionRate

GEMS.AgeDependentTransmissionRate

GEMS.ConstantTransmissionRate

GEMS.TransmissionFunction

Overview Functions

GEMS.disease_progression!

GEMS.get_infections_between

GEMS.id

GEMS.infectionsDF

GEMS.name

GEMS.transmission_probability

Structs

GEMS.AgeDependentTransmissionRate — Type

AgeDependentTransmissionRate <: TransmissionFunction

A TransmissionFunction type that allows to define transmission probabilities for specific age groups.

GEMS.ConstantTransmissionRate — Type

ConstantTransmissionRate <: TransmissionFunction

A TransmissionFunction type that uses a constant transmission rate.

GEMS.TransmissionFunction — Type

TransmissionFunction

Abstract type for all transmission functions.

Constructors

GEMS.AgeDependentTransmissionRate — Method

AgeDependentTransmissionRate(;transmission_rate, ageGroups, ageTransmissions, distribution)

Constructor for the age-dependent transmission rate struct. The parameters for the constructor include:

transmission_rate::Vector: A vector containing the parameters from which the distribution of the transmission_rate is constructed.

ageGroups::Vector: A vector containing the age groups. Should contain a vector for each age group consisting of two integers.

ageTransmissions::Vector: A vector containing the parameters for the distributions of the transmission rates for the specific age groups. Should contain a vector for each age group consisting of as many real number as parameters required for the spec. distribution.

distribution::String: A string that corresponds to a distribution of the distribution package.

Functions

GEMS.disease_progression! — Method

disease_progression!(infectee::Individual, pathogen::Pathogen, exposedtick::Int16)

Assigns a disease progression to the infectee by assigning event ticks based on the exposedtick.

GEMS.get_infections_between — Method

get_infections_between(logger::InfectionLogger, infecter::Int32, start_tick::Int16, end_tick::Int16)

Returns the id of infected individuals who's infection time t is start_tick <= t <= end_tick)

Parameters

logger::InfectionLogger: Logger instance

infecter::Int32: ID of infecter individual that is used to filter secondary infections

start_tick::Int16: Lower bound (time)

end_tick::Int16: Upper bound (time)

Returns

Vector{Int32}: List of IDs of infected individuals

GEMS.id — Method

id(vaccine)

Returns the unique identifier of the vaccine.

GEMS.infectionsDF — Method

infectionsDF(postProcessor::PostProcessor)

Returns the internal flat infections DataFrame.

Columns

Name	Type	Description

infection_id	Int32	Unique identifier of an infection

tick	Int16	Tick of the infection event

id_a	Int32	Infecter id

id_b	Int32	Infectee id

infectious_tick	Int16	Tick at which infectee becomes infectious

removed_tick	Int16	Tick at which infectee becomes removed (recovers)

symptoms_tick	Int16	Tick at which infectee develops symptoms

severeness_tick	Int16	Tick at which infectee's symptoms become severe

hospital_tick	Int16	Tick at which infectee is hospitalized

icu_tick	Int16	Tick at which infectee is admitted to ICU

ventilation_tick	Int16	Tick at which infectee requires ventilation

symptom_category	Int8	Disease progression category (asymp., mild, severe, critical)

setting_id	Int32	Id of setting in which infection happens

setting_type	Char	Setting type of the infection setting

lat	Float32	Latitude of infection location

lon	Float32	Longitude of infection location

ags	Int32	German Community Identification Number of infection

source_infection_id	Int32	ID of the infection even that caused this infection (chain)

generation_time	Int16	Time between preceeding infection and this exposure

serial_interval	Int16	Time between onset of symptoms of this and preceeding infection

test_type	String	Type of test which detected this infection

first_detected_tick	Int16	Tick of (reportable) test that first detected this infection

sex_a	Int8	Infecter sex

age_a	Int8	Infecter age

number_of_vaccinations_a	Int8	Infecter number of previous vaccinations

vaccination_tick_a	Int16	Infecter last time of vaccination

education_a	Int8	Infecter education level

occupation_a	Int8	Infecter occupation group

household_a	Int32	Infecter associated household

office_a	Int32	Infecter associated office

schoolclass_a	Int32	Infecter associated school

sex_b	Int8	Infectee sex

age_b	Int8	Infectee age

number_of_vaccinations_a	Int8	Infecter number of previous vaccinations

vaccination_tick_a	Int16	Infecter last time of vaccination

education_b	Int8	Infectee education level

occupation_b	Int8	Infectee occupation group

household_b	Int32	Infectee associated household

office_b	Int32	Infectee associated office

schoolclass_b	Int32	Infectee associated schoolclass

household_ags_b	Int32	Infectee household German Community Identification Number

GEMS.name — Method

name(vaccine)

Returns the name of the vaccine.

GEMS.transmission_probability — Function

transmission_probability(transFunc::TransmissionFunction, infecter::Individual, infected::Individual, setting::Setting, tick::Int16)

General function for TransmissionFunction struct. Should be overwritten for newly created structs, as it only serves to catch undefined transmission_probability functions.

transmission_probability(transFunc::ConstantTransmissionRate, infecter::Individual, infected::Individual, setting::Setting, tick::Int16)

Calculates the transmission probability for the ConstantTransmissionRate. Returns the transmission_rate for all individuals who have not been infected in the past. If the individual has already recovered, the function returns 0.0, assuming full indefinite natural immunity.

Parameters

transFunc::ConstantTransmissionRate: Transmission function struct

infecter::Individual: Infecting individual

infected::Individual: Individual to infect

setting::Setting: Setting in which the infection happens

tick::Int16: Current tick

Returns

Float64: Transmission probability p (0 <= p <= 1)

transmission_probability(transFunc::AgeDependentTransmissionRate, infecter::Individual, infected::Individual, setting::Setting, tick::Int16)

Calculates the transmission probability for the AgeDependentTransmissionRate. Selects the correct distribution dependent on the age of the potentially infected agent from the AgeDependentTransmissionRate, draws from it and returns the value. If no age group is found for the individual the transmission rate is drawn from the transmission_rate distribution. If the individual has already recovered, the function returns 0.0, assuming full indefinite natural immunity.

Parameters

transFunc::AgeDependentTransmissionRate: Transmission function struct

infecter::Individual: Infecting individual

infected::Individual: Individual to infect

setting::Setting: Setting in which the infection happens

tick::Int16: Current tick

Returns

Float64: Transmission probability p (0 <= p <= 1)