Misc

everything else

Overview Structs

GEMS.AGS

GEMS.Entity

Overview Functions

GEMS.GEMS_version

GEMS._int

GEMS.aggregate_df

GEMS.aggregate_dfs

GEMS.aggregate_dfs_multcol

GEMS.county

GEMS.data

GEMS.district

GEMS.escape_markdown

GEMS.foldercount

GEMS.free_mem_size

GEMS.func_from_string

GEMS.generate_combinations

GEMS.germanshapes

GEMS.get_nested_value

GEMS.getundocumented

GEMS.git_branch

GEMS.git_commit

GEMS.git_repo

GEMS.id

GEMS.in_county

GEMS.in_district

GEMS.in_state

GEMS.is_county

GEMS.is_district

GEMS.is_state

GEMS.isunset

GEMS.kernel

GEMS.markdown

GEMS.municipality

GEMS.print_aggregates

GEMS.printinfo

GEMS.remove_kw

GEMS.state

GEMS.subinfo

Structs

GEMS.Entity — Type

Supertype for all simulated objects

Functions

GEMS.GEMS_version — Function

GEMS_version(rd::ResultData)

Returns the GEMS version this ResultData object was generated with. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS_version(bd::BatchData)

Returns the GEMS version this BatchData object was generated with.

GEMS._int — Function

_int(f::Function)

Wrapper to enforce Integer returns of anonymous functions.

Missing docstring.

Missing docstring for agegroup(::String, ::Individual). Check Documenter's build log for details.

Missing docstring.

Missing docstring for aggregate_data. Check Documenter's build log for details.

GEMS.aggregate_df — Function

aggregate_dfs(dataframe, key)

Groups dataframes (with numerical values) on the provided key column and applies the aggregate_values function to each of them. The resulting dataframe has all initial columns supplemented with the suffixes min, max, mean, lower_95, upper_95, and std

GEMS.aggregate_dfs — Function

aggregate_dfs(dataframes, key)

Joins the input vector of dataframes on the key and aggregates the residual data. Requires all dataframes to provide the exact same columns and column names.

GEMS.aggregate_dfs_multcol — Function

aggregate_dfs_multcol(dataframes, key)

Aggregates data on the columns of the dataframes contained in the provided vector for each value in the key column. All dataframes must have identical columnnames. Returns a dictionary with the columnnames as keys and a dataframe as the value.

Missing docstring.

Missing docstring for create_config_files(::Any, ::Any, ::Any). Check Documenter's build log for details.

GEMS.data — Method

data(runinfo)

Returns data Dict from a Runinfo object.

GEMS.escape_markdown — Method

escape_markdown(str::String)

Replaces markdown controls (e.g. "_" or "*") with their escaped version.

GEMS.foldercount — Method

foldercount(directory)

Returns the number of subfolders in a directory.

GEMS.free_mem_size — Function

free_mem_size(rd::ResultData)

Returns the available system memory

free_mem_size(bd::BatchData)

Returns the available system memory

GEMS.func_from_string — Method

func_from_string(fct_string::String)

Convert the provided string into a applicable function.

Arguments

fct_string::String: The function in string format.

Returns

Function: The parsed function.

GEMS.generate_combinations — Method

generate_combinations(parameters_dict)

Generate combinations of the provided parameter values. For each parameter a list of values is provided in the parameters_dict. The resulting parameter combinations are all permutations of these values.

Arguments:

parameters_dict::Dict: A dictionary containing parameter lists.

Returns:

Vector: A vector of dictionaries, where each dictionary contains the parameter keys as keys and the value as an entry.

GEMS.germanshapes — Method

germanshapes(level::Int64)

Returns the Shapefile.Table object read from the respective shapefile. Downloads the file if it is not already locally available. Lookup constants.jl to find remote location of the shapefile. The level argument defines wether the state- (level = 1), county- (level = 2) or municipality- (level = 3) shapes are returned.

Returns

Shapefile.Table{Union{Missing, Shapefile.Polygon}}: Shapefile data. Lookup Shapefile.jl package for documentation.

GEMS.get_nested_value — Function

get_nested_value(dict, key_path)

Retrieve a value from a nested dictionary using a dot-separated key path (key.key.key.key...).

Arguments

dict: The nested dictionary from which to retrieve the value.

key_path::AbstractString: A dot-separated string representing the path to the desired value in the nested dictionary.

Returns

The value found at the specified key path, or nothing if the path does not exist in the dictionary.

GEMS.getundocumented — Function

getundocumented()

Returns a dataframe of all Functions and/or structs without a docstring.

GEMS.git_branch — Function

git_branch(rd::ResultData)

Returns the current git branch.

git_branch(bd::BatchData)

Returns the current git branch.

GEMS.git_commit — Function

git_commit(rd::ResultData)

Returns the current git commit.

git_commit(bd::BatchData)

Returns the current git commit.

GEMS.git_repo — Function

git_repo(rd::ResultData)

Returns the current git repository.

git_repo(bd::BatchData)

Returns the current git repository.

Missing docstring.

Missing docstring for is_toml_file. Check Documenter's build log for details.

Missing docstring.

Missing docstring for isvalidDistribution. Check Documenter's build log for details.

Missing docstring.

Missing docstring for initialize_seed. Check Documenter's build log for details.

GEMS.kernel — Function

kernel(rd::ResultData)

Returns the system kernel information Returns an empty dictionary if the data is not available in the input ResultData object.

kernel(bd::BatchData)

Returns the system kernel information

GEMS.markdown — Function

markdown(infFrac::InfectedFraction)

Returns a markdown string with all parameters (mainly for documentation purposes).

markdown(tu::TimesUp)

Returns a markdown string with all parameters (mainly for documentation purposes).

markdown(pathogen::Pathogen)

Returns a markdown string with all parameters (mainly for documentation purposes).

markdown(unDist::Uniform)

Returns a markdown string with all parameters (mainly for documentation purposes).

markdown(poisDist::Poisson)

Returns a markdown string with all parameters (mainly for documentation purposes).

markdown(binDist::Binomial)

Returns a markdown string with all parameters (mainly for documentation purposes).

markdown(dist::Distribution)

Return a markdown string with all parameters (mainly for documentation purposes). Fallback function for arbitrary distributions without dedicated formatting.

markdown(stngs::SettingsContainer, sim::Simulation)

Returns a markdown string with all parameters (mainly for documentation purposes).

GEMS.print_aggregates — Function

print_aggregates(agg; unit, multiplier, digits)

Pretty-prints the outcomes of an aggregate_values() function call.

GEMS.printinfo — Method

printinfo(str::String)

Prints an @info-Text with the provided string and a HH:MM:SS timestamp.

Missing docstring.

Missing docstring for process_funcs(::Dict). Check Documenter's build log for details.

GEMS.remove_kw — Function

remove_kw(to_remove::Symbol, kwrds)

Removes a particular keyword from a Named Tuple keyword ist. This is mainly used to exclude certain keywords in plots when passing plotargs... to subplots.

Example

function my_function(;args...)
    new_kwrds = remove_kw(:a, args)
    my_other_function(new_kwrds...)
end

If you call the above function with my_function(a = 2, b = 3), it will call the inner function with my_other_function(b = 3).

Missing docstring.

Missing docstring for set_nested!. Check Documenter's build log for details.

GEMS.subinfo — Method

subinfo(str::String)

Returns a string with a "sub-line-hook" indicating subtasks of an @info task (as generated by printinfo())

AGS

Structs

GEMS.AGS — Type

AGS

Struct to represent Germany's Community Identification Number (Amtlicher Gemeindeschlüssel).

Taken from Wikipedia (https://en.wikipedia.org/wiki/Community_Identification_Number):

The municipality key consists of eight digits, which are generated as follows: The first two digits designate the individual German state. The third digit designates the government district (in areas without government districts a zero is used instead). The fourth and fifth digits designate the number of the urban area (in a district-free city) or the district (in a city with districts). In GEMS they are called 'counties'. The sixth, seventh, and eighth digits indicate the municipality or the number of the unincorporated area.

Example

münster = AGS("05515000")

The above code generates the community identification number for the city of Münster.

The first two digits identify the state of North Rhine-Westphalia (05)

The third digit identifies the government district Münster (5)

The fourth and fifth digit identify the district-free city of Münster (15)

Being a district-free city, Münster does not have a municipality code. Examples for municipalities are: Emsdetten (05566008), Warendorf (05570052), or Steinfurt (05566084).

There are multiple helper functions to navigate through the hierarchical AGS structure such as: state, district, county, or municiaplity which each return the AGS of the respective geographic level. The is_state, is_district, and is_county return true, if the provided AGS is on the respective geographic level. Moreover, in_state, in_district, and in_county verify whether a provided AGS is contained in a parent region.

AGS are directly comparable (==) and broadcastable (.==).

Note:

There is no is_municiaplity function as this could be confused as a check wether the AGS is on the lowest geographic level. However, as there are district-free cities which are technically municipalities but without the last three digits given, this check could only be done with context information on the actual region. Based purely on the number, we do not know which county has contains municipalites and which conty is a district-free city.

Functions

GEMS.county — Method

county(ags::AGS)

Returns the parent county's Community Identification Number (AGS) of the provided AGS.

GEMS.district — Method

district(ags::AGS)

Returns the parent district's Community Identification Number (AGS) of the provided AGS.

GEMS.id — Method

id(ags::AGS)

Retuns the AGSs internal number (int).

GEMS.in_county — Method

in_county(ags::AGS, parent::AGS)

Returns true if the county-section of both Community Identification Numbers (AGS) match. Both AGSs are in the same county.

GEMS.in_district — Method

in_district(ags::AGS, parent::AGS)

Returns true if the district-section of both Community Identification Numbers (AGS) match. Both AGSs are in the same district.

GEMS.in_state — Method

in_state(ags::AGS, parent::AGS)

Returns true if the state-section of both Community Identification Numbers (AGS) match. Both AGSs are in the same state.

GEMS.is_county — Method

is_district(ags::AGS)

Returns true if the Community Identification Number (AGS) belongs to a conty. First five digits given, last three being 0s.

GEMS.is_district — Method

is_district(ags::AGS)

Returns true if the Community Identification Number (AGS) belongs to a governmental district. First three digits given, everything else being 0s.

GEMS.is_state — Method

is_state(ags::AGS)

Returns true if the Community Identification Number (AGS) belongs to a state. First two digits given, everything else being 0s.

GEMS.isunset — Method

isunset(ags::AGS)

Returns true if the AGS is the "empty" default and does not actually belongs to a any geographic region.

GEMS.municipality — Method

municipality(ags::AGS)

Returns the Community Identification Number (AGS) regardless of its geographic level.

GEMS.state — Method

state(ags::AGS)

Returns the parent state's Community Identification Number (AGS) of the provided AGS.

to be assigned

Missing docstring.

Missing docstring for aggregation_interval_steps. Check Documenter's build log for details.

Missing docstring.

Missing docstring for find_alpha. Check Documenter's build log for details.

Missing docstring.

Missing docstring for focal_objects(::Handover). Check Documenter's build log for details.