Pathogens

Overview Structs

GEMS.DiseaseProgressionStrat

GEMS.DiseaseProgressionStrat

GEMS.Pathogen

Overview Functions

GEMS.critical_death_rate

GEMS.disease_progression_strat

GEMS.hospitalization_rate

GEMS.icu_rate

GEMS.id

GEMS.infection_rate

GEMS.infectious_offset

GEMS.length_of_stay

GEMS.mild_death_rate

GEMS.name

GEMS.onset_of_severeness

GEMS.onset_of_symptoms

GEMS.sample_critical_death_rate

GEMS.sample_hospitalization_rate

GEMS.sample_icu_rate

GEMS.sample_infectious_offset

GEMS.sample_length_of_stay

GEMS.sample_mild_death_rate

GEMS.sample_onset_of_severeness

GEMS.sample_onset_of_symptoms

GEMS.sample_self_quarantine_rate

GEMS.sample_severe_death_rate

GEMS.sample_time_to_hospitalization

GEMS.sample_time_to_icu

GEMS.sample_time_to_recovery

GEMS.sample_ventilation_rate

GEMS.severe_death_rate

GEMS.time_to_hospitalization

GEMS.time_to_icu

GEMS.time_to_recovery

GEMS.transmission_function

GEMS.transmission_function!

GEMS.ventilation_rate

Structs

GEMS.Pathogen — Type

Pathogen <: Parameter

A type representing a pathogen.

Instantiation

The instantiation requires at least an id and a name that must be supplied as keyword arguments. All other fields are optional parameters.

covid = Pathogen(id = 1, name = "COVID19")
flu = Pathogen(id = 2, name = "Flu", onset_of_symptoms = Poission(3))

Parameters

id::Int8: Unique identifier pathogen

name::String: Name of the pathogen

dpr::DiseaseProgressionStrat = DiseaseProgressionStrat() (optional): Defines the distribution of symptom categories across age groups

infection_rate::Distribution = Uniform(0,1) (optional): Distribution of the infection rate

mild_death_rate::Distribution = Uniform(0,0.005) (optional): Distribution of the death rate for mild cases

severe_death_rate::Distribution = Uniform(0,0.1) (optional): Distribution of the death rate for severe cases

critical_death_rate::Distribution = Uniform(0,0.1) (optional): Distribution of the death rate for critical cases

hospitalization_rate::Distribution = Uniform(0, 0.1) (optional): Distribution of the probability for a severe case to be hospitalized

ventilation_rate::Distribution = Uniform(0, 0.1) (optional): Distribution of the probability for a critical case to need ventilation

icu_rate::Distribution = Uniform(0, 0.1) (optional): Distribution of the probability for a critical case to need ICU

onset_of_symptoms::Distribution = Uniform(2,3) (optional): Distribution of time till onset of symptoms from becoming exposed

onset_of_severeness::Distribution = Uniform(2,3) (optional): Distribution of time till onset of severeness from onset of symptoms

infectious_offset::Distribution = Uniform(0,1) (optional): Distribution of the offset to become infectious compared to the onset of symptoms

time_to_hospitalization::Distribution = Uniform(0,1) (optional): Distribution of time till hospitalization from onset of symptoms

time_to_icu::Distribution = Uniform(0,1) (optional): Distribution of time till ICU from hospitalization

time_to_recovery::Distribution = Uniform(5,6) (optional): Distribution of time till recovery from becoming exposed

length_of_stay::Distribution = Uniform(6,7) (optional): Distribution of time duration to stay in hospital

GEMS.DiseaseProgressionStrat — Type

DiseaseProgressionStrat <: AgeStratification

A wrapper type for an age stratification matrix regarding the disease progression.

Fields

age_groups::Vector{String}: A list of the age groups in order of the rows of the stratification matrix. The first entry in age_groups belongs to the first row in the matrix.

disease_compartments::Vector{String}: A list of the compartments/terminal states in the disease progression in order of the columns of the stratification matrix. First entry in disease_compartments belongs to the first column in the stratification matrix.

stratification_matrix::Vector{Vector{T}} where T <: Real: The matrix that holds the probabilities for the age stratified disease progression. The entries of each row have to add up to 1.

Functions

GEMS.critical_death_rate — Method

critical_death_rate(pathogen::Pathogen)

Returns the death rate (distribution) of the pathogen in the case of critical progression.

GEMS.DiseaseProgressionStrat — Method

DiseaseProgressionStrat()

A constructor for a default Disease Progression, where every individual will only be assigned to be asymptomatic. Mainly for testing purposes.

GEMS.disease_progression_strat — Method

disease_progression_strat(pathogen::Pathogen)

Returns the stratified probabilities for disease progression.

GEMS.hospitalization_rate — Method

hospitalization_rate(pathogen::Pathogen)

Returns the hospitalization rate (distribution) of the pathogen in the case of a severe disease progression.

GEMS.icu_rate — Method

icu_rate(pathogen::Pathogen)

Returns the icu rate (distribution) of the pathogen in the case of a critical disease progression.

GEMS.id — Method

id(pathogen::Pathogen)

Returns the id of the pathogen.

GEMS.infection_rate — Method

infection_rate(pathogen::Pathogen)

Returns the infection rate (distribution) of the pathogen.

GEMS.infectious_offset — Method

infectious_offset(pathogen::Pathogen)

Returns the time distribution for the offset of becoming infectious, before developing symptoms.

GEMS.length_of_stay — Method

length_of_stay(pathogen::Pathogen)

Returns the time distribution of the duration an individual will be hospitalized.

GEMS.mild_death_rate — Method

mild_death_rate(pathogen::Pathogen)

Returns the death rate (distribution) of the pathogen in the case of mild symptoms.

GEMS.name — Method

name(pathogen::Pathogen)

Returns the name of the pathogen.

GEMS.onset_of_severeness — Method

onset_of_severeness(pathogen::Pathogen)

Returns the time distribution for the onset of severe symptoms.

GEMS.onset_of_symptoms — Method

onset_of_symptoms(pathogen::Pathogen)

Returns the time distribution for the onset of symptoms.

GEMS.sample_critical_death_rate — Method

sample_critical_death_rate(pathogen::Pathogen, indiv::Individual)

Returns a randomly drawn value from the death rate distribution for critical disease progressions.

GEMS.sample_hospitalization_rate — Method

sample_hospitalization_rate(pathogen::Pathogen, indiv::Individual)

Returns a randomly drawn value from the hospitalization rate distribution.

GEMS.sample_icu_rate — Method

sample_icu_rate(pathogen::Pathogen, indiv::Individual)

Returns a randomly drawn value from the ICU rate distribution.

GEMS.sample_infectious_offset — Method

sample_infectious_offset(pathogen::Pathogen,  indiv::Individual)

Returns a randomly drawn (rounded) value from the infectious offset distribution.

GEMS.sample_length_of_stay — Method

sample_length_of_stay(pathogen::Pathogen,  indiv::Individual)

Returns a randomly drawn (rounded) value from the time_to_hopsitalization distribution for cases with severe disease progression.

GEMS.sample_mild_death_rate — Method

sample_mild_death_rate(pathogen::Pathogen, indiv::Individual)

Returns a randomly drawn value from the death rate distribution for cases with mild symptoms.

GEMS.sample_onset_of_severeness — Method

sample_onset_of_severeness(pathogen::Pathogen,  indiv::Individual)

Returns a randomly drawn (rounded) value from the onset_of_symptoms distribution.

GEMS.sample_onset_of_symptoms — Method

sample_onset_of_symptoms(pathogen::Pathogen,  indiv::Individual)

Returns a randomly drawn (rounded) value from the onset_of_symptoms distribution.

GEMS.sample_self_quarantine_rate — Method

sample_self_quarantine_rate(pathogen::Pathogen, indiv::Individual)

Returns a randomly drawn value from the ICU rate distribution.

GEMS.sample_severe_death_rate — Method

sample_severe_death_rate(pathogen::Pathogen, indiv::Individual)

Returns a randomly drawn value from the death rate distribution for severe disease progressions.

GEMS.sample_time_to_hospitalization — Method

sample_time_to_hospitalization(pathogen::Pathogen,  indiv::Individual)

Returns a randomly drawn (rounded) value from the time_to_hospitalization distribution for cases with severe disease progression.

GEMS.sample_time_to_icu — Method

sample_time_to_icu(pathogen::Pathogen,  indiv::Individual)

Returns a randomly drawn (rounded) value from the time_to_icu distribution for cases with critical disease progression.

GEMS.sample_time_to_recovery — Method

sample_time_to_recovery(pathogen::Pathogen,  indiv::Individual)

Returns a randomly drawn (rounded) value from the time_to_recovery distribution.

GEMS.sample_ventilation_rate — Method

sample_ventilation_rate(pathogen::Pathogen, indiv::Individual)

Returns a randomly drawn value from the Ventilation rate distribution.

GEMS.severe_death_rate — Method

severe_death_rate(pathogen::Pathogen)

Returns the death rate (distribution) of the pathogen in the case of severe symptoms.

GEMS.time_to_hospitalization — Method

time_to_hospitalization(pathogen::Pathogen)

Returns the distribution for the time till hospitalization starting from the onset of symptoms.

GEMS.time_to_icu — Method

time_to_icu(pathogen::Pathogen)

Returns the distribution for the time till ICU starting from the hospitalization time.

GEMS.time_to_recovery — Method

time_to_recovery(pathogen::Pathogen)

Returns the time to recovery (distribution) of the pathogen.

GEMS.transmission_function! — Method

transmission_function!(pathogen::Pathogen, transFunc::TransmissionFunction)

Returns the transmission_function for the pathogen.

GEMS.transmission_function — Method

transmission_function(pathogen::Pathogen)

Returns the transmission_function for the pathogen.

GEMS.ventilation_rate — Method

ventilation_rate(pathogen::Pathogen)

Returns the ventilation rate (distribution) of the pathogen in the case of a critical disease progression.