Plotting

Overview Structs

GEMS.ActiveDarkFigure

GEMS.AggregatedSettingAgeContacts

GEMS.BatchEffectiveReproduction

GEMS.BatchPlot

GEMS.BatchPopulationPyramid

GEMS.BatchSettingAgeContacts

GEMS.BatchTickCases

GEMS.BatchTickIsolations

GEMS.BatchTickTests

GEMS.CompartmentFill

GEMS.CumulativeCases

GEMS.CumulativeDiseaseProgressions

GEMS.CumulativeIsolations

GEMS.CustomLoggerPlot

GEMS.DetectedCases

GEMS.EffectiveReproduction

GEMS.GMTWrapper

GEMS.GenerationTime

GEMS.HospitalOccupancy

GEMS.HouseholdAttackRate

GEMS.Incidence

GEMS.InfectionDuration

GEMS.InfectiousHistogram

GEMS.LatencyHistogram

GEMS.ObservedReproduction

GEMS.ObservedSerialInterval

GEMS.PopulationPyramid

GEMS.SettingAgeContacts

GEMS.SettingSizeDistribution

GEMS.SimulationPlot

GEMS.SymptomCategories

GEMS.TestPositiveRate

GEMS.TickCases

GEMS.TickCasesBySetting

GEMS.TickTests

GEMS.TimeToDetection

GEMS.TotalTests

Overview Functions

GEMS.emptyplot

GEMS.gemsplot

GEMS.plot_ageGroup_contact_distribution

GEMS.plot_age_contact_distribution

GEMS.plot_all_age_contact_distributions

GEMS.plot_multiple_ageGroup_contact_distributions

GEMS.plot_multiple_age_contact_distributions

GEMS.plotseries!

GEMS.saveplot

GEMS.splitlabel

GEMS.splitplot

Structs

GEMS.ActiveDarkFigure — Type

ActiveDarkFigure <: SimulationPlot

A simulation plot type for generating an active-darkfigure-per-tick plot.

GEMS.AggregatedSettingAgeContacts — Type

AggregatedSettingAgeContacts <: SimulationPlot

A simulation plot type for sampling contacts from the model and build an age group x age group matrix for a given Setting type. The plot displays aggregated age groups and their mean number of contacts.

GEMS.BatchEffectiveReproduction — Type

BatchEffectiveReproduction <: BatchPlot

A batch plot type for generating a the effective-reproduction-number-per-tick plot. It includes values per tick, min-max-ranges as well as the 95% confidence band.

GEMS.BatchPlot — Type

Supertype for all plots that go into batch-run simulation reports

GEMS.BatchPopulationPyramid — Type

BatchPopulationPyramid <: BatchPlot

A batch plot type for generating a population pyramid for a specified population file. It has to be passed to the constructor via a string identifier.

GEMS.BatchSettingAgeContacts — Type

BatchSettingAgeContacts <: BatchPlot

A batch plot type for sampling contacts from the model and build an age x age matrix for a given Setting type and population file.

GEMS.BatchTickCases — Type

BatchTickCases <: BatchPlot

A batch plot type for generating a new-cases-per-tick plot. It includes mean cases per tick, min-max-ranges as well as the 95% confidence band.

GEMS.BatchTickIsolations — Type

BatchTickIsolations <: BatchPlot

A batch plot type for generating a current-isolations-per-tick plot. It includes mean number of isolated individuals per tick, min-max-ranges as well as the 95% confidence band.

GEMS.BatchTickTests — Type

BatchTickTests <: BatchPlot

A batch plot type for generating a tests-per-tick plot. It includes mean tests per tick, min-max-ranges as well as the 95% confidence band.

GEMS.CompartmentFill — Type

CompartmentFill <: SimulationPlot

A simulation plot type for generating a cumulative infections plot.

GEMS.CumulativeCases — Type

CumulativeCases <: SimulationPlot

A simulation plot type for generating a cumulative infections plot.

GEMS.CumulativeDiseaseProgressions — Type

CumulativeDiseaseProgressions <: SimulationPlot

A simulation plot type for generating a stacked bar chart on the cumulative disease progression.

GEMS.CumulativeIsolations — Type

CumulativeIsolations <: SimulationPlot

A simulation plot type for generating a plot displaying the cumulative number of individuals currently in isolation.

GEMS.CustomLoggerPlot — Type

CustomLoggerPlot <: SimulationPlot

A simulation plot type for generating a time series plot displaying values stored in a ResultDatas CustomLogger dataframe.

GEMS.DetectedCases — Type

DetectedCases <: SimulationPlot

A simulation plot type for generating a new-DETECTED-cases-per-tick plot.

GEMS.EffectiveReproduction — Type

EffectiveReproduction <: SimulationPlot

A simulation plot type for generating an effective reproduction number plot.

GEMS.GMTWrapper — Type

GMTWrapper

Wrapper-struct to inform saveplot function about the existance of a GMT-generated plot in the temp-folder (path_to_map).

GEMS.GenerationTime — Type

GenerationTime <: SimulationPlot

A simulation plot type for generating a generation-time-per-tick.

GEMS.HospitalOccupancy — Type

HospitalOccupancy <: SimulationPlot

A simulation plot type for generating a plot with hospitalization numbers etc.

GEMS.HouseholdAttackRate — Type

HouseholdAttackRate <: SimulationPlot

A simulation plot type for generating a household-attack-rate plot.

GEMS.Incidence — Type

Incidence <: SimulationPlot

A simulation plot type for generating an incidence plot.

GEMS.InfectionDuration — Type

InfectionDuration <: SimulationPlot

A simulation plot type for visualizing the distribution of infection durations as a histogram.

GEMS.InfectiousHistogram — Type

InfectiousHistogram <: SimulationPlot

A simulation plot type for generating the distribution of infectious period lengths.

GEMS.LatencyHistogram — Type

LatencyHistogram <: SimulationPlot

A simulation plot type for generating the distribution of latency period lengths.

GEMS.ObservedReproduction — Type

ObservedReproduction <: SimulationPlot

A simulation plot type for generating a observed-reproduction-number-plot.

GEMS.ObservedSerialInterval — Type

ObservedSerialInterval <: SimulationPlot

A simulation plot type for generating a observed-serial-interval-plot.

GEMS.PopulationPyramid — Type

PopulationPyramid <: SimulationPlot

A simulation plot type for generating a population pyramid for the associated population model.

GEMS.SettingAgeContacts — Type

SettingAgeContacts <: SimulationPlot

A simulation plot type for sampling contacts from the model and build an age x age matrix for a given Setting type.

GEMS.SettingSizeDistribution — Type

SettingSizeDistribution <: SimulationPlot

A simulation plot type for generating a population pyramid for the associated population model.

GEMS.SimulationPlot — Type

Supertype for all plots that go into single-run simulation reports

GEMS.SymptomCategories — Type

SymptomCategories <: SimulationPlot

A simulation plot type for visualizing the symptom categories of infections by age.

GEMS.TestPositiveRate — Type

TestPositiveRate <: SimulationPlot

A simulation plot type for generating a tests-positive-rate-per-tick plot.

GEMS.TickCases — Type

TickCases <: SimulationPlot

A simulation plot type for generating a new-cases-per-tick plot.

GEMS.TickCasesBySetting — Type

TickCasesBySetting <: SimulationPlot

A simulation plot type for generating tick cases for each included setting type.

GEMS.TickTests — Type

TickTests <: SimulationPlot

A simulation plot type for generating a new-tests-per-tick plot.

GEMS.TimeToDetection — Type

TimeToDetection <: SimulationPlot

A simulation plot type for generating a time-to-detection plot.

GEMS.TotalTests — Type

TotalTests <: SimulationPlot

A simulation plot type for generating a total-tests-per-tick plot.

Functions

GEMS.emptyplot — Method

emptyplot(message::String)

Returns an empty plot with the argument message as a text overlay. This can be used as a default for plot-generate()-functions, if no data is available in the ResultData object.

GEMS.gemsplot — Function

gemsplot(rd::ResultData; type::Symbol = :nothing, plotargs...)
gemsplot(bd::BatchData; type::Symbol = :nothing, combined::Symbol = :all, plotargs...)
gemsplot(rd::Vector{ResultData}; type::Symbol = :nothing, combined::Symbol = :all, plotargs...)

Facilitates the usage of GEMS' inbuilt plots. Just pass the name of the plot-type as a Symbol (must be exactly the same as the respective SimulationPlot-struct). The plotargs... can be any keyworded argument that is available in the standard Plots.jl package.

You can pass a ResultData object to get a plot for one simulation run or even pass a vector of ResultData objects or even a BatchData object. Passing data of multiple simulation runs (ResultData-vector or BatchData) will generate the respective plot with the data of all simulation runs inside. E.g., the :TickCases plot will show one line for each run. Simulations with the same label attribute will be grouped using the same color.

The keyword combined (only applicable for ResultData-vectors or BatchData-objects) determines whether all data is combined in a single plot (:all), each simulation run gets its own subplot (:single), or the plots are separated by label (:bylabel). Note: There might be plots without a multi-plot implementation. They will always be printed as if combined = :single was passed. Check the table below to see which plots are available for single- and multiplots.

Parameters

rd/bd: Data object to plot. Can be ResultData, Vector{ResultData} or BatchData

type = :nothing (optional): Plot type (instantiates a plot with the exact same struct name). You can pass a tuple of plots to generate one graph with multiple visualizations

combined::Symbol = :all (optional): all data in one plot (:all), all individual plots (:singe), plot separated by label (:bylabel).

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Plot using the Plots.jl package's struct.

Examples

Given that rd is a valid ResultData object that came out of a simulation, you can plot a summary like this:

gemsplot(rd)

If you want a specific plot, say the cases-per-tick plot, try this and add a custom title:

gemsplot(rd, type = :TickCases, title = "My Tick Case Plot")

You can also generate multiple plots in one figure like so:

gemsplot(rd, type = (:CompartmentFill, :EffectiveReproduction), layout = (2,1))

The layout keyword makes the plots appear on top of each other in stead of side-by-side.

Generate a multi-plot for a vector of ResultData objects and separate plots by label like so:

# assuming you have a baseline, and a lockdown scenario with two simulation runs each.
# the simulations of these ResultData objects must have the same label ("Baseline" and "Lockdown") to group them
rds = [rd_baseline_1, rd_baseline_2, rd_lockdown_1, rd_lockdown_2]
gemsplot(rds, type = :EffectiveReproduction, combined = :bylabel)

Plot Types

The following plot types can be generated using the gemsplot() function. The mutli-plot column describes which plots can work with ResultData-vector or BatchData-objects. Some of the plots have keyword arguments that are only applicable to that very plot.

Type	Description	Multi-Plot	Plot-Specific Parameters

:ActiveDarkFigure	Fraction of current infections per tick that are "known"/detected at that tick.	Yes

:AggregatedSettingAgeContacts	Age-X-Age contact matrix for all setting types in the simulation (e.g., Household).	No	settingtype::DataType: Filter for specific setting type (e.g., Household), show_values::Bool: Enable or disable printed values in cells.

:CompartmentFill	Current number of individuals per compartment (classic SEIR curves).	No

:CumulativeCases	Summed-up cases over time.	Yes

:CumulativeDiseaseProgressions	Number of people in a certain disease state at a given tick after their exposure.	No

:CumulativeIsolations	Number of currently isolated individuals over time.	Yes	series::Symbol: Filter for :workers, :students, :other, or :all

:CustomLoggerPlot	CustomLogger values. One subplot per logging function in the multi-plot variant.	Yes

:DetectedCases	New detected cases per time step.	No

:EffectiveReproduction	Effective reproduction number (R_eff) over time.	Yes

:GenerationTime	Mean generation time over time.	Yes

:HospitalOccupancy	Number of hospitalized, ventilated, and ICU-admitted indivudual over time.	No

:HouseholdAttackRate	In-Household attack rate per household size.	Yes

:Incidence	Indicende over time by 10-year age group (stacked chart).	No

:InfectionDuration	Histogram of total infection durations.	Yes

:InfectiousHistogram	Histogram of infectious period duratons.	No

:LatencyHistogram	Histogram of latency period durations.	No

:ObservedReproduction	Observed rffective reproduction number (R_eff) over time (based on detected cases).	No

:ObservedSerialInterval	Observed serial interval (SI) over time (based on detected cases).	No

:PopulationPyramid	Population Pyramid by age and sex.	No

:SettingSizeDistribution	Histograms of setting sizes for all setting types (e.g., Households)	No

:SymptomCategories	Heatmap of the fraction of symptom categories (asymptomatic, mild, severe...) by age.	No

:TestPositiveRate	Fraction of all performed tests that were positive per test type (e.g., PCR).	No

:TickCases	New cases per time step. For single sim. also with new infectious, removed, dead.	Yes

:TickCasesBySetting	New cases stratified by setting type (e.g., Household or Office)	No

:TickTests	Performed tests per time step, including positive and total tests and reported cases.	No

:TimeToDetection	Average time between exposure and detected for all infections over time.	No

:TotalTests	Total number of performed tests per test type (e.g., PCR)	Yes

Some Useful Keyword Arguments

Here are some examples of the Plots.jl package's keyword arguments that you can also pass to the gemsplot() function and might find helpful:

- `xlims = (0, 100)`: Setting the X-axis range between 0 and 100
- `ylims = (0, 200)`: Setting the Y-axis range between 0 and 200
- `size = (300, 400)`: Resizing the plot
- `plot_title = "My New Title"`: Changing the plot title
- `xlabel = "New X-label"`: Changing the x-axis label
- `ylabel = "New Y-label"`: Changing the y-axis label
- `legend = :topright`: Changing the legend position (`false` to disable)
- `aspect_ratio = :equal`: Having the axis of equal size

Please consult the Plots.jl package documentation for a comprehensive list

GEMS.plot_ageGroup_contact_distribution — Function

plot_ageGroup_contact_distribution(ageGroup_contact_distribution::AgeGroupContactDistribution; style::Symbol=:histogram, titlefontsize::Union{Nothing, Int64}=nothing)

Create a single plot of a "AgeGroup x AgeGroup Contact Distribution". This function supports different styles to plot the contact distribution.

Parameters

ageGroup_contact_distribution::AgeGroupContactDistribution: AgeGroup x AgeGroup Contact Distribution

style::Symbol: different style options. Currently supported:

:abs_histogram: displays the contact distribution as a histogram with absolute frequencies.

:rel_histogram: displays the contact distribution as a histogram with relative frequencies.

titlefontsize::Union{Nothing, Int64}: optional fontsize for all fonts used in the plot

GEMS.plot_age_contact_distribution — Function

plot_age_contact_distribution(age_contact_distribution::AgeContactDistribution; style::Symbol=:rel_histogram, titlefontsize::Union{Nothing, Int64}=nothing)

Create a single plot of a "Age x Age Contact Distribution". This function supports different styles to plot the contact distribution.

Parameters

age_contact_distribution::AgeContactDistribution: Age x Age Contact Distribution

style::Symbol: different style options. Currently supported:

:abs_histogram: displays the contact distribution as a histogram with absolute frequencies.

:rel_histogram: displays the contact distribution as a histogram with relative frequencies.

titlefontsize::Union{Nothing, Int64}: optional fontsize for all fonts used in the plot

GEMS.plot_all_age_contact_distributions — Function

plot_all_age_contact_distributions(age_contact_distribution_matrix::Matrix{AgeContactDistribution}; style::Symbol=:rel_histogram)

Generates a subplot containing plots with the style style for every entry of distribution_matrix. This function supports different styles to plot the contact distribution.

Parameters

age_contact_distribution_matrix::Matrix{AgeContactDistribution}: matrix containing a distribution of contacts in each cell.

style::Symbol: different style options. Currently supported:

:abs_histogram: displays the contact distribution as a histogram with absolute frequencies.

:rel_histogram: displays the contact distribution as a histogram with relative frequencies.

GEMS.plot_multiple_ageGroup_contact_distributions — Function

plot_multiple_ageGroup_contact_distributions(age_group_contact_distributions::Vector{AgeGroupContactDistribution}; style=:rel_histogram)

Generates a subplot containing plots with the style style for every entry of age_group_contact_distributions. For each entry, a plot of the 'AgeGroup x AgeGroup Contact Distribution' will be created. This function supports different styles to plot the contact distribution.

Parameters

age_group_contact_distributions::Vector{AgeGroupContactDistribution}: 'AgeGroup x AgeGroup Contact Distribution's that should be plotted.

style::Symbol: different style options. Currently supported:

:abs_histogram: displays the contact distribution as a histogram with absolute frequencies.

:rel_histogram: displays the contact distribution as a histogram with relative frequencies.

GEMS.plot_multiple_age_contact_distributions — Function

plot_multiple_age_contact_distributions(age_contact_distributions::Vector{AgeContactDistribution}; style::Symbol=:rel_histogram)

Generate a subplot containing plots with the style style for every AgeContactDistribution defined in age_contact_distributions This function supports different styles to plot the contact distribution.

Parameters

age_contact_distributions::Vector{AgeContactDistribution}: Vector of AgeContactDistributions that should be plotted.

style::Symbol: different style options. Currently supported:

:abs_histogram: displays the contact distribution as a histogram with absolute frequencies.

:rel_histogram: displays the contact distribution as a histogram with relative frequencies.

Missing docstring.

Missing docstring for plot_specific_age_contact_distributions. Check Documenter's build log for details.

GEMS.plotseries! — Function

plotseries!(p::Plots.Plot, extract_function::Function, rds::Vector{ResultData}; plotargs...)

Extends a plot and adds data grouped by the labels of the ResultData objects. The extract_function argument must be a one-argument function that is the called with a single ResultData object. It must return the data series that should be plotted.

Parameters

p::Plots.Plot: Input plot object to extend

extract_function::Function: One-argument lambda function that will be called with each of the ResultData objects in the input vector. This function must return a vector of numberical values that is being added as a line series to input plot p.

rds::Vector{ResultData}: Vector of ResultData objects to plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Plot using the Plots.jl package's struct.

Example

This code generates a plot of the effective reproduction number with one line per ResultData object and grouped by label.

p = plotseries!(plot(), rd -> effectiveR(rd)[!,"rolling_R"], rds)

GEMS.saveplot — Function

saveplot(plot::Plots.Plot, path::AbstractString)

Stores a plot from the juliaplots package to the provided path.

saveplot(plot::VegaLite.VLSpec, path::AbstractString)

Stores a plot from the VegaLite package to the provided path.

saveplot(plot::GMTWrapper, path::AbstractString)

Copies a GMT plot from the temp folder into the provided path (removes the temporary file).

GEMS.splitlabel — Function

splitlabel(plt::SimulationPlot, rds::Vector{ResultData}; plotargs...)

Returns a split side-by-side plot for multiple ResultData objects but groups simulation runs by label. If you have 2 scenarios with 10 simulation runs each, this function will generate two plots with 10 data series each.

plt::SimulationPlot: Simulation plot struct (e.g. TickCases())

rds::Vector{ResultData}: Vector of ResultData objects to plot (requires the rd objects to have the label attribute)

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Plot using the Plots.jl package's struct.

GEMS.splitplot — Function

splitplot(plt::SimulationPlot, rds::Vector{ResultData}; plotargs...)

Returns a split side-by-side plot for multiple ResultData objects.

Parameters

plt::SimulationPlot: Simulation plot struct (e.g. TickCases())

rds::Vector{ResultData}: Vector of ResultData objects to plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Plot using the Plots.jl package's struct.