Population

Overview Structs

GEMS.Population

Overview Functions

GEMS.add!

GEMS.dataframe

GEMS.each!

GEMS.get_individual_by_id

GEMS.group_by_age

GEMS.maxage

GEMS.num_of_infected

GEMS.params

GEMS.populationDF

GEMS.populationDF

GEMS.population_size

GEMS.populationfile

GEMS.populationfile

GEMS.populationfile

GEMS.remove!

GEMS.save

Structs

GEMS.Population — Type

Population

A Type for a simple population. Acts as a container for a collection of individuals.

Fields

individuals::Vector{Individual}: List of associated individuals

maxage: Age of the oldest individual

Functions

GEMS.add! — Method

add!(population::Population, individual::Individual)

Appends specified individual to a population.

GEMS.dataframe — Method

dataframe(population::Population)

Returns a DataFrame representing the given population.

Returns

DataFrame with the following columns:

Name	Type	Description

id	Int32	Individual id

sex	Int8	Individual sex

age	Int8	Individual age

education	Int8	Individual education level

occupation	Int8	Individual occupation group

household	Int32	Individual associated household

office	Int32	Individual associated office

school	Int32	Individual associated school

GEMS.each! — Function

each!(f, population::Population)

Applies function f to all individuals in the population.

Example

each!(i -> i.age = i.age + 1, pop) lets all individuals in the populaiton pop age by one year.

GEMS.get_individual_by_id — Method

get_individual_by_id(population::Population, ind::Int32)

Returns an individual contained in the Population selected by its id.

GEMS.group_by_age — Method

group_by_age(df::DataFrame)

Groups a dataframe by its "age" column. The column "sum" contains the sum of each individual in this age. This also adds rows for each age from 0 till the maximum age in the "age" column.

Returns

returns a DataFrame with an additional column "sum". This column contains the count of the age in the DataFrame.

Missing docstring.

Missing docstring for is_pop_file. Check Documenter's build log for details.

GEMS.maxage — Method

maxage(population::Population)

Returns the maximum age of any individual in the population

GEMS.num_of_infected — Method

num_of_infected(population::Population)

Returns the number of infected individuals in a given population.

Missing docstring.

Missing docstring for obtain_remote_files(::String; ::Bool). Check Documenter's build log for details.

Missing docstring.

Missing docstring for obtain_remote_files(::String). Check Documenter's build log for details.

GEMS.params — Method

params(population::Population)

Returns the parameters that were used to generate this population.

GEMS.populationDF — Method

populationDF(postProcessor)

Returns the internal flat population DataFrame.

Columns

Name	Type	Description

id	Int32	Individual id

sex	Int8	Individual sex

age	Int8	Individual age

education	Int8	Individual education level

occupation	Int8	Individual occupation group

household	Int32	Individual associated household

office	Int32	Individual associated office

school	Int32	Individual associated school

GEMS.populationDF — Method

populationDF(simulation::Simulation)

Calls the dataframe() function on the simulation's Population object.

GEMS.population_size — Method

population_size(rd::ResultData)

Returns the size of the population model in memory.

GEMS.populationfile — Method

populationfile(batchPopulationPyramid)

Returns the populationfile from an associated BatchPopulationPyramid object.

GEMS.populationfile — Method

populationfile(batchSettingAgeContacts)

Returns the populationfile from an associated BatchSettingAgeContacts object.

GEMS.populationfile — Method

populationfile(population::Population)

Returns the population file that was used to generate this population.

GEMS.remove! — Method

remove!(population::Population, individual::Individual)

Remove a specified individual from a population.

GEMS.save — Method

save(population::Population, path::AbstractString)

Saves the given population as a CSV-file at path.