Reporting

Overview Structs

GEMS.AbstractSection

GEMS.BatchReport

GEMS.PlotSection

GEMS.Report

GEMS.ReportPlot

GEMS.Section

GEMS.SimulationReport

Overview Functions

GEMS.abstract

GEMS.abstract!

GEMS.addsection!

GEMS.addtimer!

GEMS.author

GEMS.author!

GEMS.buildreport

GEMS.content

GEMS.content!

GEMS.date

GEMS.date!

GEMS.description

GEMS.description!

GEMS.dpi

GEMS.dpi!

GEMS.filename

GEMS.filename!

GEMS.fontfamily

GEMS.fontfamily!

GEMS.generate

GEMS.glossary

GEMS.glossary!

GEMS.plt

GEMS.reportdata

GEMS.savepath

GEMS.sections

GEMS.subsections

GEMS.subtitle

GEMS.subtitle!

GEMS.title

GEMS.title!

Structs

GEMS.AbstractSection — Type

Supetype for all report sections

GEMS.BatchReport — Type

BatchReport <: Report

A Type for generating a batch-run simulation report.

Fields

Batch Data

bd::BatchData: Batch data object used to generate this report

Meta Information

title::String: Report title

subtitle::String: Report subtitle

author::String: Report authors

date::String: Simulation execution date (post processing)

abstract::String: Report abstract

Configuration

glossary::Bool: flag to add or remove glossary with term definitions

dpi::Int64: dots per inch for report plots (default: 300)

fontfamily::String: font family for report (default: Times New)

GEMS.PlotSection — Type

PlotSection <: AbstractSection

This type wraps a ReportPlot into an AbstractSection struct. Thus the only field is a report plot. Its mainly used to add plots to reports. You can look up the aviailable default report styles to get an idea how that might look like.

You can either pass an instance of a SimulationPlot to the constructor like:

PlotSection(TickCases())

or pass the SimulationPlot type as a Symbol like:

PlotSection(:TickCases)

The gemsplot() function contains a comprehensive list on the available plot types.

GEMS.Report — Type

Supertype for all simulation reports

GEMS.ReportPlot — Type

Supertype for all plots that go into reports

GEMS.Section — Type

Section(title::String = "", content::String = "", subsections::Vector{AbstractSection} = [])
Section(rd::ResultData, type::Symbol)
Section(rd::BatchData, type::Symbol)

A type for Report Sections. All reports consist of (nested) sections. A report section will be parsed into a markdown section (indicated by "#"s) during report generation. Each section must have a title and content. It can have an arbitrary number of subsections.

Example

This code creates two sections. The first one having a title and some content. The second section contains the first secion as a subsection.

sec_1 = Section(
    title = "My Section Title",
    content = "Great section contents"
)

sec_2 = Section(
    title = "My Second Section
    subsections = [sec_1]
)

Defaut Sections

There are a number of default sections that you can simply plug into your report by calling the Section() constructor and pass a ResultData or BatchData object and a Symbol qualifying the type of section you'd like to get:

# run simulation
sim = Simulation()
run!(sim)
rd = ResultData(sim)

overview_section = Section(rd, :Overview)

Default Sections for Single Simulation Reports

Input must be ResultData.

Type	Title	Content

:Debug	Debug Information	Contains :Memory, :Processor,:Repo, and :System section.

:General	General	Geeral simulation info; tick unit, start conditions, and others.

:InputFiles	Input Files	Config- and population file paths.

:Interventions	Interventions	Triggers, strategies, and measures.

:Memory	Memory	Available and used system memory.

:Model	Model Configuration	Contains :InputFiles, :Interventions, :General, :Pathogens, and :Settings

:Observations	Observations	Summary of observed progression, detection rate, dark figure, and others.

:Overview	Overview	Simulation summmary with initial infections, total attack rate, and others.

:Pathogens	Pathogens	Pathogen configuration with one subsection per pathogen.

:Processor	Processor	Processor model, cores, and others.

:Repo	Repository	Current repo-, branch- and commit-ID.

:Settings	Settings	Number of settings per type and min, max, average number of individuals.

:System	System Information	Julia config, number of threads, and others.

Default Sections for Batch Reports

Input must be BatchData.

Type	Title	Content

:Allocations	Allocations	Summary statistics on memory usage across simulation runs in BatchData object.

:Debug	Debug Information	Contains :Memory, :Processor,:Repo, and :System section.

:General	Genera	Geeral simulation info; tick unit, start conditions, and others.

:Memory	Memory	Available and used system memory.

:Overview	Overview	Summary statistics on total infections, attack rates, and others.

:Processor	Processor	Processor model, cores, and others.

:Repo	Repository	Current repo-, branch- and commit-ID.

:Resouces	Resources	Contains :Runtime and :Allocations sections.

:Runtime	Runtime	Summary statistics on runtime across simulation runs in BatchData object.

:Settings	Settings	Number of settings per type and min, max, average number of individuals.

:System	System Information	Julia config, number of threads, and others.

Missing docstring.

Missing docstring for SectionBuilder. Check Documenter's build log for details.

GEMS.SimulationReport — Type

SimulationReport <: Report

A Type for generating a single-run simulation report.

Fields

Result Data

rd::ResultData: Result data object used to generate this report

Meta Information

title::String: Report title

subtitle::String: Report subtitle

author::String: Report authors

date::String: Simulation execution date (post processing)

abstract::String: Report abstract

Configuration

glossary::Bool: flag to add or remove glossary with term definitions

dpi::Int64: dots per inch for report plots (default: 300)

fontfamily::String: font family for report (default: Times New)

Functions

GEMS.abstract — Function

abstract(report)

Returns the abstract of a Report.

GEMS.abstract! — Function

abstract!(report, abstract)

setss the abstract of a Report.

GEMS.addsection! — Function

addsection!(section::Section, subsection::AbstractSection)

Adds a subsection to a provided section.

addsection!(section::Section, subsections::Vector)

Adds multiple subsection to a provided section.

addsection!(report, section)

Adds a Section to a Report. It can either be a regular section or a plot section.

GEMS.addtimer! — Function

addstimer!(report, timeroutput)

Generates a Section from the consolue output of a TimerOutput object and adds it to a SimulationReport.

GEMS.author! — Function

author!(report, author)

Sets the author of a Report.

GEMS.author — Function

author(report)

Returns the author of a Report.

GEMS.buildreport — Function

buildreport(rd::ResultData, config::Dict = Dict())

Initializes and configures a simulation report with the tile, abstract, sections etc. provided in config. If config is an empty dictionary all available standard sections, plots, and the glossary etc. will be used. It returns a full SimulationReport object which can then be generated using the generate() function.

GEMS.content! — Function

content!(section::Section, content::String)

Sets a section's content.

GEMS.content — Function

content(section::Section)

Returns a section's content.

GEMS.date! — Function

date!(report, date)

Sets the date of a Report.

GEMS.date — Function

date(report)

Returns the date of a Report.

Missing docstring.

Missing docstring for debugrep(::ResultData). Check Documenter's build log for details.

GEMS.description! — Method

description!(plot::ReportPlot, description::String)

Setter for report plot description.

GEMS.description — Method

description(plot::ReportPlot)

Return description from report plot object.

GEMS.dpi — Function

dpi(report)

Returns dpi (dots per inch) from a report object. dpi conditions the resolution of images generated for the report.

GEMS.dpi! — Method

dpi!(report, dpi)

Setter for report dpi (dot per inch) for images.

GEMS.filename! — Method

filename!(plot::ReportPlot, filename::String)

Setter for report plot filename.

GEMS.filename — Method

filename(plot::ReportPlot)

Return filename from report plot object.

GEMS.fontfamily! — Function

fontfamily!(report, fontfamily)

Setter for report fontfamily.

GEMS.fontfamily — Function

fontfamily(report)

Returns font family config from a report object.

Missing docstring.

Missing docstring for generalrep. Check Documenter's build log for details.

GEMS.generate — Function

Abstract wrapper function for all simulation report plots. Requires concrete implementation in subtypes

generate(batchEffectiveReproduction, batchData)

Generates a plot with the effecive reproduction number per tick for a specified batch of simulations.

generate(batchPopulationPyramid, batchData)

Generates population pyramid for a specified population file within a batch.

generate(batchSettingAgeContacts, batchData)

Generates and returns an age x age matrix from sampled contacts for a given Setting type.

generate(batchTickCases, batchData)

Generates a plot with cases per tick for a specified batch of simulations.

generate(batchTickIsolations, batchData)

Generates a plot with currently quarantined people per tick for a specified batch of simulations.

generate(BatchTickTests, batchData)

Generates a plot with subplots for each test type. Each includes tests per tick for a specified batch of simulations.

generate(plt::ActiveDarkFigure, rd::ResultData; plotargs...)

Generates and returns a active-darkfigure-per-tick plot for a provided ResultData object. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::ActiveDarkFigure: SimulationPlot struct with meta data (i.e. title, description, and filename)

rd::ResultData: Input data used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Active Dark Figure plot

generate(plt::ActiveDarkFigure, rds::Vector{ResultData}; plotargs...)

Generates and returns a active-darkfigure-per-tick plot for a provided vector of ResultData objects. In contrast to the single-ResultData version of this plot, this implementation will visualize the fraction of detected cases and not the absolute number of cases. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::ActiveDarkFigure: SimulationPlot struct with meta data (i.e. title, description, and filename)

rds::Vector{ResultData}: Vector of input data objects used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Active Dark Figure multi plot

generate(plt::AggregatedSettingAgeContacts, rd::ResultData;
    settingtype::Union{DataType, Nothing} = nothing, plotargs...)

Generates and returns an age group x age group matrix from sampled contacts for a given Setting type. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::AggregatedSettingAgeContacts: SimulationPlot struct with meta data (i.e. title, description, and filename)

rd::ResultData: Input data used to generate plot

settingtype::Union{DataType, Nothing} = nothing (optional): Setting type (e.g. "Household"). If nothing is passed, all setting types in the ResultData object are being ploted.

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Aggregated Setting Age Contacts plot

generate(plt::CompartmentFill, rd::ResultData; plotargs...)

Generates and returns a plot of the current compartment fill for a provided ResultData object. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::CompartmentFill: SimulationPlot struct with meta data (i.e. title, description, and filename)

rd::ResultData: Input data used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Compartment Fill plot

generate(plt::CumulativeCases, rd::ResultData; plotargs...)

Generates and returns a cumulative infections plot for a provided simulation object. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::CumulativeCases: SimulationPlot struct with meta data (i.e. title, description, and filename)

rd::ResultData: Input data used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Cumulative Cases plot

generate(plt::CumulativeCases, rds::Vector{ResultData}; plotargs...)

Generates and returns a cumulative infections plot for a provided vector of ResultData objects. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::CumulativeCases: SimulationPlot struct with meta data (i.e. title, description, and filename)

rds::Vector{ResultData}: Vector of input data objects used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Cumulative Cases multi plot

generate(plt::CumulativeDiseaseProgressions, rd::ResultData; plotargs...)

Generates a stacked bar chart of cumulative disease progressions for all infections. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::CumulativeDiseaseProgressions: SimulationPlot struct with meta data (i.e. title, description, and filename)

rd::ResultData: Input data used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Cumulative Disease Progressions plot

generate(plt::CumulativeIsolations, rd::ResultData; plotargs...)

Generates a plot for the cumulative number of isolated individuals per tick. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::CumulativeIsolations: SimulationPlot struct with meta data (i.e. title, description, and filename)

rd::ResultData: Input data used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Cumulative Isolations plot

generate(plt::CumulativeIsolations, rds::Vector{ResultData};
    series::Symbol = :all, plotargs...)

Generates and returns a plot for the cumulative isolations for a provided vector of ResultData objects. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

The series argument allows to specifcy if quarantined :workers, :students, or :other should be plotted.

Parameters

plt::CumulativeIsolations: SimulationPlot struct with meta data (i.e. title, description, and filename)

rds::Vector{ResultData}: Vector of input data objects used to generate plot

series::Symbol = :all (optional): specifcy if quarantined :workers, :students, or :other should be plotted

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Cumulative Isolations multi plot

generate(plt::CustomLoggerPlot, rd::ResultData; plotargs...)

Generates and returns a plot for the values contained in the custom logger of a ResultData object. It will contain one individual plot per function that was passed to the custom logger. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package. However, be aware that the keyword arguments might be applied to each of the subplots individually.

Parameters

plt::CustomLoggerPlot: SimulationPlot struct with meta data (i.e. title, description, and filename)

rd::ResultData: Input data used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Custom Logger Plot plot

generate(plt::CustomLoggerPlot, rds::Vector{ResultData}; plotargs...)

Generates and returns a plot for the values contained in the custom loggers of a provided vector of ResultData objects. It will contain one individual plot per function that was passed to the custom logger. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::CumulativeIsolations: SimulationPlot struct with meta data (i.e. title, description, and filename)

rds::Vector{ResultData}: Vector of input data objects used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Cumulative Isolations multi plot

generate(plt::DetectedCases, rd::ResultData; plotargs...)

Generates and returns a new-DETECTED-cases-per-tick plot for a provided simulation object. Sorts infections dataframe by test_tickand filters for tested individuals. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::DetectedCases: SimulationPlot struct with meta data (i.e. title, description, and filename)

rd::ResultData: Input data used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Detected Cases plot

generate(plt::EffectiveReproduction, rd::ResultData; plotargs...)

Generates a plot for the effective reproduction number per tick. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::EffectiveReproduction: SimulationPlot struct with meta data (i.e. title, description, and filename)

rd::ResultData: Input data used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Effective Reproduction Number plot

generate(plt::EffectiveReproduction, rds::Vector{ResultData}; plotargs...)

Generates a plot for the effective reproduction number per tick for a provided vector of ResultData objects. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::EffectiveReproduction: SimulationPlot struct with meta data (i.e. title, description, and filename)

rds::Vector{ResultData}: Vector of input data objects used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Effective Reproduction Number multi plot

generate(plt::GenerationTime, rd::ResultData; plotargs...)

Generates and returns a generation-time-per-tick plot for a provided simulation object. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::GenerationTime: SimulationPlot struct with meta data (i.e. title, description, and filename)

rd::ResultData: Input data used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Generation Time plot

generate(plt::GenerationTime, rds::Vector{ResultData}; plotargs...)

Generates and returns a plot for the mean generation-time-per-tick for a provided vector of ResultData objects. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::TickCases: SimulationPlot struct with meta data (i.e. title, description, and filename)

rds::Vector{ResultData}: Vector of input data objects used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Generation Time multi plot

generate(plt::HospitalOccupancy, rd::ResultData; plotargs...)

Generates a plot of the number of hospitalized, ventilated and ICU admitted agents for each tick. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::HospitalOccupancy: SimulationPlot struct with meta data (i.e. title, description, and filename)

rd::ResultData: Input data used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Hospital Occupancy plot

generate(plt::HouseholdAttackRate, rd::ResultData; plotargs...)

Generates and returns a household-attack-rate plot for a provided simulation object. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package. However, be aware that the keyword arguments might be applied to each of the subplots individually.

Parameters

plt::HouseholdAttackRate: SimulationPlot struct with meta data (i.e. title, description, and filename)

rd::ResultData: Input data used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Household Attack Rate plot

generate(plt::HouseholdAttackRate, rds::Vector{ResultData}; plotargs...)

Generates and returns a household-attack-rate plot for a vector of provided ResultData objects. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package. However, be aware that the keyword arguments might be applied to each of the subplots individually.

Parameters

plt::HouseholdAttackRate: SimulationPlot struct with meta data (i.e. title, description, and filename)

rds::Vector{ResultData}: Vector of input data objects used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Household Attack Rate multi plot

generate(plt::Incidence, rd::ResultData; plotargs...)

Generates an age-stratified incidence plot. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::Incidence: SimulationPlot struct with meta data (i.e. title, description, and filename)

rd::ResultData: Input data used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Incidence plot

generate(plt::InfectionDuration, rd::ResultData; plotargs...)

Generates and returns a histogram of the total infection durations in ticks. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::InfectionDuration: SimulationPlot struct with meta data (i.e. title, description, and filename)

rd::ResultData: Input data used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Infection Duration plot

generate(plt::InfectionDuration, rds::Vector{ResultData}; plotargs...)

Generates and returns a histogram of the total infection durations in ticks for a provided vector of ResultData objects. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::InfectionDuration: SimulationPlot struct with meta data (i.e. title, description, and filename)

rd::ResultData: Input data used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Infection Duration plot

generate(plt::InfectionMap, rd::ResultData; plotargs...)

Generates a map of the infections if geodata is provided. If more than the maximum of allowed data points is given (lookup MAX_MAP_POINTS in constants.jl), a sample of size MAX_MAP_POINTS is taken from the original data for visualization.

The current implementation does not offer the option for additional keyworded arguments. The plotargs... argument is just a placeholder.

Parameters

plt::InfectionMap: SimulationPlot struct with meta data (i.e. title, description, and filename)

rd::ResultData: Input data used to generate plot

plotargs... (optional): PLACEHOLDER. Currently not implemented.

Returns

GMTWrapper: Custom struct containing the storage location of the generated map

generate(plt::InfectiousHistogram, rd::ResultData; plotargs...)

Generates a histrogram of the infectious period distribution of collected infections. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::InfectiousHistogram: SimulationPlot struct with meta data (i.e. title, description, and filename)

rd::ResultData: Input data used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Infectious Histogram plot

generate(plt::LatencyHistogram, rd::ResultData; plotargs...)

Generates a histrogram of the latency period distribution of collected infections. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::LatencyHistogram: SimulationPlot struct with meta data (i.e. title, description, and filename)

rd::ResultData: Input data used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Latency Histogram plot

 generate(plt::ObservedReproduction, rd::ResultData; plotargs...)

Generates a plot for the effective reproduction number per tick. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::ObservedReproduction: SimulationPlot struct with meta data (i.e. title, description, and filename)

rd::ResultData: Input data used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Observed Reproduction Number plot

generate(plt::ObservedSerialInterval, rd::ResultData; plotargs...)

Generates a plot for the estimation on the observed serial interval per tick. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::ObservedSerialInterval: SimulationPlot struct with meta data (i.e. title, description, and filename)

rd::ResultData: Input data used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Observed Serial Interval plot

generate(plt::PopulationPyramid, rd::ResultData; plotargs...)

Generates population pyramid for a the associated population model.

The current implementation does not offer the option for additional keyworded arguments. The plotargs... argument is just a placeholder.

Parameters

plt::PopulationPyramid: SimulationPlot struct with meta data (i.e. title, description, and filename)

rd::ResultData: Input data used to generate plot

plotargs... (optional): PLACEHOLDER. Currently not implemented.

Returns

Plots.Plot: Population Pyramid plot

generate(plt::SettingAgeContacts, rd::ResultData; plotargs...)

Generates and returns an age x age matrix from sampled contacts for a given Setting type. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::SettingAgeContacts: SimulationPlot struct with meta data (i.e. title, description, and filename)

rd::ResultData: Input data used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Setting Age Contacts plot

generate(plt::SettingSizeDistribution, rd::ResultData; plotargs...)

Generates the setting size distributions for all included settings. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::SettingSizeDistribution: SimulationPlot struct with meta data (i.e. title, description, and filename)

rd::ResultData: Input data used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Setting Size Distribution plot

generate(plt::SymptomCategories, rd::ResultData; plotargs...)

Generates and returns a symptom_category x age matrix as heatmap. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::SymptomCategories: SimulationPlot struct with meta data (i.e. title, description, and filename)

rd::ResultData: Input data used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Symptom Categories plot

generate(plt::TestPositiveRate, rd::ResultData; plotargs...)

Generates and returns a tests-positive-rate-per-tick plot for a provided simulation object. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::TestPositiveRate: SimulationPlot struct with meta data (i.e. title, description, and filename)

rd::ResultData: Input data used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Test Positive Rate plot

generate(plt::TickCases, rd::ResultData; plotargs...)

Generates and returns a new-cases-per-tick plot for a provided simulation object. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::TickCases: SimulationPlot struct with meta data (i.e. title, description, and filename)

rd::ResultData: Input data used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Tick Cases plot

generate(plt::TickCases, rds::Vector{ResultData}; plotargs...)

Generates and returns a plot for the new-cases-per-tick for a provided vector of ResultData objects. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::TickCases: SimulationPlot struct with meta data (i.e. title, description, and filename)

rds::Vector{ResultData}: Vector of input data objects used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Tick Cases multi plot

generate(plt::TickCasesBySetting, rd::ResultData; plotargs...)

Generates a plot of tick cases for each included setting type. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::TickCasesBySetting: SimulationPlot struct with meta data (i.e. title, description, and filename)

rd::ResultData: Input data used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Tick Cases By Setting plot

generate(plt::TickTests, rd::ResultData; plotargs...)

Generates and returns a tests-per-tick plot for a provided simulation object. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::TickTests: SimulationPlot struct with meta data (i.e. title, description, and filename)

rd::ResultData: Input data used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Tick Tests plot

generate(plt::TimeToDetection, rd::ResultData; plotargs...)

Generates and returns a time-to-detection plot for a provided simulation object. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::TimeToDetection: SimulationPlot struct with meta data (i.e. title, description, and filename)

rd::ResultData: Input data used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Time To Detection plot

generate(plt::TotalTests, rd::ResultData; plotargs...)

Generates and returns a total-tests-per-tick plot for a provided simulation object. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::TotalTests: SimulationPlot struct with meta data (i.e. title, description, and filename)

rd::ResultData: Input data used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Total Tests plot

generate(plt::TotalTests, rds::Vector{ResultData}; plotargs...)

Generates and returns a total tests plot for a provided vector of ResultData objects. It will contain one individual plot per test type available in the ResultData object. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::TotalTests: SimulationPlot struct with meta data (i.e. title, description, and filename)

rds::Vector{ResultData}: Vector of input data objects used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Total Tests multi plot

Abstract wrapper function for all map plots. Requires concrete implementation in subtypes

generate(plt::AgeMap, sim::Simulation; level::Int = 3, plotargs...)

Generates and returns a map showing the average age per region for a provided simulation object. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::AgeMap: MapPlot struct with meta data (i.e. title, description, and filename)

sim::Simulation: Input data used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Average age map

generate(plt::AttackRateMap, sim::Simulation; level::Int = 3, plotargs...)

Generates and returns a map showing the fraction of individuals that got infected at least one time for a provided ResultData object. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::AttackRateMap: MapPlot struct with meta data (i.e. title, description, and filename)

rd::ResultData: Input data used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Attack rate map plot

generate(plt::CaseFatalityMap, sim::Simulation; level::Int = 3, plotargs...)

Generates and returns a map showing the fraction of infecftions that led to death for a provided ResultData object. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::CaseFatalityMap: MapPlot struct with meta data (i.e. title, description, and filename)

rd::ResultData: Input data used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Case fatality map plot

generate(plt::HouseholdSizeMap, sim::Simulation; level::Int = 3, plotargs...)

Generates and returns a map showing the average household size per region for a provided simulation object. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::AgeMap: MapPlot struct with meta data (i.e. title, description, and filename)

sim::Simulation: Input data used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Average household size map

generate(plt::PopDensityMap, sim::Simulation; level::Int = 3, plotargs...)

Generates and returns a map showing the population density per region for a provided simulation object. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::PopDensityMap: MapPlot struct with meta data (i.e. title, description, and filename)

sim::Simulation: Input data used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Population density map

generate(plt::SinglesMap, sim::Simulation; level::Int = 3, plotargs...)

Generates and returns a map showing the fraction of single-person households for a provided simulation object. You can pass any additional keyword arguments using plotargs... that are available in the Plots.jl package.

Parameters

plt::SinglesMap: MapPlot struct with meta data (i.e. title, description, and filename)

sim::Simulation: Input data used to generate plot

plotargs... (optional): Any argument that the plot() function of the Plots.jl package can take.

Returns

Plots.Plot: Fraction of single-person households map plot

generate(section::Section, depth::Int64, rd::AbstractResultData, directory::AbstractString)

Generates markdown string from a Section object including its nested subsections. The depth parameter controls the number of leading "#"s before the title. A directory must be provided in case subsections generate images that will be stored outside of the report.

generate(section::PlotSection, depth::Int64, rd::AbstractResultData, directory::AbstractString)

Generates markdown string from a PlotSection object. The depth parameter controls the number of leading "#"s before the title. A directory must be provided as the generated images (of the plots) will be stored outside of the report in dedicated "/img" folder.

generate(section::AbstractSection, rd::AbstractResultData, directory::AbstractString)

Generates markdown string from an AbstractSection object including its potential nested subsections starting from depth 1 (one leading "#" for titles). A directory must be provided in case subsections generate images that will be stored outside of the report.

generate(report, directory)

Generates markdown string from a Report object including its nested subsections. The report will be stored as PDF, HTML, and MD in the provided directory. Images are generated and stored in an "/img" subfolder.

GEMS.glossary! — Function

glossary!(report, glossary)

Sets the glossary flag of a Report. If true, the glossary will be copied into the report upon generation.

GEMS.glossary — Function

glossary(report)

Returns the glossary flag of a Report.

Missing docstring.

Missing docstring for inputfilesrep. Check Documenter's build log for details.

Missing docstring.

Missing docstring for memory_sec. Check Documenter's build log for details.

Missing docstring.

Missing docstring for modelconfigrep. Check Documenter's build log for details.

Missing docstring.

Missing docstring for observationsummaryrep(::ResultData). Check Documenter's build log for details.

Missing docstring.

Missing docstring for pathogensrep. Check Documenter's build log for details.

Missing docstring.

Missing docstring for plotpackage(::PlotSection). Check Documenter's build log for details.

GEMS.plt — Function

plt(section::PlotSection)

Return the nested ReportPlot from a PlotSection object.

Missing docstring.

Missing docstring for population_pyramid_rep(::BatchData). Check Documenter's build log for details.

Missing docstring.

Missing docstring for population_pyramid_sec(::BatchData). Check Documenter's build log for details.

Missing docstring.

Missing docstring for processor_sec(::ResultData). Check Documenter's build log for details.

GEMS.reportdata — Method

reportdata(report)

Returns the associated ReportData object from a Report.

Missing docstring.

Missing docstring for repository_sec(::ResultData). Check Documenter's build log for details.

Missing docstring.

Missing docstring for resourcesrep(::BatchData). Check Documenter's build log for details.

Missing docstring.

Missing docstring for runtime_sec(::BatchData). Check Documenter's build log for details.

GEMS.savepath — Method

savepath(str::String)

Replaces all backslashes in a path-string with forward slashes to prevent weaving issues while generating TeX-PDFs as backslashes indicate control structures in TeX.

GEMS.sections — Method

sections(report)

Returns the array of sections of a Report.

Missing docstring.

Missing docstring for setting_age_contacts_rep. Check Documenter's build log for details.

Missing docstring.

Missing docstring for setting_age_contacts_sec. Check Documenter's build log for details.

Missing docstring.

Missing docstring for settingrep. Check Documenter's build log for details.

Missing docstring.

Missing docstring for simoverviewrep. Check Documenter's build log for details.

GEMS.subsections — Method

subsections(section::Section)

Returns a section's array of subsections.

GEMS.subtitle! — Method

subtitle!(report, subtitle)

Sets the subtitle of a Report.

GEMS.subtitle — Method

subtitle(report)

Returns the subtitle of a Report.

Missing docstring.

Missing docstring for sysinformationsec. Check Documenter's build log for details.

Missing docstring.

Missing docstring for sys_information_sec. Check Documenter's build log for details.

GEMS.title! — Function

title!(plot::ReportPlot, title::String)

Setter for report plot title.

title!(section::Section, title::String)

Sets a section's title.

title!(report, title)

Sets the title of a Report.

GEMS.title — Function

title(plot::ReportPlot)

Returns title from report plot object.

title(section::Section)

Returns a section's title.

title(report)

Returns the title of a Report.

Missing docstring.

Missing docstring for vaccinationrep(::ResultData). Check Documenter's build log for details.

Missing docstring.

Missing docstring for varied_parameters_sec. Check Documenter's build log for details.