Result data

Overview Structs

GEMS.DefaultResultData

GEMS.LightRD

GEMS.ResultData

GEMS.ResultData

GEMS.ResultData

GEMS.ResultData

GEMS.ResultData

GEMS.ResultDataStyle

Overview Functions

GEMS.GEMS_version

GEMS.aggregated_compartment_periods

GEMS.allempty

GEMS.attack_rate

GEMS.compartment_fill

GEMS.compartment_periods

GEMS.config_file

GEMS.config_file_val

GEMS.cpu_data

GEMS.cumulative_cases

GEMS.cumulative_deaths

GEMS.cumulative_disease_progressions

GEMS.cumulative_quarantines

GEMS.cumulative_vaccinations

GEMS.customlogger

GEMS.data_hash

GEMS.dataframes

GEMS.deaths

GEMS.detected_tick_cases

GEMS.detection_rate

GEMS.effectiveR

GEMS.execution_date

GEMS.execution_date_formatted

GEMS.exportJLD

GEMS.final_tick

GEMS.free_mem_size

GEMS.git_branch

GEMS.git_commit

GEMS.git_repo

GEMS.hashes

GEMS.hospital_df

GEMS.household_attack_rates

GEMS.id

GEMS.import_resultdata

GEMS.infections

GEMS.infections_hash

GEMS.info

GEMS.initial_infections

GEMS.julia_version

GEMS.kernel

GEMS.label

GEMS.meta_data

GEMS.model_size

GEMS.number_of_individuals

GEMS.observed_R

GEMS.pathogens

GEMS.population_file

GEMS.population_params

GEMS.population_pyramid

GEMS.population_size

GEMS.region_info

GEMS.rolling_observed_SI

GEMS.setting_data

GEMS.setting_sizes

GEMS.sim_data

GEMS.someempty

GEMS.start_condition

GEMS.stop_criterion

GEMS.strategies

GEMS.symptom_triggers

GEMS.system_data

GEMS.tests

GEMS.testtypes

GEMS.threads

GEMS.tick_cases

GEMS.tick_cases_per_setting

GEMS.tick_deaths

GEMS.tick_generation_times

GEMS.tick_pooltests

GEMS.tick_serial_intervals

GEMS.tick_tests

GEMS.tick_unit

GEMS.tick_vaccinations

GEMS.time_to_detection

GEMS.timer_output

GEMS.timer_output!

GEMS.total_infections

GEMS.total_mem_size

GEMS.total_quarantines

GEMS.total_tests

GEMS.word_size

ResultData

Constructors

GEMS.ResultData — Method

ResultData(sim::Simulation; style::String = "")

Create a ResultData object using a Simulation and the name of a ResultDataStyle, that describes the level of detail for the fields to be calculated. This constructor instantiates a default PostProcessor for the passed simulation object. If you want to manually configure the PostProcessor, you need to instantiate it first and pass the PostProcessor to the ResultData constructor instead. Post Processing requires a simulation to be done.

GEMS.ResultData — Method

ResultData(postProcessor::PostProcessor; style::String="")

Create a ResultData object using a PostProcessor and a key, that describes the level of detail for the fields to be calculated. Post Processing requires a simulation to be done.

GEMS.ResultData — Method

ResultData(postProcessors::Vector{PostProcessor}; style::String="", print_infos::Bool = false)

Create a vector ResultData objects using a vector of associated PostProcessor objects and a key, that describes the level of detail for the fields to be calculated. Post Processing requires a simulation to be done. It supresses the usual info outputs that are being made during the ResultData generation. If you want to enable them, pass print_infos = true.

GEMS.ResultData — Method

ResultData(sim::Vector{Simulation}; style::String = "", print_infos::Bool = false)

Create a vector ResultData objects using a vector of Simulation objects and the name of a ResultDataStyle, that describes the level of detail for the fields to be calculated. If you want to manually configure the PostProcessor, you need to instantiate it first and pass the PostProcessor to the ResultData constructor instead. Post Processing requires a simulation to be done. It supresses the usual info outputs that are being made during the ResultData generation. If you want to enable them, pass print_infos = true.

GEMS.ResultData — Method

ResultData(batch::Batch; style::String = "", print_infos::Bool = false)

Create a vector ResultData objects using a Batch object and the name of a ResultDataStyle, that describes the level of detail for the fields to be calculated. If you want to manually configure the PostProcessors, you need to instantiate them first and pass the PostProcessors to the ResultData constructor instead. Post Processing requires a simulation to be done. It supresses the usual info outputs that are being made during the ResultData generation. If you want to enable them, pass print_infos = true.

Functions

GEMS.aggregated_compartment_periods — Function

aggregated_compartment_periods(rd::ResultData)

Returns the DataFrame with disease state durations (normalized). Look up the PostProcessor docs to find the column definitions. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.allempty — Function

allempty(f::Function, rds::Vector{ResultData})

Returns true if the provided function returns an empty dictionary for all ResultData objects in the provided vector.

GEMS.attack_rate — Method

attack_rate(rd::ResultData)

Returns the simulation's attack rate. It's total infections divided by population size. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.compartment_fill — Method

compartment_fill(rd::ResultData)

Returns the compartment_fill infections over time. Look up the PostProcessor docs to find the column definitions. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.compartment_periods — Method

compartment_periods(rd::ResultData)

Returns the DataFrame with duration of exposed and infectious states for all infections. Look up the PostProcessor docs to find the column definitions. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.config_file — Method

config_file(rd::ResultData)

Returns the path to the config file

GEMS.config_file_val — Function

config_file_val(rd::ResultData)

Returns the parsed config file. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.cpu_data — Method

cpu_data(rd::ResultData)

Returns the processor information (not available for ARM Macs)

GEMS.cumulative_cases — Method

cumulative_cases(rd::ResultData)

Returns the cumulative infections over time. Look up the PostProcessor docs to find the column definitions. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.cumulative_deaths — Method

cumulative_deaths(rd::ResultData)

Returns the cumulative deaths over time. Look up the PostProcessor docs to find the column definitions. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.cumulative_disease_progressions — Method

cumulative_disease_progressions(rd::ResultData)

Returns the DataFrame with cumultive number of individuals in certain disease states per tick. Look up the PostProcessor docs to find the column definitions. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.cumulative_quarantines — Method

cumulative_quarantines(rd::ResultData)

Returns the DataFrame with number of isolated individuals per tick Look up the PostProcessor docs to find the column definitions. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.cumulative_vaccinations — Method

cumulative_vaccinations(rd::ResultData)

Returns the cumulative vaccinations over time. Look up the PostProcessor docs to find the column definitions. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.customlogger — Method

customlogger(rd::ResultData)

Returns the DataFrame of the Simulation object's internal custom logger. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.data_hash — Method

data_hash(rd::ResultData)

Returns a SHA1 hash value for the ResultData object.

GEMS.dataframes — Method

dataframes(rd::ResultData)

Returns the dataframes of result data. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.deaths — Method

deaths(rd::ResultData)

Returns the deaths DataFrame joined with individuals' attributes. Look up the PostProcessor docs to find the column definitions. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.detected_tick_cases — Method

detected_tick_cases(rd::ResultData)

Returns the detected cases per tick DataFrame. Look up the PostProcessor docs to find the column definitions. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.detection_rate — Method

detection_rate(rd::ResultData)

Returns the fraction of detected infections. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.effectiveR — Method

effectiveR(rd::ResultData)

Returns the Effective R value over time DataFrame. Look up the PostProcessor docs to find the column definitions. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.execution_date — Method

execution_date(rd::ResultData)

Returns the timestamp of result data generation. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.execution_date_formatted — Function

execution_date_formatted(rd::ResultData)

Returns the (formatted for report) timestamp of result data generation. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.exportJLD — Method

exportJLD(rd::ResultData, directory::AbstractString)

Exports the ResultData object as a JLD2 file, storing it in the specified directory.

GEMS.final_tick — Method

final_tick(rd::ResultData)

Returns the tick counter at the end of the simulation run. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.free_mem_size — Method

free_mem_size(rd::ResultData)

Returns the available system memory

GEMS.GEMS_version — Method

GEMS_version(rd::ResultData)

Returns the GEMS version this ResultData object was generated with. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.git_branch — Method

git_branch(rd::ResultData)

Returns the current git branch.

GEMS.git_commit — Method

git_commit(rd::ResultData)

Returns the current git commit.

GEMS.git_repo — Method

git_repo(rd::ResultData)

Returns the current git repository.

GEMS.hashes — Method

hashes(rd::ResultData)

Returns the dataframes of result data. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.hospital_df — Method

hospital_df(rd::ResultData)

Returns the tests per tick DataFrame. Look up the PostProcessor docs to find the column definitions. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.household_attack_rates — Method

household_attack_rates(rd::ResultData)

Returns householdattackrates DataFrame. Look up the PostProcessor docs to find the column definitions. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.id — Method

id(rd::ResultData)

Returns the unique identifer of the ResultData object.

GEMS.import_resultdata — Function

import_resultdata(filepath::String)

Import the ResultData object from a jld2 file. Returns the ResultData object.

import_resultdata(filepath::String, config::Dict=Dict())

Import the ResultData object from a jld2 file. Also accepts a config dictionary that includes the fields that should be obtained from the file. If there are fields in the config file that are not yet present in the ResultData object the creation of these fields is being attempted. If there are fields present in the ResultData object that are not in the config file, these fields are ommited. Providing an empty config dictionary will lead to the generation of all fields.

GEMS.infections — Method

infections(rd::ResultData)

Returns the infection DataFrame joined with individuals' attributes. Look up the PostProcessor docs to find the column definitions. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.infections_hash — Method

infections_hash(rd::ResultData)

Returns a SHA1 hash value for the infections DataFrame based on the tick, id_a, and id_b column. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.info — Method

info(rd::ResultData)

Prints info about available fields in the ResultData object.

GEMS.initial_infections — Method

initial_infections(rd::ResultData)

Returns the number of individuals who are marked as infected during initialization. This happens before the actual simulation run. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.julia_version — Method

julia_version(rd::ResultData)

Returns the Julia version that was used to generate this result data object. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.kernel — Method

kernel(rd::ResultData)

Returns the system kernel information Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.label — Method

label(rd::ResultData)

Returns the lable of the simulation run. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.meta_data — Method

meta_data(rd::ResultData)

Returns the meta_data dictionary of the ResultData. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.model_size — Method

model_size(rd::ResultData)

Returns the size of the simulation model in memory.

GEMS.number_of_individuals — Method

number_of_individuals(rd::ResultData)

Returns the total number of individuals in the population model. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.observed_R — Method

observed_R(rd::ResultData)

Returns the observed reproduction number estimation DataFrame. Look up the PostProcessor docs to find the column definitions. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.pathogens — Method

pathogens(rd::ResultData)

Returns an array of pathogen parameters. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.population_file — Method

population_file(rd::ResultData)

Returns the path to the population file Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.population_params — Method

population_params(rd::ResultData)

Returns parameters that were used to generate the population. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.population_pyramid — Method

population_pyramid(rd::ResultData)

Returns the DataFrame required to plot population pyramid (age, sex, count) Look up the PostProcessor docs to find the column definitions. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.population_size — Method

population_size(rd::ResultData)

Returns the size of the population model in memory.

GEMS.region_info — Method

region_info(rd::ResultData)

Returns a Dataframe with population size and area per municiaplity (if model is geolocalized). Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.rolling_observed_SI — Method

rolling_observed_SI(rd::ResultData)

Returns the rolling observed serial interval DataFrame. Look up the PostProcessor docs to find the column definitions. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.setting_data — Method

setting_data(rd::ResultData)

Returns a DataFrame containing information on all setting types. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.setting_sizes — Method

setting_sizes(rd::ResultData)

Returns a Dictionary containing information on all setting sizes. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.sim_data — Method

sim_data(rd::ResultData)

Returns the sim_data of result data. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.someempty — Function

someempty(f::Function, rds::Vector{ResultData})

Returns true if the provided function returns an empty dictionary for at least one of the ResultData objects.

GEMS.start_condition — Method

start_condition(rd::ResultData)

Returns the StartCondition object the simulation was initialized with. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.stop_criterion — Method

stop_criterion(rd::ResultData)

Returns the StopCriterion object of the simulation. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.strategies — Method

strategies(rd::ResultData)

Returns the strategies included in the simulation. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.symptom_triggers — Method

symptom_triggers(rd::ResultData)

Returns the symptom triggers included in the simulation. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.system_data — Method

system_data(rd::ResultData)

Returns the system_data of result data. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.tests — Method

tests(rd::ResultData)

Returns the tests DataFrame . Look up the PostProcessor docs to find the column definitions. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.testtypes — Method

testtypes(rd::ResultData)

Returns the test types included in the simulation. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.threads — Method

threads(rd::ResultData)

Returns the number of threads this Julia instance was started with.

GEMS.tick_cases — Method

tick_cases(rd::ResultData)

Returns the infections per tick DataFrame. Look up the PostProcessor docs to find the column definitions. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.tick_cases_per_setting — Method

tick_cases_per_setting(rd::ResultData)

Returns the tests per tick DataFrame. Look up the PostProcessor docs to find the column definitions. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.tick_deaths — Method

tick_deaths(rd::ResultData)

Returns the deaths per tick DataFrame. Look up the PostProcessor docs to find the column definitions. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.tick_generation_times — Method

tick_generation_times(rd::ResultData)

Returns the generation times per tick DataFrame. Look up the PostProcessor docs to find the column definitions. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.tick_pooltests — Method

tick_pooltests(rd::ResultData)

Returns the pool tests per tick DataFrame. Look up the PostProcessor docs to find the column definitions. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.tick_serial_intervals — Method

tick_serial_intervals(rd::ResultData)

Returns the serial intervals per tick DataFrame. Look up the PostProcessor docs to find the column definitions. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.tick_tests — Method

tick_tests(rd::ResultData)

Returns the tests per tick DataFrame. Look up the PostProcessor docs to find the column definitions. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.tick_unit — Method

tick_unit(rd::ResultData)

Returns the unit of time that one tick corresponds to. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.tick_vaccinations — Method

tick_vaccinations(rd::ResultData)

Returns the vaccinations per tick DataFrame. Look up the PostProcessor docs to find the column definitions. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.time_to_detection — Method

time_to_detection(rd::ResultData)

Returns time to detection DataFrame. Look up the PostProcessor docs to find the column definitions. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.timer_output! — Method

timer_output!(rd::ResultData, timer_output::TimerOutput)

Sets the TimerOutput object for a ResultData object

GEMS.timer_output — Method

timer_output(rd::ResultData)

Returns the TimerOutput object used to supply debug report with execution time information Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.total_infections — Method

total_infections(rd::ResultData)

Returns the row count of the PostProcessors' infections-DataFrame. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.total_mem_size — Method

total_mem_size(rd::ResultData)

Returns the total system memory

GEMS.total_quarantines — Method

total_quarantines(rd::ResultData)

Returns the total quarantined agent over the course of the simulation. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.total_tests — Method

total_tests(rd::ResultData)

Returns a dictionary with the the total number of tests per TestType. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.word_size — Method

word_size(rd::ResultData)

Returns the system word size.

ResultDataStyle

Constructors

GEMS.ResultDataStyle — Type

ResultDataStyle

Abstract type, whose implementations define the structure of a ResutltData object.

GEMS.DefaultResultData — Type

DefaultResultData <: ResultDataStyle

The default style for ResultData objects. It contains all that can currently be calculated in the PostProcessor. Therefore, it is both, the most comprehensive and computationally intensive (memory & runtime) option.

Fields

data::Dict{String, Any}:

meta_data::Dict{String, Any}

timer_output::TimerOutput: TimerOutput object

execution_date::String: Time this ResultData object was generated

GEMS_version::VersionNumber: GEMS version this ResultData object was generated with

config_file::String: Path to the config file

config_file_val::Dict{String, Any}: Deep copy of the supplied TOML config file

population_file::String: Path to the population file

population_params::Dict{String, Any}: Parameters used to generate population

sim_data::Dict{String, Any}

label::String: Label of this simulation run (needed for plotting)

final_tick::Int16: Tick counter at the end of the simulation run

number_of_individuals::Int64: Total number of individuals in the population model

initial_infections::Int64: Number of initial infected individuals

total_infections::Int64: Row count of the PostProcessor's infections DataFrame

attack_rate::Float64: Fraction of overall infected individuals

setting_data::DataFrame: DataFrame containing information on all setting types

setting_sizes::Dict{Any, Any}: Dictionary containing the setting sizes distributions for all included settingtypes

region_info::Dataframe: Municipality population size and area (if geolocalized model is used)

pathogens::Vector{Pathogen}: Array of pathogen parameters

vaccine::Vaccine: Vaccine parameter [CURRENTLY DEACTIVATED]

vaccination_strategy::VaccineScheduler: Vaccination strategy used in this model [CURRENTLY DEACTIVATED]

tick_unit::String: Unit of time that one tick corresponds to

start_condition::StartCondition: Initial setup of the simulation

stop_criterion::StopCriterion: Termination conditions

strategies::Vector{Strategy}: Intervention strategies

symptom_triggers::Vector{ITrigger}: Strategies that are triggered upon experiencing symptoms

testtypes::Vector{AbstractTestType}: Test types used in the model (e.g. Antigen Tests)

total_quarantines::Int64: Total person-ticks (e.g. days) spent in isolation

total_tests::Dict{Any, Any}: Total number of performed tests per TestType

detection_rate::Float64: Fraction of detected infections (by testing)

system_data::Dict{String, Any}

kernel::String: System kernel

julia_version::String: Julia version that was used to generate this data object

word_size::Int64: System word size

threads::Int64: Number of threads this Julia instance was started with

cpu_data::Markdown.MD: Information on the processor (not available for ARM Macs)

total_mem_size::Float64: Total system memory

free_mem_size::Float64: Available system memory

git_repo::SubString{String}: Current Git repository

git_branch::SubString{String}: Current Git branch

git_commit::SubString{String}: Current Git commit ID

model_size::Int64: Size of the simulation model in memory [CURRENTLY DACTIVATED]

population_size::Int64: Size of the population model in memory [CURRENTLY DACTIVATED]

setting_age_contacts::Dict{String, Any}

Household: age X age contact matrix for Households based on sampled data [CURRENTLY DACTIVATED]

GlobalSetting: age X age contact matrix for GlobalSettings based on sampled data [CURRENTLY DACTIVATED]

aggregated_setting_age_contacts::Dict{String, Any}

Household::ContactMatrix{Float64}: age group x age group "ContactMatrix" object for Households based on sampled data

SchoolClass::ContactMatrix{Float64}: age group x age group "ContactMatrix" object for SchoolClass based on sampled data

School::ContactMatrix{Float64}: age group x age group "ContactMatrix" object for School based on sampled data

SchoolComplex::ContactMatrix{Float64}: age group x age group "ContactMatrix" object for SchoolComplex based on sampled data

Office::ContactMatrix{Float64}: age group x age group "ContactMatrix" object for Office based on sampled data

Department::ContactMatrix{Float64}: age group x age group "ContactMatrix" object for Department based on sampled data

Workplace::ContactMatrix{Float64}: age group x age group "ContactMatrix" object for Workplace based on sampled data

WorkplaceSite::ContactMatrix{Float64}: age group x age group "ContactMatrix" object for WorkplaceSite based on sampled data

Municipality::ContactMatrix{Float64}: age group x age group "ContactMatrix" object for Municipality based on sampled data

GlobalSetting::ContactMatrix{Float64}: age group x age group "ContactMatrix" object for GlobalSetting based on sampled data

dataframes::Dict{String, Any}

infections::DataFrame: Infection data joined with individuals' attributes

vaccinations::DataFrame: Vaccination data joined with individuals' attributes

deaths::DataFrame: Death data joined with individuals' attributes

tests::DataFrame: Test data joined with individuals' attributes

effectiveR::DataFrame: Effective R value over time

compartment_periods::DataFrame: Duration of exposed and infectious states for all infections

aggregated_compartment_periods::DataFrame: Statistics on time individuals spend in each disease compartment

tick_cases::DataFrame: Infections per tick

tick_deaths::DataFrame: Deaths per tick

tick_vaccinations::DataFrame: Vaccinations per tick

tick_serial_intervals::DataFrame: Aggregated data on serial intervals per tick

tick_generation_times::DataFrame: Aggregated data on generation timess per tick

tick_tests::DataFrame: Number of tests performed per tick

tick_pooltests::DataFrame: Number of (pooled) tests per tick

tick_cases_per_setting::DataFrame: Tick cases aggregated by settingtype,

detected_tick_cases::DataFrame: Number of detected infections per tick

compartment_fill::DataFrame: Number of individuals currently in any of the disease compartments

cumulative_cases::DataFrame: Cumulative infections over time

cumulative_deaths::DataFrame: Cumulative deaths over time

cumulative_vaccinations::DataFrame: Cumulative vaccinations over time

cumulative_disease_progressions::DataFrame: Cumulative information on disease states N ticks after exposure

cumulative_quarantines::DataFrame: Number of quarantined individuals per tick

age_incidence::DataFrame: Incidence over time stratified by age groups

population_pyramid::DataFrame: Data required to plot population pyramid (age, sex, count)

rolling_observed_SI::DataFrame: Serial interval estimation based on the last 14 days of detected cases

observed_R::DataFrame: Reproduction number estimation based on detected cases and the SI estimation

hospital_df::DataFrame: DataFrame containing the daily hospitalizations etc.

time_to_detection::DataFrame: Statistics on the time between exposure and first detection of an infection through a test

household_attack_rates::DataFrame: Statistics on the seconary infections in households

customlogger::DataFrame: Dataframe obtained from any custom logger that might have been set

GEMS.LightRD — Type

LightRD <: ResultDataStyle

Similar to the default style for ResultData objects but without any raw data (e.g. the ìnfections-, deaths- or tests- dataframes) as the raw data makes around 80% of the DefaultResultData style memory footprint. It contains everything that can currently be calculated in the PostProcessor.

This RD-style cannot be used to generate geographical maps or infection videos.

Fields

data::Dict{String, Any}: Dictionary holding the following sub-dictionaries

meta_data::Dict{String, Any}

timer_output::TimerOutput: TimerOutput object

(Note: This data is only available if the simulation runs were done via the main() function)

execution_date::String: Time this ResultData object was generated

GEMS_version::VersionNumber: GEMS version this ResultData object was generated with

config_file::String: Path to the config file

config_file_val::Dict{String, Any}: Deep copy of the supplied TOML config file

population_file::String: Path to the population file

population_params::Dict{String, Any}: Parameters used to generate population

sim_data::Dict{String, Any}

label::String: Label of this simulation run (needed for plotting)

final_tick::Int16: Tick counter at the end of the simulation run

number_of_individuals::Int64: Total number of individuals in the population model

initial_infections::Int64: Number of initial infected individuals

total_infections::Int64: Row count of the PostProcessor's infections DataFrame

attack_rate::Float64: Fraction of overall infected individuals

setting_data::DataFrame: DataFrame containing information on all setting types

setting_sizes::Dict{Any, Any}: Dictionary containing the setting sizes distributions for all included settingtypes

region_info::Dataframe: Municipality population size and area (if geolocalized model is used)

pathogens::Vector{Pathogen}: Array of pathogen parameters

tick_unit::String: Unit of time that one tick corresponds to

start_condition::StartCondition: Initial setup of the simulation

stop_criterion::StopCriterion: Termination conditions

strategies::Vector{Strategy}: Intervention strategies

symptom_triggers::Vector{ITrigger}: Strategies that are triggered upon experiencing symptoms

testtypes::Vector{AbstractTestType}: Test types used in the model (e.g. Antigen Tests)

total_quarantines::Int64: Total person-ticks (e.g. days) spent in isolation

total_tests::Dict{Any, Any}: Total number of performed tests per TestType

detection_rate::Float64: Fraction of detected infections (by testing)

system_data::Dict{String, Any}

kernel::String: System kernel

julia_version::String: Julia version that was used to generate this data object

word_size::Int64: System word size

threads::Int64: Number of threads this Julia instance was started with

cpu_data::Markdown.MD: Information on the processor (not available for ARM Macs)

total_mem_size::Float64: Total system memory

free_mem_size::Float64: Available system memory

git_repo::SubString{String}: Current Git repository

git_branch::SubString{String}: Current Git branch

git_commit::SubString{String}: Current Git commit ID

model_size::Int64: Size of the simulation model in memory [CURRENTLY DACTIVATED]

population_size::Int64: Size of the population model in memory [CURRENTLY DACTIVATED]

aggregated_setting_age_contacts::Dict{String, Any}

Household::ContactMatrix{Float64}: age group x age group "ContactMatrix" object for Households based on sampled data

SchoolClass::ContactMatrix{Float64}: age group x age group "ContactMatrix" object for SchoolClass based on sampled data

School::ContactMatrix{Float64}: age group x age group "ContactMatrix" object for School based on sampled data

SchoolComplex::ContactMatrix{Float64}: age group x age group "ContactMatrix" object for SchoolComplex based on sampled data

Office::ContactMatrix{Float64}: age group x age group "ContactMatrix" object for Office based on sampled data

Department::ContactMatrix{Float64}: age group x age group "ContactMatrix" object for Department based on sampled data

Workplace::ContactMatrix{Float64}: age group x age group "ContactMatrix" object for Workplace based on sampled data

WorkplaceSite::ContactMatrix{Float64}: age group x age group "ContactMatrix" object for WorkplaceSite based on sampled data

Municipality::ContactMatrix{Float64}: age group x age group "ContactMatrix" object for Municipality based on sampled data

GlobalSetting::ContactMatrix{Float64}: age group x age group "ContactMatrix" object for GlobalSetting based on sampled data

dataframes::Dict{String, Any}

effectiveR::DataFrame: Effective R value over time

tick_cases::DataFrame: Infections per tick

tick_deaths::DataFrame: Deaths per tick

tick_serial_intervals::DataFrame: Aggregated data on serial intervals per tick

tick_generation_times::DataFrame: Aggregated data on generation timess per tick

tick_tests::DataFrame: Number of tests performed per tick

tick_pooltests::DataFrame: Number of (pooled) tests per tick

tick_cases_per_setting::DataFrame: Tick cases aggregated by settingtype,

detected_tick_cases::DataFrame: Number of detected infections per tick

compartment_fill::DataFrame: Number of individuals currently in any of the disease compartments

aggregated_compartment_periods::DataFrame: Statistics on time individuals spend in each disease compartment

cumulative_cases::DataFrame: Cumulative infections over time

cumulative_deaths::DataFrame: Cumulative deaths over time

cumulative_disease_progressions::DataFrame: Cumulative information on disease states N ticks after exposure

cumulative_quarantines::DataFrame: Number of quarantined individuals per tick

age_incidence::DataFrame: Incidence over time stratified by age groups

population_pyramid::DataFrame: Data required to plot population pyramid (age, sex, count)

rolling_observed_SI::DataFrame: Serial interval estimation based on the last 14 days of detected cases

observed_R::DataFrame: Reproduction number estimation based on detected cases and the SI estimation

hospital_df::DataFrame: DataFrame containing the daily hospitalizations etc.

time_to_detection::DataFrame: Statistics on the time between exposure and first detection of an infection through a test

household_attack_rates::DataFrame: Statistics on the seconary infections in households

customlogger::DataFrame: Dataframe obtained from any custom logger that might have been set