Simulation

alles aus setup folder configuration des sim objects

Overview Structs

GEMS.AbstractScheduler

GEMS.DiseaseProgressionStrat

GEMS.DiseaseProgressionStrat

GEMS.HospitalizationTrigger

GEMS.InfectedFraction

GEMS.NoneInfected

GEMS.Parameter

GEMS.Runinfo

GEMS.Simulation

GEMS.Simulation

GEMS.Simulation

GEMS.StartCondition

GEMS.StopCriterion

GEMS.TimesUp

Overview Functions

GEMS.add_hospitalization_trigger!

GEMS.add_strategy!

GEMS.add_symptom_trigger!

GEMS.add_testtype!

GEMS.add_tick_trigger!

GEMS.configfile

GEMS.create_contact_sampling_method

GEMS.create_transmission_function

GEMS.customlogger

GEMS.customlogger!

GEMS.customlogs

GEMS.deathlogger

GEMS.deaths

GEMS.dequeue!

GEMS.enqueue!

GEMS.evaluate

GEMS.event_queue

GEMS.fire_custom_loggers!

GEMS.fraction

GEMS.hospitalization_triggers

GEMS.incidence

GEMS.increment!

GEMS.infections

GEMS.info

GEMS.initial_infections

GEMS.initialize!

GEMS.interval

GEMS.label

GEMS.limit

GEMS.main

GEMS.main_r

GEMS.model_size

GEMS.parameters

GEMS.pathogen

GEMS.pathogen!

GEMS.pooltestlogger

GEMS.pooltests

GEMS.population

GEMS.populationfile

GEMS.quarantinelogger

GEMS.quarantines

GEMS.region_info

GEMS.remove_empty_settings!

GEMS.reset!

GEMS.run!

GEMS.schedule!

GEMS.settingfile_sim

GEMS.settingfile_test

GEMS.should_fire

GEMS.spread_infection!

GEMS.start_condition

GEMS.step!

GEMS.stepmod

GEMS.stop_criterion

GEMS.switch_tick

GEMS.symptom_triggers

GEMS.test_sim

GEMS.test_sim_r

GEMS.test_state

GEMS.testlogger

GEMS.tests

GEMS.testtypes

GEMS.tick

GEMS.tick_triggers

GEMS.ticks

GEMS.tickunit

GEMS.try_to_infect!

GEMS.vaccinate!

GEMS.waning

Structs

GEMS.AbstractScheduler — Type

Supertype for all Scheduler Structs

Missing docstring.

Missing docstring for Asymptomatic. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Critical. Check Documenter's build log for details.

GEMS.DiseaseProgressionStrat — Type

DiseaseProgressionStrat <: AgeStratification

A wrapper type for an age stratification matrix regarding the disease progression.

Fields

age_groups::Vector{String}: A list of the age groups in order of the rows of the stratification matrix. The first entry in age_groups belongs to the first row in the matrix.

disease_compartments::Vector{String}: A list of the compartments/terminal states in the disease progression in order of the columns of the stratification matrix. First entry in disease_compartments belongs to the first column in the stratification matrix.

stratification_matrix::Vector{Vector{T}} where T <: Real: The matrix that holds the probabilities for the age stratified disease progression. The entries of each row have to add up to 1.

GEMS.HospitalizationTrigger — Type

HospitalizationTrigger <: ITrigger

A struct defining an IStrategy that shall be fired upon an individual's admittance to hospital.

Missing docstring.

Missing docstring for Mild. Check Documenter's build log for details.

GEMS.NoneInfected — Type

NoneInfected <: StopCriterion

A StopCriterion that stops the simulation once no individual is infected.

GEMS.Parameter — Type

Supertype for all simulation parameters

GEMS.Runinfo — Type

Runinfo

An immutable struct providing information on a given run (instance) of the simulation. The main function stores this data into the result output directory (as TOML). The internal data structure is a Dict to facilitate TOML-conversion. Currently containing:

date

configfile

populationfile

Missing docstring.

Missing docstring for Severe. Check Documenter's build log for details.

GEMS.Simulation — Type

Simulation

A struct for the management of a single run, holding all necessary informations.

Fields

Data Sources

configfile::String: Path to config file

General

tick::Int16: Current tick/timestep

tickunit::Char: Time unit of one simulation step (tick)

startdate::Date: Start date of the simulation

enddate::Date: End date of the simulation

start_condition::StartCondition: Starting condition that sets the initial situation

stop_criterion::StopCriterion: Criterion that terminates a simulation run

label::String: Label for plot visualizations

Model

population::Population: Container to hold all present individuals

settings::SettingsContainer: All settings present in the simulation

pathogen::Pathogen: The pathogen of which infections are simulated

Logger

infectionlogger::InfectionLogger: A logger tracking all infections

deathlogger::DeathLogger: A logger specifically for the deaths of individuals

testlogger::TestLogger: A logger tracking all individual tests

pooltestlogger::PoolTestLogger: A logger tracking all pool tests

quarantinelogger::QuarantineLogger: A tracking cumulative quarantines per tick

customlogger::CustomLogger: A logger running custom methods on the Simulation object in each tick

Interventions

symptom_triggers::Vector{ITrigger}: List of all SymptomTriggers

tick_triggers::Vector{TickTrigger}: List of all TickTriggers

hospitalization_triggers::Vector{ITrigger}: List of all HospitalizationTriggers

event_queue::EventQueue: Event Queue to apply intervention measures

strategies::Vector{Strategy}: List of all registered intervention strategies

testtypes::Vector{AbstractTestType}: List of all TestTypes (e.g. Antigen- or PCR-Test)

Runtime Modifiers

stepmod::Function: Single-argment function that runs custom code on the simulation object in each tick

GEMS.StartCondition — Type

supertype for all start conditions

GEMS.StopCriterion — Type

supertype for all stop criteria

Missing docstring.

Missing docstring for SymptomCategory. Check Documenter's build log for details.

GEMS.TimesUp — Type

TimesUp <: StopCriterion

A StopCriterion that specifies a time limit.

Fields

limit::Int16: A time limit. When reached, the simulation should be terminated.

GEMS.waning — Function

waning(vaccine)

Returns the waning struct associated with the vaccine.

Constructors

GEMS.DiseaseProgressionStrat — Method

DiseaseProgressionStrat()

A constructor for a default Disease Progression, where every individual will only be assigned to be asymptomatic. Mainly for testing purposes.

GEMS.Simulation — Method

Simulation(config_file::String, start_condition::StartCondition, stop_criterion::StopCriterion, population::Population, settings::SettingsContainer, label::String)

Used as initializer and is called by the other constructors.

GEMS.Simulation — Method

Simulation(; simargs...)

Constructor that creates and initializes a Simulation object based on various parameters.

It uses the default configuration for the created run which is saved in data/DefaultConf.toml but should not be edited. See Usage and Examples on how to edit the simulation configuration.

The constructor can be called without any parameters which will create a default simulation.

A population can be provided. This can either be the path to a population file (.csv or .jdl2) or a Population object created. Either way the poulation will overwrite any population configurations in the config file.

Arguments can be provided overwriting just certain parameters. For all possible arguments see the table below. These arguments can be used in combination with a custom config file and will overwrite it where possible. Arguments can be combined.

All arguments can alse be put into Dictionary and provided to function. This dictionary must have Symbols as keys. The keys must be the same as the argument names.

All possibilities above can also be combined. Providing arguments will overwrite custom config file where possible. Custom populations will override custom arguments.

The optional stepmod argument allows to pass a custom single-argument function that takes the simulation object as its argument and allows to do custom operations on the simulation object in each step.

Note: Be careful with the stepmod option as there's no option to check whether your custom code invalidates simulation outputs or causes internal model inconsistencies. Please only use this option if you are sure you know what you're doing :-)

Examples

sim = Simulation() # returns a default simulation 

sim = Simulation(population = "data/TestPop.csv") # returns a Simulation using the assigned population file.

sim = Simulation(population = Population()) # returns a Simulation using the created Population.

my_population = Population(n=1_000);
sim = Simulation(population = my_population) # returns a Simulation using the created Population that has its own parameters.

sim = Simulation(label = "My First Simulation") # returns a Simulation with a custom label

sim = Simulation(startdate = "2020.1.1", n = 100_000, transmission_rate = 0.2) # returns a Simulation starting at the 1.1.2020, having 100,000 individuals and a transmission rate of 0.2

my_arguments = Dict(
                    :transmission_rate => 0.3,
                    :pop_size => 10_000_000
                )
sim = Simulation(my_arguments) # returns a `Simulation` the arguments provided in the dictionary.

sim = Simulation(population = Population(n=100_000), label = "My First Simulation") # returns a Simulation with a custom label and a population partly overwriting the default configuration.

Parameters

Name	Type	Description

population	String, Population	Path to a CSV or JLD2 file containing popultion data or a Population object

settingsfile	String	Path to a setting-file

label	String	Label used to name simulation runs during analysis

stepmod	Function	One-agument function which will be executed each simulation step, called on the Simulation object

seed	Int64	Random seed

global_setting	Bool	Enable or disable a global setting that contains every individual

startdate	Date, String	Simulation start date (format: YYYY.MM.DD)

enddate	Date, String	Simulation end date (format: YYYY.MM.DD)

infected_fraction	Float64	Fraction of the initially infected agents for the InfectedFraction start condition

transmission_rate	Float64	Infection probability (0-1) during a contact where one individual is infectious

onset_of_symptoms	Float64	Average time from exposure to onset of symptoms (Poisson-distributed)

onset_of_severeness	Float64	Average time from onset of symptoms to onset of severeness (Poisson-distributed)

infectious_offset	Float64	Average time from onset of infectiousness to onset of symptoms (Poisson-distributed); cannot be before exposure, obviously.

mild_death_rate	Float64	Probability of dying (0-1) with a mild disease progression

severe_death_rate	Float64	Probability of dying (0-1) with a severe disease progression

critical_death_rate	Float64	Probability of dying (0-1) with a critical disease progression

hospitalization_rate	Float64	Probability of being hospitalized (0-1) with a severe disease progression

ventilation_rate	Float64	Probability of being ventilated (0-1) with a critical disease progression

icu_rate	Float64	Probability of being admitted to ICU (0-1) with a critical disease progression

time_to_recovery	Float64	Average time from onset of symptoms to recovery (Poisson-distributed)

time_to_hospitalization	Float64	Average time from onset of symptoms to hospitalization (Poisson-distributed)

time_to_icu	Float64	Average time from hospitalization to ICU-addmitance (Poisson-distributed)

length_of_stay	Float64	Average duration of hospitalization (Poisson-distributed)

progression_categories	Vector{Float64}	Four-value vector indicating the fraction of individuals with an asymptomatic, mild, severe, and critical progression (must sum to 1)

office_contact_rate	Float64	Average number of office contacts per timestep (Poisson-distributed)

household_contact_rate	Float64	Average number of household contacts per timestep (Poisson-distributed)

school_contact_rate	Float64	Average number of school contacts per timestep (Poisson-distributed)

schoolClass_contact_rate	Float64	Average number of school-class contacts per timestep (Poisson-distributed)

schoolYear_contact_rate	Float64	Average number of school-year contacts per timestep (Poisson-distributed)

schoolComplex_contact_rate	Float64	Average number of school-complex contacts per timestep (Poisson-distributed)

workplaceSite_contact_rate	Float64	Average number of workplace-site contacts per timestep (Poisson-distributed)

workplace_contact_rate	Float64	Average number of workplace contacts per timestep (Poisson-distributed)

department_contact_rate	Float64	Average number of department contacts per timestep (Poisson-distributed)

municipality_contact_rate	Float64	Average number of municipality contacts per timestep (Poisson-distributed)

global_contact_rate	Float64	Average number of contacts in the global per timestep (Poisson-distributed)

Population Parameters

pop_size	Int64	Number of individuals in the population

avg_household_size	Int64	Average size of households

avg_office_size	Int64	Average size of offices

avg_school_size	Int64	Average size of schools

Missing docstring.

Missing docstring for Simulation(::String, ::String, ::String, ::String, ::Function, ::Dict). Check Documenter's build log for details.

Functions

GEMS.add_hospitalization_trigger! — Function

add_hospitalization_trigger!(simulation, trigger)

Adds a HospitalizationTrigger to the simulation.

GEMS.add_strategy! — Function

add_strategy!(simulation, strategy)

Adds an intervention Strategy to the simulation object. A strategy must be added to the simulation object to make it appear in the report. In order to execute a strategy during the simulation run, you must define a Trigger and link this strategy. Just adding it here will not execute the strategy.

GEMS.add_symptom_trigger! — Function

add_symptom_trigger!(simulation, trigger)

Adds a SymptomTrigger to the simulation.

GEMS.add_testtype! — Function

add_testtype!(simulation, testtype)

Adds a test type to the simulation.

GEMS.add_tick_trigger! — Function

add_tick_trigger!(simulation, trigger)

Adds a TickTrigger to the simulation.

GEMS.configfile — Method

configfile(simulation)

Returns configfile that was used to initialize simulation.

Missing docstring.

Missing docstring for create_simulation. Check Documenter's build log for details.

Missing docstring.

Missing docstring for create_waning. Check Documenter's build log for details.

GEMS.customlogger! — Method

customlogger(simulation, customlogger)

Sets the Simulation's CustomLogger.

Missing docstring.

Missing docstring for customlogger(::Simulation, ::CustomLogger). Check Documenter's build log for details.

GEMS.customlogger — Method

customlogger(simulation)

Returns the CustomLogger of the simulation.

GEMS.customlogs — Function

customlogs(simulation::Simulation)

Calls the dataframe() function on the internal simulation's CustomLogger.

GEMS.deaths — Method

deaths(simulation::Simulation)

Calls the dataframe() function on the internal simulation's DeathLogger.

GEMS.deathlogger — Method

deathlogger(simulation)

Returns the DeathLogger of the simulation.

GEMS.dequeue! — Method

dequeue!(queue::EventQueue)

Removes and returns the first Event of the EventQueue.

GEMS.enqueue! — Method

enqueue!(queue::EventQueue, event::Event, tick::Int16)

Adds a new Event to the EventQueue at the specified tick.

GEMS.evaluate — Function

evaluate(simulation, criterion)

Evaluates whether the specified stop criterion is met for the simulation model. Return True if criterion was met. This is an 'abstract' function that must be implemented for concrete criterion types.

evaluate(simulation, criterion)

Returns true if specified termination tick has been met.

evaluate(simulation, criterion)

Returns true if none of the individuals are infected.

GEMS.event_queue — Method

event_queue(simulation)

Returns the simulation's intervention event queue.

Missing docstring.

Missing docstring for final_tick(::ResultData). Check Documenter's build log for details.

GEMS.fire_custom_loggers! — Method

fire_custom_loggers!(sim::Simulation)

Executes all custom functions that are stored in the CustomLogger on the Simulation object and stores them in the internal dataframe.

GEMS.hospitalization_triggers — Method

hospitalization_triggers(simulation)

Returns the list of HospitalizationTriggers registered in the simulation.

GEMS.incidence — Function

incidence(simulation::Simulation, pathogen::Pathogen, base_size::Int = 100_000, duration::Int16 = Int16(7))

Returns the incidence at a particular pathogen and a point in time (current simulation tick). The duration defines a time-span for which the incidence is measured (default: 7 ticks). The base_size provides the population size reference (default: 100_000 individuals)

Parameters

simulation::Simulation: Simulation object

pathogen::Pathogen: Pathogen for which the incidence shall be calculated

base_size::Int = 100_000 (optional): Reference popuation size for incidence calculation

duration::Int16 = Int16(7) (optional): Reference duration (in ticks) for the incidence calculation

Returns

Float64: Incidence

incidence(simulation::Simulation, base_size::Int = 100_000, duration::Int16 = Int16(7))

Returns the incidence at a particular point in time (current simulation tick). NOTE: This will only work for the single-pathogen version. The duration defines a time-span for which the incidence is measured (default: 7 ticks). The base_size provides the population size reference (default: 100_000 individuals)

Parameters

simulation::Simulation: Simulation object

base_size::Int = 100_000 (optional): Reference popuation size for incidence calculation

duration::Int16 = Int16(7) (optional): Reference duration (in ticks) for the incidence calculation

Returns

Float64: Incidence

GEMS.increment! — Method

increment!(simulation)

Increments the current simulation's tick counter by 1.

GEMS.info — Method

info(sim::Simulation)

Summary output for Simulation object configuration.

GEMS.interval — Function

interval(trigger::TickTrigger)

Returns the interval associated with a TickTrigger.

GEMS.label — Method

label(simulation::Simulation)

Returns simulation object's string label.

GEMS.main — Function

main(configfile::String, populationfile::String, outputdir::String;
settingfile::String="", report::Bool = false, csv::Bool = false,
popexport::Bool = false, batch::Bool = false, seed = missing, 
model_validation::Bool = true,  mod!::Function = x -> x, stepmod!::Function = x -> x,
customlogger::CustomLogger = CustomLogger(), with_progressbar::Bool = true)

Runs the full simulation pipeline, including:

Initialization (loading config- and population files)

Model validation

Running the simulation

Exporting raw data in provided output folder

Post processing

Exporting processed data

Generating report and exporting it to provided folder (indicated by optional boolean flag report. Default is false)

Parameters

configfile::String: Path to the configuration file.

populationfile::String: Path to the population file.

outputdir::String: Path to the output directory.

settingfile::String = "" (optional): Path to the setting file. No setting file is used if default ("") is provided.

report::Bool = false (optional): Indicates whether to generate and export a report (default is false).

csv::Bool = false (optional): Indicates whether raw data from loggers shall be exported in CSV format

popexport::Bool = false (optional): Indicates whether a CSV of the population shall be exported into the provided result folder.

batch::Bool = false (optional): Indicates whether the function is being run in batch mode, i.e. data exportation is limited to the resultdata, configfile and runinfo (default is false).

seed = missing (optional): Seed value for random number generation (default is missing).

model_validation::Bool = true (optional): Indicates whether to perform model validation (default is true).

mod!::Function = x -> x (optional): Function for the modding of the simulation object (default is x -> x). The function will be applied to the simulation object after its construction and before the model execution.

stepmod!::Function = x -> x (optional): Function for custom logic added to the simulation object in each step (default is x -> x). The function will be applied to the simulation object in the end of each simulation step.

customlogger::CustomLogger = CustomLogger() (optional): Can be used to add custom logging mechanics that are being executed each step. Lookup CustomLogger docs for more information.

with_progressbar::Bool = true (optional): Flag indicating, if the run should display the ProgressBar.

Output Files

The main() function generates multiple folders and files in the output directory (outputdir) if all options are switched on:

/config: Contains a copy of the used config file

/img: Contains plots as .png images generated for the report

/output_JLD2: Contains raw output data of the simulation loggers

/output_processed: Contains the ResultData object

report.html: Simulation report in HTML format

report.md: Simulation report in Markdown format

report.pdf: Simulation report in PDF format

runinfo.toml: Simulation run meta data

Returns

ResultData: Result data object.

main(args)

Outer main()-function to run GEMS simulation from console. Provide parameters as String array in this order:

Path to config file

Path to population file

Path to output directory

Path to setting file (optional)

This function calls the inner main()-function.

GEMS.main_r — Function

main_r(configfile::String, popidentifier::String, outputdir::String;
    report::Bool = false, csv::Bool = false, popexport::Bool = false,
    batch::Bool = false, seed = missing, model_validation::Bool = true,
    mod!::Function = x -> x, stepmod!::Function = x -> x,
    customlogger::CustomLogger = CustomLogger())

The "_r" suffix is for remote!

A wrapper for the main()-function to specify population- and setting files using a popidentifier. Downloads the specified files from the population file server (currently on Münster University Sciebo Cloud) and store them locally. If a local copy is already available, the download will be skipped. Identifiers can be German state abbreviations (e.g. "NRW"), depending on the availability of the respective population files on the server.

For further documentation of the arguments and inner workings, please look up the main() functions documentation.

Returns

ResultData: Result data object.

GEMS.model_size — Method

model_size(rd::ResultData)

Returns the size of the simulation model in memory.

GEMS.parameters — Function

parameters(d::Distribution)

Returns a dictionary containing the string() of the distribution, the mean and the std.

parameters(dps::DiseaseProgressionStrat)::Dict

Obtains a dictionary containing the parameters of the DiseaseProgressionStrat.

parameters(transfunc::TransmissionFunction)

Fallback function for transmission functions that returns the type of the transmission function as an entry in a dictionary

parameters(transfunc::AgeDependentTransmissionRate)

Returns the parameters of the AgeDependentTransmissionRate as a dictionary.

parameters(transfunc::ConstantTransmissionRate)

Returns the parameters of the ConstantTransmissionRate as a dictionary.

parameters(pathogen::Pathogen):::Dict

Obtains a dictionary containing the parameters of the Pathogen. Distributions and the disease progression are formatted using their parameters function.

parameters(v::Vaccine)::Dict

Returns a dictionary containing the parameters of the vaccine.

parameters(s::StartCondition)

Returns an empty dictionary.

parameters(inffrac::InfectedFraction)

Returns a dictionary containing the parameters of the infected fraction start condition.

parameters(p0::PatientZero)

Returns a dictionary containing the parameters of the patient zero start condition.

parameters(p0::PatientZero)

Returns a dictionary containing the parameters of the patient zero start condition.

parameters(criterion::NoneInfected)

Fallback parameters function for the StopCriterion.

parameters(criterion::TimesUp)

Returns a dictionary with the type and limit of the criterion.

parameters(criterion::NoneInfected)

Returns a dictionary with the type of the NoneInfected stop criterion.

GEMS.pathogen! — Method

pathogen!(simulation, pathogen)

Sets the pathogen of the simulation.

GEMS.pathogen — Function

pathogen(simulation)

Retursn the pathogen of the simulation.

pathogen(patientZero)

Returns pathogen used to infect individuals at the beginning in this start condition.

pathogen(patientZeros)

Returns pathogen used to infect individuals at the beginning in this start condition.

pathogen(infectedFraction)

Returns pathogen used to infect individuals at the beginning in this start condition.

pathogen(tt::TestType)

Returns the TestType's associated pathogen.

GEMS.pooltests — Function

pooltests(simulation::Simulation)

Calls the dataframe() function on the internal simulation's PoolTestLogger.

GEMS.pooltestlogger — Function

pooltestlogger(simulation)

Returns the PoolTestLogger of the simulation.

GEMS.population — Method

population(simulation)

Returns the population associated with the simulation run.

GEMS.populationfile — Method

populationfile(simulation)

Returns populationfile that was used to initialize simulation.

Missing docstring.

Missing docstring for process_events!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for process_events!(::Simulation). Check Documenter's build log for details.

GEMS.quarantinelogger — Method

quarantinelogger(simulation)

Returns the QuarantineLogger of the simulation.

GEMS.quarantines — Function

quarantines(simulation::Simulation)

Calls the dataframe() function on the internal simulation's QuarantineLogger.

GEMS.remove_empty_settings! — Method

remove_empty_settings!(sim::Simulation)

Removes all settings that have no individuals associated with them. In the end update all ids to ensure that they are still correspond to the entry in the settings container.

GEMS.reset! — Method

reset!(simulation)

Resets the current simulation's tick counter to 0.

GEMS.region_info — Method

region_info(sim::Simulation)

Returns a DataFrame containing information about the Municipalitys in the model with the following columns:

Name	Type	Description

ags	AGS	Amtlicher Gemeindeschlüssel (Community Identification Code)

pop_size	Int64	Number of individuals in that municipality

area	Float64	Area size of this municipality in km²

Note: This function will download the Germany shapefile, if it's not available locally, and return missing values for pop_size and area if the download cannot be completed.

GEMS.run! — Method

run!(simulation::Simulation; with_progressbar::Bool = true)

Takes and initializes Simulation object and calls the stepping function (step!) until the stop criterion is met.

Returns

Simulation: Simulation object

GEMS.schedule! — Function

schedule!(scheduler, individual, vaccine, tick)

Schedules the individual for vaccination at time tick with the vaccine.

schedule!(simulation, vaccine, strategy)

Schedules the population of the simulation according to the provided strategy for vaccination with vaccine.

schedule!(simulation, vaccine, dailydosestrategy)

Schedules the population of the simulation with a fixed amount of vaccine doses per day defined in dailydoses(dailydosestrategy) and starting from available_from(dailydosestrategy).

Missing docstring.

Missing docstring for scheduled(scheduler, tick). Check Documenter's build log for details.

GEMS.settingfile_sim — Function

settingfile_sim()

Returns a simulation object with data from the full population file and settingfile from Münster.

GEMS.settingfile_test — Function

settingfile_test()

Runs a simulation with the population file and settingfile from Münster incl. the post-processing and report generation and returns the resultdata object.

Missing docstring.

Missing docstring for setup. Check Documenter's build log for details.

Missing docstring.

Missing docstring for setup_r. Check Documenter's build log for details.

GEMS.should_fire — Function

should_fire(trigger::TickTrigger, tick::Int16)

Evaluates whether a trigger should be fired at a given tick. Considers switch_tick and interval.

Returns

Bool: True, if the trigger should fire at this tick. False otherwise.

Missing docstring.

Missing docstring for Simulation(). Check Documenter's build log for details.

GEMS.spread_infection! — Method

spread_infection!(setting::Setting, sim::Simulation, pathogen::Pathogen)

Spreads the infection of pathogen inside the provided setting. This will simulate the infection dynamics at the time tick(sim) inside setting within the context of the simulation sim. This will also update all settings, the individual is part of, if the infection is successful.

Parameters

setting::Setting: Setting in which the pathogen shall be spreaded

sim::Simulation: Simulation object

pathogen::Pathogen: Pathogen to spread

GEMS.start_condition — Function

start_condition(simulation)

Returns start condition associated with the simulation run.

start_condition(rd::ResultData)

Returns the StartCondition object the simulation was initialized with. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.step! — Function

step!(simulation::Simulation)

Increments the simulation status by one tick and executes all events that shall be handled during this tick.

Missing docstring.

Missing docstring for steps. Check Documenter's build log for details.

GEMS.stepmod — Function

stepmod(simulation::Simulation)

Returns the defined step mod.

GEMS.stop_criterion — Method

stop_criterion(simulation)

Returns stop criterion associated with the simulation run.

GEMS.switch_tick — Function

switch_tick(trigger::TickTrigger)

Returns the switch_tick associated with a TickTrigger.

GEMS.symptom_triggers — Method

symptom_triggers(simulation)

Returns the list of SymptomTriggers registered in the simulation.

GEMS.test_sim — Function

test_sim()

Returns a simulation object with test data.

GEMS.test_sim_r — Function

test_sim_r()

Returns a simulation object with test data using the predefined (remote) population models.

GEMS.test_state — Method

test_state(state::String)

Runs a simulation for the German state specified in state. Requires a folder in the project root called "localdata" which contains a subfolder with the state abbreviation and two containing files: people_$state.jld2" and settings_state.jld2.

GEMS.testlogger — Method

testlogger(simulation)

Returns the TestLogger of the simulation.

GEMS.tests — Method

tests(simulation::Simulation)

Calls the dataframe() function on the internal simulation's TestLogger.

GEMS.testtypes — Method

testtypes(simulation)

Returns the test types registered in the simulation.

GEMS.tick — Method

tick(simulation)

Returns current tick of the simulation run.

GEMS.tick_triggers — Method

tick_triggers(simulation)

Returns the list of TickTriggers registered in the simulation.

GEMS.ticks — Function

ticks(logger::InfectionLogger)

Returns a vector of ticks with logging events.

ticks(scheduler)

Returns all the ticks, when individuals are scheduled for vaccination.

GEMS.tickunit — Method

tickunit(simulation)

Returns the unit of the ticks as a char like in date formats, i.e. 'd' means days, 'h' mean hours, etc.

Missing docstring.

Missing docstring for trigger. Check Documenter's build log for details.

GEMS.try_to_infect! — Function

try_to_infect!(infctr::Individual, infctd::Individual, sim::Simulation, pathogen::Pathogen, setting::Setting;
    source_infection_id::Int32 = DEFAULT_INFECTION_ID)

Tries to infect the infctd with the given pathogen transmitted by infctrat time tick(sim) with sim being the simulation. Success depends on whether the agent is alive, not already infected an whether an infection event was sampled using the provided distribution or probability. Returns true if infection was successful.

Parameters

infctr::Individual: Infecting individual

infctd::Individual: Individual to infect

sim::Simulation: Simulation object

pathogen::Pathogen: Pathogen to infect the individual with

setting::Setting: Setting this infection happens in

source_infection_id::Int32 = DEFAULT_INFECTION_ID (optional): Current infection ID of the infecting individual

Returns

Bool: True if infection was successful, false otherwise

Missing docstring.

Missing docstring for update_properties. Check Documenter's build log for details.

GEMS.vaccinate! — Function

vaccinate!(individual::Individual, vaccine::Vaccine, tick::Int16)

Vaccinates the individual with the given vaccine at time tick.

vaccinate!(scheduler, tick)

Vaccinates all individuals that are scheduled to be vaccinated at time tick.

Missing docstring.

Missing docstring for vaccination_schedule(::Simulation). Check Documenter's build log for details.

Missing docstring.

Missing docstring for vaccination_strategy(::ResultData). Check Documenter's build log for details.

Missing docstring.

Missing docstring for vaccine!(::Simulation, ::Vaccine). Check Documenter's build log for details.

Missing docstring.

Missing docstring for vaccine(::Simulation). Check Documenter's build log for details.

Initialization

Structs

GEMS.InfectedFraction — Type

InfectedFraction <: StartCondition

A StartCondition that specifies a fraction of infected individuals (drawn at random).

Fields

fraction::Float64: A fraction of the whole population that has ot be infected

pathogen::Pathogen: The pathogen with which the fraction has to be infected

Functions

Missing docstring.

Missing docstring for ags(PatientZeros). Check Documenter's build log for details.

GEMS.create_contact_sampling_method — Method

create_contact_sampling_method(config::Dict)

Creates a ContactSamplingMethod (CSM) struct using the details specified in the provided dictionary. The dictionary must contain the keys "type" where type corresponds to the name of the ContactSamplingMethod struct to be used. Optionaly the Dict can have the key "parameters". These will be used, to construct the CSM defined by "type". When "type" doesn't have attributes, "parameters" can be ommited.

Missing docstring.

Missing docstring for create_distribution(::Real, ::String). Check Documenter's build log for details.

Missing docstring.

Missing docstring for create_parameter_list. Check Documenter's build log for details.

Missing docstring.

Missing docstring for create_pathogens(::Dict). Check Documenter's build log for details.

GEMS.create_transmission_function — Method

create_transmission_function(config::Dict)

Creates a transmission function struct using the details specified in the provided dictionary. The dictionary must contain the keys type and parameters where type corresponds to the name of the TransmissionFunction struct to be used and parameters holds the keyword arguments for the constructer of this TransmissionFunction. If the provided type does not correspond to the name of a TransmissionFunction an error is thrown.

Returns

<:TransmissionFunction: New instance of a TransmissionFunction struct.

GEMS.fraction — Method

fraction(infectedFraction)

Returns fraction of individuals that shall be infected at the beginning in this start condition.

GEMS.infections — Method

infections(simulation::Simulation)

Calls the dataframe() function on the internal simulation's InfectionLogger.

GEMS.initial_infections — Method

initial_infections(rd::ResultData)

Returns the number of individuals who are marked as infected during initialization. This happens before the actual simulation run. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.initialize! — Function

initialize!(simulation, condition)

Initializes the simulation model according to a provided start condition. This is an 'abstract' function that must be implemented for concrete start condition types.

initialize!(simulation)

Initializes the simulation model with a provided start condition.

initialize!(simulation, infectedFraction)

Initializes the simulation model with a fraction of infected individuals, provided by the start condition.

GEMS.limit — Function

limit(timesUp)

Returns time limit of a timesUp stop criterion.

Missing docstring.

Missing docstring for load_start_condition(::Dict, ::Pathogen). Check Documenter's build log for details.

Missing docstring.

Missing docstring for load_stop_criterion(::Dict). Check Documenter's build log for details.