Batch Runs

Overview Structs

GEMS.Batch

GEMS.BatchData

GEMS.BatchProcessor

GEMS.BatchProcessor

GEMS.DefaultBatchData

Overview Functions

GEMS.add!

GEMS.allocations

GEMS.attack_rate

GEMS.attack_rate

GEMS.config_files

GEMS.cpu_data

GEMS.create_batch_configs

GEMS.cumulative_disease_progressions

GEMS.cumulative_disease_progressions

GEMS.cumulative_quarantines

GEMS.cumulative_quarantines

GEMS.cumulative_quarantines

GEMS.customlogger!

GEMS.dataframes

GEMS.effectiveR

GEMS.effectiveR

GEMS.execution_date

GEMS.exportJLD

GEMS.exportJSON

GEMS.generate_batch_info

GEMS.id

GEMS.import_batchdata

GEMS.julia_version

GEMS.meta_data

GEMS.number_of_individuals

GEMS.number_of_runs

GEMS.parameterset

GEMS.pathogens

GEMS.pathogens_by_name

GEMS.population_files

GEMS.population_pyramid

GEMS.remove!

GEMS.run!

GEMS.run_ids

GEMS.rundata

GEMS.runs

GEMS.runtime

GEMS.settingdata

GEMS.sim_data

GEMS.simulations

GEMS.start_conditions

GEMS.stop_criteria

GEMS.symptom_triggers

GEMS.system_data

GEMS.tests

GEMS.tests

GEMS.testtypes

GEMS.threads

GEMS.tick_cases

GEMS.tick_cases

GEMS.tick_unit

GEMS.total_infections

GEMS.total_infections

GEMS.total_mem_size

GEMS.total_quarantines

GEMS.total_quarantines

GEMS.total_tests

GEMS.total_tests

GEMS.totalonly

GEMS.totalonly!

GEMS.word_size

Structs

GEMS.Batch — Type

Batch

A batch is a container to run and analyze multiple simulations at once. It allows running a single simulation multiple times or combining multiple simulations and scenarios in one analysis.

GEMS.BatchData — Type

BatchData <: AbstractResultData

A struct that stores all processed data of a batch of simulation runs. It holds four internal dictionaries with meta-, simulation-, system- and data as well as multiple DataFrames that are the BatchProcessor outcomes. Note that some information (i.e., execution date or GEMS version) is only read out upon generation of this BatchData object. Thus, there can be inconsistencies if the BatchData object is not generated right after simulation execution.

GEMS.BatchProcessor — Type

BatchProcessor

A type to provide data processing features supplying reports, plots, or other data analyses.

GEMS.DefaultBatchData — Type

DefaultBatchData <: BatchDataStyle

The default style for BatchData objects. It contains all that can currently be calculated in the BatchProcessor and all ResultData objects of the individual runs.

Fields

data::Dict{String, Any}

meta_data::Dict{String, Any}

execution_date::String: Time this BatchData object was generated

GEMS_version::VersionNumber: GEMS version this BatchData object was generated with

system_data::Dict{String, Any}

kernel::String: System kernel

julia_version::String: Julia version that was used to generate this data object

word_size::Int64: System word size

threads::Int64: Number of threads this Julia instance was started with

cpu_data::Markdown.MD: Information on the processor (not available for ARM Macs)

total_mem_size::Float64: Total system memory

free_mem_size::Float64: Available system memory

git_repo::SubString{String}: Current Git repository

git_branch::SubString{String}: Current Git branch

git_commit::SubString{String}: Current Git commit ID

sim_data::Dict{String, Any}

runs::Vector{ResultData}: ResultData objects of individual simulation runs runs

number_of_runs::Int64: Number of simulation runs

total_infections::Dict{String, Real}: Summary statistics on total infections across simulation runs

attack_rate::Dict{String, Real}: Summary statistics on attack rates across simulation runs

total_quarantines::Dict{String, Real}: Summary statistics on total quarantines across simulation runs

total_tests::Dict{String, Real}: Summary statistics on total tests across simulation runs

dataframes::Dict{String, Any}

tick_cases::Dataframe: Aggregated data on infections per tick across simulation runs

effectiveR::Dataframe: Aggregated data on the effective reproduction number per tick across simulation runs

tests::Dataframe: Aggregated data on tests per tick across simulation runs

cumulative_quarantines::Dataframe: Aggregated data on cumulative quarantines per tick across simulation runs

cumulative_disease_progressions::Dataframe: Aggregated data on cumulative disease progressions per tick across simulation runs

Missing docstring.

Missing docstring for DefaultBatchDataStyle. Check Documenter's build log for details.

Constructors

GEMS.BatchProcessor — Method

BatchProcessor(batch::Batch; stysle::String = "LightRD", print_infos::Bool = false)

Creates a BatchProcessor object from a Batch object. This constructor generates the ResultData object for each of the Simulations contained in the Batch. It supresses the usual info outputs that are being made during the ResultData generation. If you want to enable them, pass print_infos = true.

Functions

GEMS.add! — Method

add!(sim::Simulation, batch::Batch)

Adds a Simulation to a Batch.

GEMS.customlogger! — Method

customlogger!(batch::Batch, cl::CustomLogger)

Adds a CustomLogger to each of the Simulation objects that are contained in this batch. Note: This function will generate a duplicate of the passed logger for each of the simulation objects as otherwise, als data from all simulations would we written to the same logger.

GEMS.exportJLD — Method

exportJLD(batchData, directory)

Exports the BatchData object as a JLD2 file, storing it in the specified directory.

GEMS.exportJSON — Method

exportJSON(batchData, directory)

Exports the BatchData object as a JSON file, storing it in the specified directory. All DataFrames are excluded from the JSON export. This feature is only available in the JLD2-export. In that case, please use exportJLD(resultData, directory)

GEMS.import_batchdata — Method

import_batchdata(filepath::AbstractString)

Import the BatchData object from a jld2 file. Returns the BatchData object.

GEMS.remove! — Method

remove!(sim::Simulation, batch::Batch)

Removes a Simulation from a Batch.

GEMS.run! — Method

run!(batch::Batch; stepmod!::Function = x -> x)

Run all simulations within a Batch. You can pass a stepmod! function that is being executed for each simulation in each step. But be aware that it will be the same instance of the stepmod! function. So if you have any internal data collection, this might cause inconsistencies.

GEMS.runs — Method

runs(bd::BatchData)

Returns the ResultData objects of each of the runs in the the batch data object.

GEMS.sim_data — Method

sim_data(bd::BatchData)

Returns the simdata dict of the batch data object.

GEMS.simulations — Method

simulations(batch::Batch)

Returns the list of Simulations associated with a Batch.

GEMS.system_data — Method

system_data(bd::BatchData)

Returns the systemdata dict of the batch data object.

Deprecated Batches

GEMS.allocations — Function

allocations(batchProcessor::BatchProcessor)

Returns a dataframe for the memory allocations of the associated simulation runs in this batch. The columns correspond to the names of the inner timers while each row corresponds to one run. (Note: This data is only available if the simulation runs were done via the main() function)

allocations(bd::BatchData)

Returns allocation data of the simulation runs in this batch. (Note: This data is only available if the simulation runs were done via the main() function)

GEMS.attack_rate — Method

attack_rate(bd::BatchData)

Returns aggregated values for the attack_rate accross the simulation runs in this batch. It returns mean, standard deviation, range, and confidence intervals.

GEMS.attack_rate — Method

attack_rate(batchProcessor::BatchProcessor)

Returns a vector of the attack_rate accross the simulation runs in this batch.

Missing docstring.

Missing docstring for batch. Check Documenter's build log for details.

GEMS.config_files — Function

config_files(batchProcessor::BatchProcessor)

Returns the associated config file.

GEMS.cpu_data — Method

cpu_data(bd::BatchData)

Returns the processor information (not available for ARM Macs)

GEMS.create_batch_configs — Function

create_batch_configs(config_dict, base_config, directory)

Generates batch configurations. A folder config will be created in the directory provided as an argument. In this folder the base_config file and the batch config will both be saved as TOML files. The parametervariation will be saved as a parameter list in the batch config. In the config folder a subfolder files will be created where for each parameterset a config file will be saved.

Arguments:

batch_config::Dict: A dictionary containing information about the batch parameters.

directory::String: The directory where the configuration files will be saved.

Returns:

Array: The file paths of the created config files.

GEMS.cumulative_disease_progressions — Method

cumulative_disease_progressions(batchProcessor::BatchProcessor)

Returns cumulative quarantines per tick accross the simulation runs in this batch.

GEMS.cumulative_disease_progressions — Method

cumulative_disease_progressions(bd::BatchData)

Returns aggregated values for cumulativediseaseprogressions per tick accross the simulation runs in this batch. It returns mean, standard deviation, range, and confidence intervals.

GEMS.cumulative_quarantines — Function

cumulative_quarantines(postProcessor::PostProcessor)

Returns a DataFrame containing cumulative information about days spent in isolation.

Columns

Name	Type	Description

tick	Int16	Simulation tick (time)

quarantined	Int64	Total number of individuals in isolation during that tick

students	Int64	Total number of students in isolation during that tick

workers	Int64	Total number of workers in isolation during that tick

other	Int64	Total number of non-students and -workers in isolation during that tick

cumulative_quarantines(rd::ResultData)

Returns the DataFrame with number of isolated individuals per tick Look up the PostProcessor docs to find the column definitions. Returns an empty dictionary if the data is not available in the input ResultData object.

cumulative_quarantines(batchProcessor::BatchProcessor)

Returns cumulative quarantines per tick accross the simulation runs in this batch.

cumulative_quarantines(bd::BatchData)

Returns aggregated values cumulative_quarantines per tick accross the simulation runs in this batch. It returns mean, standard deviation, range, and confidence intervals.

GEMS.cumulative_quarantines — Method

cumulative_quarantines(bd::BatchData)

Returns aggregated values cumulative_quarantines per tick accross the simulation runs in this batch. It returns mean, standard deviation, range, and confidence intervals.

GEMS.cumulative_quarantines — Method

cumulative_quarantines(batchProcessor::BatchProcessor)

Returns cumulative quarantines per tick accross the simulation runs in this batch.

GEMS.dataframes — Method

dataframes(batchData)

Returns the dataframes dict of the batch data object.

GEMS.effectiveR — Method

effectiveR(bd::BatchData)

Returns aggregated values for the effective R value for each tick accross the simulation runs in this batch. It returns mean, standard deviation, range, and confidence intervals.

GEMS.effectiveR — Method

effectiveR(batchProcessor::BatchProcessor)

Returns the effective R value for each tick accross the simulation runs in this batch. It returns mean, standard deviation, range, and confidence intervals.

GEMS.execution_date — Method

execution_date(bd::BatchData)

Returns the timestamp of batch data generation.

GEMS.generate_batch_info — Function

generate_batch_info(date, seeds, configfile, populationfile, directory)

Exports the batchinfo including the date, used seeds, configfile and population file. It is being saved in the provided directory as a json.

GEMS.id — Method

id(bd::BatchData)

Returns the stringified SHA1 hash that serves as a unique identifier.

GEMS.julia_version — Method

julia_version(bd::BatchData)

Returns the Julia version that was used to generate this result data object.

GEMS.meta_data — Method

meta_data(bd::BatchData)

Returns the metadata dict of the batch data object.

GEMS.number_of_individuals — Method

number_of_individuals(batchProcessor::BatchProcessor)

Returns a vector of number_of_individuals that were used in the associated simulations. Values are unique and can originate from multiple simulation runs.

GEMS.number_of_runs — Method

number_of_runs(bd::BatchData)

Returns the number of runs in this batch.

GEMS.parameterset — Method

parameterset(batchProcessor::BatchProcessor, parameters::Vector{String})

Takes the batchProcessor and a vector of strings corresponding to the keys of parameter values in the config file(in the format key.key.key...). Returns a DataFrame containing the columns id with the individual simulation ids and parameter_set with the parameter values corresponding to the provided keys. The parameter_set dictionaries have the same keys as the provided strings.

Dataframe Columns

Name	Type	Description

run_id	String	id of the simulation run

parameter_set	Dict	Parameter set of the provided parameter keys.

GEMS.pathogens — Method

pathogens(batchProcessor::BatchProcessor)

Returns a vector of pathogens that were used in the associated simulations. Values are unique and can originate from multiple simulation runs.

GEMS.pathogens_by_name — Function

pathogens_by_name(batchProcessor::BatchProcessor)

Returns a {name, pathogen[]} dictionary of pathogens that were used in the associated simulations. This is done to keep multiple pathogen configurations of the same name with varying parameters. Names are unique and can originate from multiple simulation runs.

GEMS.population_files — Function

population_files(batchProcessor::BatchProcessor)

Returns the associated population file.

GEMS.population_pyramid — Method

population_pyramid(batchProcessor::BatchProcessor)

Returns a {String, DataFrame} dictionary containing data to generate a population pyramid for each population files of this batch.

Dataframe Columns

Name	Type	Description

age	Int8	1-year age classes

sex	Int8	Sex according to population DataFame (0 = female, 1 = male)

gender	String	String variant of Sex [Female, Male]

sum	Int64	Total of all genders in all ages (females multiplied by -1)

GEMS.run_ids — Function

run_ids(batchProcessor::BatchProcessor)

Returns the vector of simulation run IDs that are associated with this batch.

GEMS.rundata — Function

rundata(batchProcessor::BatchProcessor)

Returns the vector of ResultData objects of simulation runs that are associated with this batch.

GEMS.runtime — Function

runtime(batchProcessor::BatchProcessor)

Returns dataframe of the runtimes of the associated simulation runs in this batch. The columns correspond to the names of the inner timers while each row corresponds to one run. (Note: This data is only available if the simulation runs were done via the main() function)

runtime(bd::BatchData)

Returns runtime data of the simulation runs in this batch. (Note: This data is only available if the simulation runs were done via the main() function)

GEMS.settingdata — Method

settingdata(batchProcessor::BatchProcessor)

Returns a {String, DataFrame} dictionary containing information about setting types in the population files in the simulation runs of this batch. Populations are distinguished by their population file name stored in the key of the result dictionary.

Dataframe Columns

Name	Type	Description

setting_type	String	Setting type identifier (name)

number_of_settings	Int64	Overall number of settings of that type

min_individuals	Float64	Lowest number of individuals assigned to a setting of this type

max_individuals	Float64	Highest number of individuals assigned to a setting of this type

avg_individuals	Float64	Average number of individuals assigned to a setting of this type

GEMS.start_conditions — Function

start_conditions(batchProcessor::BatchProcessor)

Returns a vector of start_conditions that were used in the associated simulations. Values are unique and can originate from multiple simulation runs.

GEMS.stop_criteria — Function

stop_criteria(batchProcessor::BatchProcessor)

Returns a vector of stop_criteria that were used in the associated simulations. Values are unique and can originate from multiple simulation runs.

GEMS.symptom_triggers — Method

symptom_triggers(batchProcessor::BatchProcessor)

Returns a vector of the symptom_triggers accross the simulation runs in this batch.

Missing docstring.

Missing docstring for systemdata(::BatchData). Check Documenter's build log for details.

GEMS.tests — Method

tests(bd::BatchData)

Returns aggregated values for tests per tick accross the simulation runs in this batch. It returns mean, standard deviation, range, and confidence intervals.

GEMS.tests — Method

tests(batchProcessor::BatchProcessor)

Returns newly exposed inviduals per tick accross the simulation runs in this batch. It returns mean, standard deviation, range, and confidence intervals.

GEMS.testtypes — Method

testtypes(batchProcessor::BatchProcessor)

Returns a vector of the testtypes accross the simulation runs in this batch.

GEMS.threads — Method

threads(bd::BatchData)

Returns the number of threads this Julia instance was started with

GEMS.tick_cases — Method

tick_cases(bd::BatchData)

Returns aggregated values for newly exposed inviduals per tick accross the simulation runs in this batch. It returns mean, standard deviation, range, and confidence intervals.

GEMS.tick_cases — Method

tick_cases(batchProcessor::BatchProcessor)

Returns newly exposed inviduals per tick accross the simulation runs in this batch. It returns mean, standard deviation, range, and confidence intervals.

GEMS.tick_unit — Method

tick_unit(batchProcessor::BatchProcessor)

Returns a vector of tick_units that were used in the associated simulations. Values are unique and can originate from multiple simulation runs.

GEMS.total_infections — Method

total_infections(bd::BatchData)

Returns aggregated values for total_infections accross the simulation runs in this batch. It returns mean, standard deviation, range, and confidence intervals.

GEMS.total_infections — Method

total_infections(batchProcessor::BatchProcessor)

Returns a vector of the total_infections accross the simulation runs in this batch.

GEMS.total_mem_size — Method

total_mem_size(bd::BatchData)

Returns the total system memory

GEMS.total_quarantines — Method

total_quarantines(batchData)

Returns aggregated values for total_quarantines accross the simulation runs in this batch. It returns mean, standard deviation, range, and confidence intervals.

GEMS.total_quarantines — Method

total_quarantines(batchProcessor::BatchProcessor)

Returns a vector of the total_quarantines accross the simulation runs in this batch.

GEMS.total_tests — Method

total_tests(bd::BatchData)

Returns aggregated values for total_tests accross the simulation runs in this batch. It returns mean, standard deviation, range, and confidence intervals.

GEMS.total_tests — Method

total_tests(batchProcessor::BatchProcessor)

Returns a Dict of vectors of the total_tests per TestType accross the simulation runs in this batch.

GEMS.totalonly! — Method

totalonly!(batchTickTests, totalonly)

Setter for the totalonly field of the BatchTickTests plot specifying whether only the total number of tests should be plotted or all data series (including positive, negative tests, etc...)

GEMS.totalonly — Method

totalonly(batchTickTests)

Getter for the totalonly field of BatchTickTests plot structs.

GEMS.word_size — Method

word_size(bd::BatchData)

Returns the system word size