Individuals

hier eine subsection: wie fragt man den Status der agents ab (hier alle funktionen, wie man states queried)

Overview Structs

GEMS.Agent

GEMS.Individual

GEMS.Individual

GEMS.Individual

Overview Functions

GEMS.age

GEMS.class_id

GEMS.comorbidities

GEMS.education

GEMS.end_quarantine!

GEMS.has_municipality

GEMS.home_quarantine!

GEMS.household_id

GEMS.id

GEMS.infection_id

GEMS.infectiousness

GEMS.infectiousness!

GEMS.is_student

GEMS.is_working

GEMS.isquarantined

GEMS.isvaccinated

GEMS.last_reported_at

GEMS.last_reported_at!

GEMS.last_test

GEMS.last_test!

GEMS.last_test_result

GEMS.last_test_result!

GEMS.mandate_compliance

GEMS.mandate_compliance!

GEMS.municipality

GEMS.municipality_id

GEMS.num_of_infected

GEMS.number_of_infections

GEMS.number_of_vaccinations

GEMS.occupation

GEMS.office

GEMS.office_id

GEMS.pathogen_id

GEMS.quarantine_release_tick

GEMS.quarantine_release_tick!

GEMS.quarantine_status

GEMS.quarantine_tick

GEMS.quarantine_tick!

GEMS.reset!

GEMS.setting_id

GEMS.setting_id!

GEMS.settings

GEMS.sex

GEMS.social_factor

GEMS.social_factor!

GEMS.vaccination_tick

GEMS.vaccine_id

Structs

GEMS.Agent — Type

Supertype for simulation agents

GEMS.Individual — Type

Individual <: Agent

A type to represent individuals, that act as agents inside the simulation.

Fields

General

id::Int32: Unique identifier of the individual

sex::Int8: Sex (Female (1), Male(2), Diverse (3))

age::Int8: Age

education::Int8: Education class (i.e. highest degree)

occupation::Int16: Occupation class (i.e. manual labour, office job, etc...)

Behaviour

social_factor::Float32: Parameter for the risk-willingness. Can be anywhere between -1 and 1 with neutral state is 0.

mandate_compliance::Float32. Paremeter which influences the probability of complying to mandates. Can be anywhere between -1 and 1 with neutral state is 0.

Health Status

comorbidities::Vector{Bool}: Indicating prevalence of certain health conditions. True, if the individual is preconditioned with the comorbidity associated to the array index.

dead::Bool: Flag indicating individual's decease

infected::Bool: Flag indicating individual's infection status

Associated Settings

household::Int32: Reference to household id

office::Int32: Reference to office id

schoolclass::Int32: Reference to schoolclass id

municipality::Int32: Reference to municipality id

Pathogen

pathogen_id::Int8: pathogen identifier

infection_id::Int32: Current infection id

infectiousness::Int8: an individuals infectiousness (1-127), i.e. for superspreaders

number_of_infections::Int8: infection count

Natural Disease History

exposure::Int16: Tick of most recent exposure

infectiousness_onset::Int16: Tick of most recent change into "infectious" state

symptom_onset::Int16: Tick of the onset of symptoms

severeness_onset::Int16: Tick of onset of severe symptoms

hospital_admission::Int16: Tick of hospitalization

icu_admission::Int16: Tick of ICU (Intensive Care Unit) admission

icu_discharge::Int16: Tick of ICU (Intensive Care Unit) discharge

ventilation_admission::Int16: Tick of ventilation admission

ventilation_discharge::Int16: Tick of ventilation discharge

hospital_discharge::Int16: Tick of hospital discharge

recovery::Int16: Tick of most recent recovery

death::Int16: Tick of Death

Testing

last_test::Int16: Tick of last test for pathogen

last_test_result::Bool: Flag for positivity of last test

last_reported_at::Int16: Tick at which this individual was last reported

Vaccination

vaccine_id::Int8: Vaccine identifier

number_of_vaccinations::Int8: Individual's vaccination counter

vaccination_tick::Int16: Tick of most recent vaccination

Interventions

quarantine_status::Int8: Status to indicate quarantine (none, household_quarantined, hospitalized, etc...)

quarantine_tick::Int16: Start tick of quarantine

quarantine_release_tick::Int16: End tick of quarantine

Constructors

GEMS.Individual — Method

Individual(properties::Dict)

Create an individual with the provided properties. Properties must have at least keys id, sex, age.

GEMS.Individual — Method

Individual(properties::DataFrameRow)

Create an individual with the provided properties. Properties must have at least keys id, sex, age.

Functions

GEMS.age — Function

age(individual::Individual)

Return an individual's age.

GEMS.class_id — Function

class_id(individual::Individual)

Returns an individual's associated class's ID.

GEMS.comorbidities — Function

comorbidities(individual::Individual)

Returns an individual's comorbidities.

GEMS.education — Method

education(individual::Individual)

Return an individual's education class

GEMS.end_quarantine! — Method

end_quarantine!(individual::Individual)

Ends an individuals quarantine.

GEMS.has_municipality — Method

has_municipality(individual::Individual)

Returns true if individual is assigned to an instance of type Municipality.

GEMS.home_quarantine! — Method

home_quarantine!(individual::Individual)

Quarantines an individual in their household.

GEMS.household_id — Method

household_id(individual::Individual)

Returns an individual's associated household's ID.

GEMS.id — Method

id(individual::Individual)

Return the unique identifier of the individual.

GEMS.infection_id — Method

infection_id(individual::Individual)

Returns an individual's infection_id (currently infected).

GEMS.infectiousness! — Method

infectiousness!(individual::Individual, infectiousness)

Assigns a specified infectiousness (0-127) to an individual.

GEMS.infectiousness — Method

infectiousness(individual::Individual)

Returns an individual's infectiousness (currently infected).

GEMS.is_student — Method

is_student(individual::Individual)

Returns true if individual is assigned to an instance of type SchoolClass.

GEMS.is_working — Method

is_working(individual::Individual)

Returns true if individual is assigned to an instance of type Office.

GEMS.isquarantined — Method

isquarantined(individual::Individual)

Returns wether the individual is in quarantine or not.

GEMS.isvaccinated — Method

isvaccinated(individual::Individual)

Returns wether the individual is vaccinated.

GEMS.last_reported_at! — Method

last_reported_at!(individual::Individual, report_tick::Int16)

Sets last tick this individual was last reported.

GEMS.last_reported_at — Method

last_reported_at(individual::Individual)

Returns the last tick this individual was a reported case.

GEMS.last_test! — Method

last_test!(individual::Individual, tick::Int16)

Sets last test date (tick).

GEMS.last_test — Method

last_test(individual::Individual)

Returns last test date (tick).

GEMS.last_test_result! — Method

last_test_result!(individual::Individual, test_result::Bool)

Sets last test result.

GEMS.last_test_result — Method

last_test_result(individual::Individual)

Returns whether last test was positive. Defaults to false.

GEMS.mandate_compliance! — Method

mandate_compliance!(individual::Individual, val::Float32)

Overwrites the individual's mandate_compliance attribute.

GEMS.mandate_compliance — Method

mandate_compliance(individual::Individual)

Return an individual's mandate_compliance value.

GEMS.municipality — Method

municipality(i::Individual, sim::Simulation)::Municipality

Returns the Municipality instance referenced in an individual.

GEMS.municipality_id — Method

municipality_id(individual::Individual)

Returns an individual's associated municipalities ID.

GEMS.num_of_infected — Method

num_of_infected(individuals::Vector{Individual})

Takes a vector of individuals and returns the number of infected individuals.

GEMS.number_of_infections — Method

number_of_infections(individual::Individual)

Returns an individual's number of infections (currently infected).

GEMS.number_of_vaccinations — Method

number_of_vaccinations(individual::Individual)

Returns the number of vaccinations.

GEMS.occupation — Method

occupation(individual::Individual)

Returns an individual's occupation class.

GEMS.office — Method

office(i::Individual, sim::Simulation)::Office

Returns the Office instance referenced in an individual.

GEMS.office_id — Method

office_id(individual::Individual)

Returns an individual's associated office's ID.

GEMS.pathogen_id — Method

pathogen_id(individual::Individual)

Returns an individual's pathogen (currently infected).

GEMS.quarantine_release_tick! — Method

quarantine_release_tick!(individual::Individual, tick::Int16)

Sets an individual's quarantine release tick.

GEMS.quarantine_release_tick — Method

quarantine_release_tick(individual::Individual)

Returns an individual's quarantine release tick.

GEMS.quarantine_status — Method

quarantine_status(individual::Individual)

Returns an individuals quarantine status.

GEMS.quarantine_tick! — Method

quarantine_tick!(individual::Individual, tick::Int16)

Sets an individual's quarantine tick.

GEMS.quarantine_tick — Method

quarantine_tick(individual::Individual)

Returns an individual's quarantine tick.

GEMS.reset! — Method

reset!(individual::Individual)

Resets all non-static values like the disease progression timing. The individual will get back into a state where it was never infected, vaccinated, tested, etc.

GEMS.setting_id! — Method

setting_id!(individual::Individual, type::DataType, id::Int32)

Changes the assigned setting id of the individual for the given type of setting to id.

GEMS.setting_id — Method

setting_id(individual::Individual, type::DataType)

Returns the id of the setting of type associated with the individual. If the settingtype is unknown or the agent isn't part of a setting of that type, -1 will be returned.

GEMS.settings — Method

settings(individual::Individual, sim::Simulation)

Return a dictionary with the assigned setting types as keys and the assigned IDs as values.

GEMS.sex — Method

sex(individual::Individual)

Return an individual's sex.

GEMS.social_factor! — Method

social_factor!(individual::Individual, val::Float32)

Overwrites the individual's social_factor attribute.

GEMS.social_factor — Method

social_factor(individual::Individual)

Returns an individual's social_factor value.

GEMS.vaccination_tick — Method

vaccination_tick(individual::Individual)

Returns the time of last vaccination.

GEMS.vaccine_id — Method

vaccine_id(individual::Individual)

Returns the id of the vaccine the individual is vaccinated with.