Individuals

hier eine subsection: wie fragt man den Status der agents ab (hier alle funktionen, wie man states queried)

Overview Structs

GEMS.Agent

GEMS.Individual

GEMS.Individual

GEMS.Individual

Overview Functions

GEMS.age

GEMS.class_id

GEMS.comorbidities

GEMS.dead!

GEMS.death

GEMS.death!

GEMS.detected!

GEMS.education

GEMS.end_quarantine!

GEMS.exposure

GEMS.exposure!

GEMS.has_municipality

GEMS.home_quarantine!

GEMS.hospital_admission

GEMS.hospital_admission!

GEMS.hospital_discharge

GEMS.hospital_discharge!

GEMS.hospitalized!

GEMS.household_id

GEMS.icu_admission

GEMS.icu_admission!

GEMS.icu_discharge

GEMS.icu_discharge!

GEMS.id

GEMS.infected!

GEMS.infection_id

GEMS.infectious!

GEMS.infectiousness

GEMS.infectiousness!

GEMS.infectiousness_onset

GEMS.infectiousness_onset!

GEMS.is_dead

GEMS.is_dead

GEMS.is_detected

GEMS.is_detected

GEMS.is_exposed

GEMS.is_exposed

GEMS.is_hospitalized

GEMS.is_hospitalized

GEMS.is_icu

GEMS.is_icu

GEMS.is_infected

GEMS.is_infected

GEMS.is_infectious

GEMS.is_infectious

GEMS.is_mild

GEMS.is_presymptomatic

GEMS.is_recovered

GEMS.is_severe

GEMS.is_severe

GEMS.is_student

GEMS.is_symptomatic

GEMS.is_symptomatic

GEMS.is_ventilated

GEMS.is_ventilated

GEMS.is_working

GEMS.isvaccinated

GEMS.last_reported_at

GEMS.last_reported_at!

GEMS.last_test

GEMS.last_test!

GEMS.last_test_result

GEMS.last_test_result!

GEMS.mandate_compliance

GEMS.mandate_compliance!

GEMS.municipality

GEMS.municipality_id

GEMS.num_of_infected

GEMS.number_of_infections

GEMS.number_of_vaccinations

GEMS.occupation

GEMS.office

GEMS.office_id

GEMS.pathogen_id

GEMS.progress_disease!

GEMS.quarantine_release_tick

GEMS.quarantine_release_tick!

GEMS.quarantine_status

GEMS.quarantine_tick

GEMS.quarantine_tick!

GEMS.recovery

GEMS.recovery!

GEMS.reset!

GEMS.setting_id

GEMS.setting_id!

GEMS.settings

GEMS.severe!

GEMS.severeness_offset

GEMS.severeness_offset!

GEMS.severeness_onset

GEMS.severeness_onset!

GEMS.sex

GEMS.social_factor

GEMS.social_factor!

GEMS.symptom_onset

GEMS.symptom_onset!

GEMS.symptomatic!

GEMS.vaccination_tick

GEMS.vaccine_id

GEMS.ventilated!

GEMS.ventilation_admission

GEMS.ventilation_admission!

GEMS.ventilation_discharge

GEMS.ventilation_discharge!

Structs

GEMS.Agent — Type

Supertype for simulation agents

GEMS.Individual — Type

Individual <: Agent

A type to represent individuals, that act as agents inside the simulation.

Fields

General

id::Int32: Unique identifier of the individual

sex::Int8: Sex (Female (1), Male(2), Diverse (3))

age::Int8: Age

education::Int8: Education class (i.e. highest degree)

occupation::Int16: Occupation class (i.e. manual labour, office job, etc...)

Behaviour

social_factor::Float32: Parameter for the risk-willingness. Can be anywhere between -1 and 1 with neutral state is 0.

mandate_compliance::Float32. Paremeter which influences the probability of complying to mandates. Can be anywhere between -1 and 1 with neutral state is 0.

Health Status

comorbidities::Vector{Bool}: Indicating prevalence of certain health conditions. True, if the individual is preconditioned with the comorbidity associated to the array index.

dead::Bool: Flag indicating individual's decease

infected::Bool: Flag indicating individual's infection status

Associated Settings

household::Int32: Reference to household id

office::Int32: Reference to office id

schoolclass::Int32: Reference to schoolclass id

municipality::Int32: Reference to municipality id

Pathogen

pathogen_id::Int8: pathogen identifier

infection_id::Int32: Current infection id

infectiousness::Int8: an individuals infectiousness (1-127), i.e. for superspreaders

number_of_infections::Int8: infection count

Natural Disease History

exposure::Int16: Tick of most recent exposure

infectiousness_onset::Int16: Tick of most recent change into "infectious" state

symptom_onset::Int16: Tick of the onset of symptoms

severeness_onset::Int16: Tick of onset of severe symptoms

hospital_admission::Int16: Tick of hospitalization

icu_admission::Int16: Tick of ICU (Intensive Care Unit) admission

icu_discharge::Int16: Tick of ICU (Intensive Care Unit) discharge

ventilation_admission::Int16: Tick of ventilation admission

ventilation_discharge::Int16: Tick of ventilation discharge

hospital_discharge::Int16: Tick of hospital discharge

recovery::Int16: Tick of most recent recovery

death::Int16: Tick of Death

Testing

last_test::Int16: Tick of last test for pathogen

last_test_result::Bool: Flag for positivity of last test

last_reported_at::Int16: Tick at which this individual was last reported

Vaccination

vaccine_id::Int8: Vaccine identifier

number_of_vaccinations::Int8: Individual's vaccination counter

vaccination_tick::Int16: Tick of most recent vaccination

Interventions

quarantine_status::Int8: Status to indicate quarantine (none, household_quarantined, hospitalized, etc...)

quarantine_tick::Int16: Start tick of quarantine

quarantine_release_tick::Int16: End tick of quarantine

Constructors

GEMS.Individual — Method

Individual(properties::Dict)

Create an individual with the provided properties. Properties must have at least keys id, sex, age.

GEMS.Individual — Method

Individual(properties::DataFrameRow)

Create an individual with the provided properties. Properties must have at least keys id, sex, age.

Functions

GEMS.age — Function

age(individual::Individual)

Return an individual's age.

GEMS.class_id — Function

class_id(individual::Individual)

Returns an individual's associated class's ID.

GEMS.comorbidities — Function

comorbidities(individual::Individual)

Returns an individual's comorbidities.

GEMS.dead! — Method

dead!(individual::Individual, dead::Bool)

Set the dead flag of the individual.

GEMS.death — Method

death(individual::Individual)

Returns an individual's last death tick. Return -1 if never died.

GEMS.death! — Method

death!(individual::Individual, tick::Int16)

Sets an individual's death tick.

GEMS.detected! — Method

detected!(individual::Individual, detected::Bool)

Sets the detected flag of the individual.

GEMS.education — Method

education(individual::Individual)

Return an individual's education class

GEMS.end_quarantine! — Method

end_quarantine!(individual::Individual)

Ends an individuals quarantine.

GEMS.exposure — Method

exposure(individual::Individual)

Returns an individual's last exposure tick. Returun -1 if never exposed.

GEMS.exposure! — Method

exposure!(individual::Individual, tick::Int16)

Sets an individual's exposure tick.

GEMS.has_municipality — Method

has_municipality(individual::Individual)

Returns true if individual is assigned to an instance of type Municipality.

GEMS.home_quarantine! — Method

home_quarantine!(individual::Individual)

Quarantines an individual in their household.

GEMS.hospital_admission — Method

hospital_admission(individual::Individual)

Returns an individual's last hospital admission tick. Return -1 if never hospitalized.

GEMS.hospital_admission! — Method

hospital_admission!(individual::Individual, tick::Int16)

Sets an individual's hospital admission tick.

GEMS.hospital_discharge — Method

hospital_discharge(individual::Individual)

Returns an individual's last hospital discharge tick. Return -1 if never discharged from hospital.

GEMS.hospital_discharge! — Method

hospital_discharge!(individual::Individual, tick::Int16)

Sets an individual's hospital discharge tick.

GEMS.hospitalized! — Method

hospitalized!(individual::Individual, hospitalized::Bool)

Sets the hospitalized flag of the individual.

GEMS.household_id — Method

household_id(individual::Individual)

Returns an individual's associated household's ID.

GEMS.icu_admission — Method

icu_admission(individual::Individual)

Returns an individual's last icu admission tick. Return -1 if never admitted to icu.

GEMS.icu_admission! — Method

icu_admission!(individual::Individual, tick::Int16)

Sets an individual's icu admission tick.

GEMS.icu_discharge — Method

icu_discharge(individual::Individual)

Returns an individual's last icu discharge tick. Return -1 if never discharged from icu.

GEMS.icu_discharge! — Method

icu_discharge!(individual::Individual, tick::Int16)

Sets an individual's icu discharge tick.

GEMS.id — Method

id(individual::Individual)

Return the unique identifier of the individual.

GEMS.infected! — Method

infected!(individual::Individual, infected::Bool)

Sets the infected flag of the individual.

GEMS.infection_id — Method

infection_id(individual::Individual)

Returns an individual's infection_id (currently infected).

GEMS.infectious! — Method

infectious!(individual::Individual, infectious::Bool)

Sets the infectious flag of the individual.

GEMS.infectiousness! — Method

infectiousness!(individual::Individual, infectiousness)

Assigns a specified infectiousness (0-127) to an individual.

GEMS.infectiousness — Method

infectiousness(individual::Individual)

Returns an individual's infectiousness (currently infected).

GEMS.infectiousness_onset — Method

infectiousness_onset(individual::Individual)

Returns an individual's last infectiousness onset tick. Return -1 if never infectious.

GEMS.infectiousness_onset! — Method

infectiousness_onset!(individual::Individual, tick::Int16)

Sets an individual's infectiousness onset tick.

GEMS.is_dead — Method

is_dead(individual::Individual)
isdead(individual::Individual)
dead(individual::Individual)

Returns true if the individual is dead.

GEMS.is_dead — Method

is_dead(individual::Individual, t::Int16)
isdead(individual::Individual, t::Int16)
dead(individual::Individual, t::Int16)

Returns true if the individual is dead at tick t.

GEMS.is_detected — Method

is_detected(individual::Individual)

Returns true if the individual is currently infected and has been detected (i.e. tested positive).

GEMS.is_detected — Method

is_detected(individual::Individual, t::Int16)

Returns true if the individual is currently infected and has been detected (i.e. tested positive) prior to or at tick t.

GEMS.is_exposed — Method

is_exposed(individual::Individual)
isexposed(individual::Individual)
exposed(individual::Individual)

Returns true if the individual is exposed at the current tick. Exposed means infected but not yet infectious.

GEMS.is_exposed — Method

is_exposed(individual::Individual, t::Int16)
isexposed(individual::Individual, t::Int16)
exposed(individual::Individual, t::Int16)

Returns true if the individual is exposed at tick t. Exposed means infected but not yet infectious.

GEMS.is_hospitalized — Method

is_hospitalized(individual::Individual)
ishospitalized(individual::Individual)
hospitalized(individual::Individual)

Returns the hospitalized flag of the individual.

GEMS.is_hospitalized — Method

is_hospitalized(individual::Individual, t::Int16)    
ishospitalized(individual::Individual, t::Int16)
hospitalized(individual::Individual, t::Int16)

Returns true if the individual is hospitalized at tick t.

GEMS.is_icu — Method

is_icu(individual::Individual)
isicu(individual::Individual)
isicu(individual::Individual)

Returns the icu flag of the individual.

GEMS.is_icu — Method

is_icu(individual::Individual, t::Int16)
isicu(individual::Individual, t::Int16)
icu(individual::Individual, t::Int16)

Returns true if the individual is in ICU at tick t.

GEMS.is_infected — Method

is_infected(individual::Individual)
isinfected(individual::Individual)
infected(individual::Individual)

Returns the infected flag of the individual.

GEMS.is_infected — Method

is_infected(individual::Individual, t::Int16)
isinfected(individual::Individual, t::Int16)
infected(individual::Individual, t::Int16)

Returns true if the individual is infected at tick t.

GEMS.is_infectious — Method

is_infectious(individual::Individual)
isinfectious(individual::Individual)
infectious(individual::Individual)

Returns the infectious flag of the individual.

GEMS.is_infectious — Method

is_infectious(individual::Individual, t::Int16)
isinfectious(individual::Individual, t::Int16)
infectious(individual::Individual, t::Int16)

Returns true if the individual is infectious at tick t.

GEMS.is_mild — Method

is_mild(individual::Individual, t::Int16)
ismild(individual::Individual, t::Int16)
mild(individual::Individual, t::Int16)

Returns true if the individual is in a mild state at tick t. Mild means symptomatic but not severe.

GEMS.is_presymptomatic — Method

is_presymptomatic(individual::Individual, t::Int16)
ispresymptomatic(individual::Individual, t::Int16)
presymptomatic(individual::Individual, t::Int16)

Returns true if the individual is presymptomatic at tick t. Presymptomatic means infected, will develop symptoms, but is not yet symptomatic.

GEMS.is_recovered — Method

is_recovered(individual::Individual, t::Int16)
isrecovered(individual::Individual, t::Int16)
recovered(individual::Individual, t::Int16)

Returns true if the individual is recovered at tick t.

GEMS.is_severe — Method

is_severe(individual::Individual)
issevere(individual::Individual)
severe(individual::Individual)

Returns the severe flag of the individual.

GEMS.is_severe — Method

is_severe(individual::Individual, t::Int16)
issevere(individual::Individual, t::Int16)
severe(individual::Individual, t::Int16)

Returns true if the individual is in a severe state at tick t.

GEMS.is_student — Method

is_student(individual::Individual)

Returns true if individual is assigned to an instance of type SchoolClass.

GEMS.is_symptomatic — Method

is_symptomatic(individual::Individual)
issymptomatic(individual::Individual)
symptomatic(individual::Individual)

Returns the symptomatic flag of the individual.

GEMS.is_symptomatic — Method

is_symptomatic(individual::Individual, t::Int16)
issymptomatic(individual::Individual, t::Int16)
symptomatic(individual::Individual, t::Int16)

Returns true if the individual is symptomatic at tick t.

GEMS.is_ventilated — Method

is_ventilated(individual::Individual)
isventilated(individual::Individual)
ventilated(individual::Individual)

Returns the ventilated flag of the individual.

GEMS.is_ventilated — Method

is_ventilated(individual::Individual, t::Int16)
isventilated(individual::Individual, t::Int16)
ventilated(individual::Individual, t::Int16)

Returns true if the individual is ventilated at tick t.

GEMS.is_working — Method

is_working(individual::Individual)

Returns true if individual is assigned to an instance of type Office.

Missing docstring.

Missing docstring for isquarantined(::Individual). Check Documenter's build log for details.

GEMS.isvaccinated — Method

isvaccinated(individual::Individual)

Returns wether the individual is vaccinated.

GEMS.last_reported_at! — Method

last_reported_at!(individual::Individual, report_tick::Int16)

Sets last tick this individual was last reported.

GEMS.last_reported_at — Method

last_reported_at(individual::Individual)

Returns the last tick this individual was a reported case.

GEMS.last_test! — Method

last_test!(individual::Individual, tick::Int16)

Sets last test date (tick).

GEMS.last_test — Method

last_test(individual::Individual)

Returns last test date (tick).

GEMS.last_test_result! — Method

last_test_result!(individual::Individual, test_result::Bool)

Sets last test result.

GEMS.last_test_result — Method

last_test_result(individual::Individual)

Returns whether last test was positive. Defaults to false.

GEMS.mandate_compliance! — Method

mandate_compliance!(individual::Individual, val::Float32)

Overwrites the individual's mandate_compliance attribute.

GEMS.mandate_compliance — Method

mandate_compliance(individual::Individual)

Return an individual's mandate_compliance value.

GEMS.municipality — Method

municipality(i::Individual, sim::Simulation)::Municipality

Returns the Municipality instance referenced in an individual.

GEMS.municipality_id — Method

municipality_id(individual::Individual)

Returns an individual's associated municipalities ID.

GEMS.num_of_infected — Method

num_of_infected(individuals::Vector{Individual})

Takes a vector of individuals and returns the number of infected individuals.

GEMS.number_of_infections — Method

number_of_infections(individual::Individual)

Returns an individual's number of infections (currently infected).

GEMS.number_of_vaccinations — Method

number_of_vaccinations(individual::Individual)

Returns the number of vaccinations.

GEMS.occupation — Method

occupation(individual::Individual)

Returns an individual's occupation class.

GEMS.office — Method

office(i::Individual, sim::Simulation)::Office

Returns the Office instance referenced in an individual.

GEMS.office_id — Method

office_id(individual::Individual)

Returns an individual's associated office's ID.

GEMS.pathogen_id — Method

pathogen_id(individual::Individual)

Returns an individual's pathogen (currently infected).

GEMS.progress_disease! — Method

progress_disease!(individual::Individual, tick::Int16)

Updates the disease progression status flags of the individual at the given tick.

GEMS.quarantine_release_tick! — Method

quarantine_release_tick!(individual::Individual, tick::Int16)

Sets an individual's quarantine release tick.

GEMS.quarantine_release_tick — Method

quarantine_release_tick(individual::Individual)

Returns an individual's quarantine release tick.

GEMS.quarantine_status — Method

quarantine_status(individual::Individual)

Returns an individuals quarantine status.

GEMS.quarantine_tick! — Method

quarantine_tick!(individual::Individual, tick::Int16)

Sets an individual's quarantine tick.

GEMS.quarantine_tick — Method

quarantine_tick(individual::Individual)

Returns an individual's quarantine tick.

GEMS.recovery — Method

recovery(individual::Individual)

Returns an individual's last recovery tick. Return -1 if never recovered.

GEMS.recovery! — Method

recovery!(individual::Individual, tick::Int16)

Sets an individual's recovery tick.

GEMS.reset! — Method

reset!(individual::Individual)

Resets all non-static values like the disease progression timing. The individual will get back into a state where it was never infected, vaccinated, tested, etc.

GEMS.setting_id! — Method

setting_id!(individual::Individual, type::DataType, id::Int32)

Changes the assigned setting id of the individual for the given type of setting to id.

GEMS.setting_id — Method

setting_id(individual::Individual, type::DataType)

Returns the id of the setting of type associated with the individual. If the settingtype is unknown or the agent isn't part of a setting of that type, -1 will be returned.

GEMS.settings — Method

settings(individual::Individual, sim::Simulation)

Return a dictionary with the assigned setting types as keys and the assigned IDs as values.

GEMS.severe! — Method

severe!(individual::Individual, severe::Bool)

Sets the severe flag of the individual.

GEMS.severeness_offset — Method

severeness_offset(individual::Individual)

Returns an individual's last severeness offset tick. Return -1 if never severe.

GEMS.severeness_offset! — Method

severeness_offset!(individual::Individual, tick::Int16)

Sets an individual's severeness offset tick.

GEMS.severeness_onset — Method

severeness_onset(individual::Individual)

Returns an individual's last severeness onset tick. Return -1 if never severe.

GEMS.severeness_onset! — Method

severeness_onset!(individual::Individual, tick::Int16)

Sets an individual's severeness onset tick.

GEMS.sex — Method

sex(individual::Individual)

Return an individual's sex.

GEMS.social_factor! — Method

social_factor!(individual::Individual, val::Float32)

Overwrites the individual's social_factor attribute.

GEMS.social_factor — Method

social_factor(individual::Individual)

Returns an individual's social_factor value.

GEMS.symptom_onset — Method

symptom_onset(individual::Individual)

Returns an individual's last symptom onset tick. Return -1 if never symptomatic.

GEMS.symptom_onset! — Method

symptom_onset!(individual::Individual, tick::Int16)

Sets an individual's symptom onset tick.

GEMS.symptomatic! — Method

symptomatic!(individual::Individual, symptomatic::Bool)

Sets the symptomatic flag of the individual.

GEMS.vaccination_tick — Method

vaccination_tick(individual::Individual)

Returns the time of last vaccination.

GEMS.vaccine_id — Method

vaccine_id(individual::Individual)

Returns the id of the vaccine the individual is vaccinated with.

GEMS.ventilated! — Method

ventilated!(individual::Individual, ventilated::Bool)

Sets the ventilated flag of the individual.

GEMS.ventilation_admission — Method

ventilation_admission(individual::Individual)

Returns an individual's last ventilation admission tick. Return -1 if never admitted to ventilation.

GEMS.ventilation_admission! — Method

ventilation_admission!(individual::Individual, tick::Int16)

Sets an individual's ventilation admission tick.

GEMS.ventilation_discharge — Method

ventilation_discharge(individual::Individual)

Returns an individual's last ventilation discharge tick. Return -1 if never discharged from ventilation.

GEMS.ventilation_discharge! — Method

ventilation_discharge!(individual::Individual, tick::Int16)

Sets an individual's ventilation discharge tick.