Individuals

hier eine subsection: wie fragt man den Status der agents ab (hier alle funktionen, wie man states queried)

Overview Structs

GEMS.Agent

GEMS.Individual

GEMS.Individual

Overview Functions

GEMS.age

GEMS.class_id

GEMS.comorbidities

GEMS.critical

GEMS.critical!

GEMS.dead

GEMS.death_tick

GEMS.death_tick!

GEMS.disease_state

GEMS.education

GEMS.end_quarantine!

GEMS.estimate_disease_progression

GEMS.exposed

GEMS.exposed_tick

GEMS.exposed_tick!

GEMS.has_municipality

GEMS.hospital_status

GEMS.hospitalize!

GEMS.hospitalized

GEMS.hospitalized_tick

GEMS.hospitalized_tick!

GEMS.household_id

GEMS.icu

GEMS.icu!

GEMS.icu_tick

GEMS.icu_tick!

GEMS.id

GEMS.individuals

GEMS.individuals!

GEMS.infect!

GEMS.infected

GEMS.infection_id

GEMS.infectious

GEMS.infectious_tick

GEMS.infectious_tick!

GEMS.infectiousness

GEMS.infectiousness!

GEMS.is_student

GEMS.is_working

GEMS.isdetected

GEMS.isquarantined

GEMS.isvaccinated

GEMS.kill!

GEMS.last_reported_at

GEMS.last_reported_at!

GEMS.last_test

GEMS.last_test!

GEMS.last_test_result

GEMS.last_test_result!

GEMS.mandate_compliance

GEMS.mandate_compliance!

GEMS.municipality

GEMS.municipality_id

GEMS.num_of_infected

GEMS.number_of_infections

GEMS.number_of_vaccinations

GEMS.occupation

GEMS.office

GEMS.office_id

GEMS.onset_of_severeness

GEMS.onset_of_severeness!

GEMS.onset_of_symptoms

GEMS.onset_of_symptoms!

GEMS.pathogen_id

GEMS.presymptomatic

GEMS.presymptomatic!

GEMS.progress_disease!

GEMS.quarantine_release_tick

GEMS.quarantine_release_tick!

GEMS.quarantine_status

GEMS.quarantine_tick

GEMS.quarantine_tick!

GEMS.recover!

GEMS.removed_tick

GEMS.removed_tick!

GEMS.reset!

GEMS.sample_critical_death_rate

GEMS.sample_hospitalization_rate

GEMS.sample_icu_rate

GEMS.sample_individuals

GEMS.sample_infectious_offset

GEMS.sample_length_of_stay

GEMS.sample_mild_death_rate

GEMS.sample_onset_of_severeness

GEMS.sample_onset_of_symptoms

GEMS.sample_self_quarantine_rate

GEMS.sample_severe_death_rate

GEMS.sample_time_to_hospitalization

GEMS.sample_time_to_icu

GEMS.sample_time_to_recovery

GEMS.sample_ventilation_rate

GEMS.setting_id

GEMS.severe

GEMS.severe!

GEMS.sex

GEMS.social_factor

GEMS.social_factor!

GEMS.symptom_category

GEMS.symptom_category!

GEMS.symptomatic

GEMS.symptomatic!

GEMS.update_individual!

GEMS.vaccination_tick

GEMS.vaccine_id

GEMS.ventilate!

GEMS.ventilated

GEMS.ventilation_tick

GEMS.ventilation_tick!

Structs

GEMS.Agent — Type

Supertype for simulation agents

GEMS.Individual — Type

Individual <: Agent

A type to represent individuals, that act as agents inside the simulation.

Fields

General

id::Int32: Unique identifier of the individual

sex::Int8: Sex (Female (1), Male(2), Diverse (3))

age::Int8: Age

education::Int8: Education class (i.e. highest degree)

occupation::Int8: Occupation class (i.e. manual labour, office job, etc...)

Behaviour

social_factor::Float32: Parameter for the risk-willingness. Can be anywhere between -1 and 1 with neutral state is 0.

mandate_compliance::Float32. Paremeter which influences the probability of complying to mandates. Can be anywhere between -1 and 1 with neutral state is 0.

Health Status

comorbidities::Vector{Bool}: Indicating prevalence of certain health conditions. True, if the individual is preconditioned with the comorbidity associated to the array index.

dead::Bool: Flag indicating individual's decease

hospital_status::Int8: State in the hopsital (0 = not hospitalized, 1 = hospitalized, 2 = ventilation, 3 = ICU)

Associated Settings

household::Int32: Reference to household id

office::Int32: Reference to office id

schoolclass::Int32: Reference to schoolclass id

municipality::Int32: Reference to municipality id

Pathogen

pathogen_id::Int8: pathogen identifier

infection_id::Int32: Current infection id

disease_state::Int8: Current State in natural disease history (0 = not infected, 1 = Presymptomatic, 2 = Symptomatic, 3 = Severe, 4 = Critical)

symptom_category::Int8: Endstate in current disease progression. The numbers should align with disease_state, but can be interpreted as the symptom category of this case (0 = None, 1 = Asymptomatic, 2 = Mild, 3 = Severe, 4 = Critical)

infectiousness::Int8: an individuals infectiousness (1-127), i.e. for superspreaders

number_of_infections::Int8: infection count

Natural Disease History

exposed_tick::Int16: Tick of most recent infection

infectious_tick::Int16: Tick of most recent change into "infectious" state (considered asymptomatic)

onset_of_symptoms::Int16: Tick of the onset of symptoms

onset_of_severeness::Int16: Tick of onset of severe symptoms

hospitalized_tick::Int16: Tick of hospitalization

ventilation_tick::Int16: Tick of ventilation

icu_tick::Int16: Tick of ICU (Intensive Care Unit)

death_tick::Int16: Tick of Death

removed_tick::Int16: Tick of most recent removal event

Testing

last_test::Int16: Tick of last test for pathogen

last_test_result::Bool: Flag for positivity of last test

last_reported_at::Int16: Tick at which this individual was last reported

Vaccination

vaccine_id::Int8: Vaccine identifier

number_of_vaccinations::Int8: Individual's vaccination counter

vaccination_tick::Int16: Tick of most recent vaccination

Interventions

quarantine_status::Int8: Status to indicate quarantine (none, household_quarantined, hospitalized, etc...)

quarantine_tick::Int16: Start tick of quarantine

quarantine_release_tick::Int16: End tick of quarantine

Constructors

GEMS.Individual — Method

Individual(properties::Dict)

Create an individual with the provided properties. Properties must have at least keys id, sex, age.

Functions

GEMS.age — Function

age(individual::Individual)

Return an individual's age.

GEMS.class_id — Function

class_id(individual::Individual)

Returns an individual's associated class's ID.

GEMS.comorbidities — Function

comorbidities(individual::Individual)

Returns an individual's comorbidities.

GEMS.critical! — Method

critical!(individual::Individual)

Marks the individual's condition as critical.

GEMS.critical — Method

critical(individual::Individual)

Returns wether the individual's condition is critical.

GEMS.dead — Method

dead(individual::Individual)

Returns an individual's death flag.

GEMS.death_tick! — Method

death_tick!(individual::Individual, tick::Int16)

Sets an individual's death tick.

GEMS.death_tick — Method

death_tick(individual::Individual)

Returns an individual's death tick.

GEMS.disease_state — Method

disease_state(individual::Individual)

Returns an individual's disease status (currently infected).

GEMS.education — Method

education(individual::Individual)

Return an individual's education class

GEMS.end_quarantine! — Method

end_quarantine!(individual::Individual)

Ends an individuals quarantine.

GEMS.estimate_disease_progression — Method

estimate_disease_progression(dpr::DiseaseProgressionStrat, indiv::Individual)

Estimates the final status an individual reaches during the progression of a disease.

GEMS.exposed — Method

exposed(individual::Individual)

Returns wether the individual is exposed (infected, but not yet infectious).

GEMS.exposed_tick! — Method

exposed_tick!(individual::Individual, tick::Int16)

Sets an individual's exposed tick.

GEMS.exposed_tick — Method

exposed_tick(individual::Individual)

Returns an individual's exposed tick (currently infected).

GEMS.has_municipality — Method

has_municipality(individual::Individual)

Returns true if individual is assigned to an instance of type Municipality.

GEMS.hospital_status — Method

hospital_status(individual::Individual)

Returns the hospital_status of the individual.

GEMS.hospitalize! — Method

hospitalize!(individual::Individual)

Hospitalizes the individual.

GEMS.hospitalized — Method

hospitalized(individual::Individual)

Returns wether the individual is hospitalized.

GEMS.hospitalized_tick! — Method

hospitalized_tick!(individual::Individual, tick::Int16)

Sets an individual's tick for when it gets hospitalized.

GEMS.hospitalized_tick — Method

hospitalized_tick(individual::Individual)

Returns an individual's tick for when it gets hospitalized.

GEMS.household_id — Method

household_id(individual::Individual)

Returns an individual's associated household's ID.

GEMS.icu! — Method

icu!(individual::Individual)

Sets the individual to be in ICU.

GEMS.icu — Method

icu(individual::Individual)

Returns wether the individual is in ICU.

GEMS.icu_tick! — Method

icu_tick!(individual::Individual, tick::Int16)

Sets an individual's tick for when it will be delivered into icu.

GEMS.icu_tick — Method

icu_tick(individual::Individual)

Returns an individual's tick when it will be delivered into icu.

GEMS.id — Method

id(individual::Individual)

Return the unique identifier of the individual.

GEMS.individuals! — Function

individuals!(setting::IndividualSetting, indivs::Vector{Individual}, simulation::Simulation)

Appends the individuals associated with a given IndividualSetting to the provided indivs vector.

Parameters

setting::IndividualSetting: Setting to get the individuals from

indivs::Vector{Individual}: List that will be appeneded with the setting's individuals

simulation::Simulation: Simulation object

individuals!(setting::ContainerSetting, indivs::Vector{Individual}, simulation::Simulation)

Appends the individuals associated with a given ContainerSetting to the provided indivs vector by recursively calling the individuals! function.

Parameters

setting::ContainerSetting: Setting to get the individuals from

indivs::Vector{Individual}: List that will be appeneded with the setting's individuals

simulation::Simulation: Simulation object

GEMS.individuals — Function

individuals(population::Population)

Return the individuals associated with the population.

individuals(setting::IndividualSetting)

Returns the individuals associated with the given setting.

individuals(setting::IndividualSetting, simulation::Simulation)

Returns the individuals associated with a given IndividualSetting.

individuals(setting::ContainerSetting, indivs::Vector{Individual}, simulation::Simulation)::Vector{Individual}

Returns the individuals associated with a given ContainerSetting by recursively getting the individuals of all contained settings using the individuals! function.

GEMS.infect! — Function

infect!(infectee::Individual, tick::Int16, pathogen::Pathogen;
    sim::Union{Simulation, Nothing} = nothing,
    infecter_id::Int32 = Int32(-1), setting_id::Int32 = Int32(-1), lon::Float32 = NaN32,
    lat::Float32 = NaN32, setting_type::Char = '?', ags::Int32 = Int32(-1),
    source_infection_id::Int32 = DEFAULT_INFECTION_ID)

Infect infectee with the specified pathogen and calculate time to infectiousness and time to recovery. Optional arguments infecter_id. setting_id, and setting_type can be passed for logging. It's not required to calulate the infection. The infection can only be logged, if Simulation object is passed (as this object holds the logger).

Parameters

infectee::Individual: Individual to infect

tick::Int16: Infection tick

pathogen::Pathogen: Pathogen to infect the individual with

sim::Simulation: Simulation object (used to get logger and current tick)

sim::Union{Simulation, Nothing} = nothing (optional) = Simulation object (used to get logger)

infecter_id::Int32 = Int32(-1) (optional): Infecting individual

setting_id::Int32 = Int32(-1) (optional): ID of setting this infection happens in

lon::Float32 = NaN32 (optional): Longitude of the infection infection location (setting)

lat::Float32 = NaN32 (optional): Latitude of the infection infection location (setting)

setting_type::Char = '?' (optional): Setting type as char (e.g. "h" for Household)

ags::Int32 = Int32(-1) (optional): Amtlicher Gemeindeschlüssel (community identification number) of the region this infection happened in as Integer value

source_infection_id::Int32 = DEFAULT_INFECTION_ID (optional): Current infection ID of the infecting individual

Returns

Int32: New infection ID

GEMS.infected — Method

infected(individual::Individual)

Returns whether an individual is currently infected.

GEMS.infection_id — Method

infection_id(individual::Individual)

Returns an individual's infection_id (currently infected).

GEMS.infectious — Method

infectious(individual::Individual)

Returns whether an individual is currently infectious.

GEMS.infectious_tick! — Method

infectious_tick!(individual::Individual, tick::Int16)

Sets an individual's infectious tick.

GEMS.infectious_tick — Method

infectious_tick(individual::Individual)

Returns an individual's infectious tick (currently infected).

GEMS.infectiousness! — Method

infectiousness!(individual::Individual, infectiousness)

Assigns a specified infectiousness (0-127) to an individual.

GEMS.infectiousness — Method

infectiousness(individual::Individual)

Returns an individual's infectiousness (currently infected).

GEMS.is_student — Method

is_student(individual::Individual)

Returns true if individual is assigned to an instance of type SchoolClass.

GEMS.is_working — Method

is_working(individual::Individual)

Returns true if individual is assigned to an instance of type Office.

GEMS.isdetected — Method

isdetected(individual::Individual)

Returns true if an individual was currently infected and already reported.

GEMS.isquarantined — Method

isquarantined(individual::Individual)

Returns wether the individual is in quarantine or not.

Missing docstring.

Missing docstring for issubset(::Individual, ::Individual). Check Documenter's build log for details.

GEMS.isvaccinated — Method

isvaccinated(individual::Individual)

Returns wether the individual is vaccinated.

GEMS.kill! — Method

kill!(individual::Individual)

Kills the individual.

GEMS.last_reported_at! — Method

last_reported_at!(individual::Individual, report_tick::Int16)

Sets last tick this individual was last reported.

GEMS.last_reported_at — Method

last_reported_at(individual::Individual)

Returns the last tick this individual was a reported case.

GEMS.last_test! — Method

last_test!(individual::Individual, tick::Int16)

Sets last test date (tick).

GEMS.last_test — Method

last_test(individual::Individual)

Returns last test date (tick).

GEMS.last_test_result! — Method

last_test_result!(individual::Individual, test_result::Bool)

Sets last test result.

GEMS.last_test_result — Method

last_test_result(individual::Individual)

Returns whether last test was positive. Defaults to false.

GEMS.mandate_compliance! — Method

mandate_compliance!(individual::Individual, val::Float32)

Overwrites the individual's mandate_compliance attribute.

GEMS.mandate_compliance — Method

mandate_compliance(individual::Individual)

Return an individual's mandate_compliance value.

GEMS.municipality — Method

municipality(i::Individual, sim::Simulation)::Municipality

Returns the Municipality instance referenced in an individual.

GEMS.municipality_id — Method

municipality_id(individual::Individual)

Returns an individual's associated municipalities ID.

GEMS.num_of_infected — Method

num_of_infected(individuals::Vector{Individual})

Takes a vector of individuals and returns the number of infected individuals.

GEMS.number_of_infections — Method

number_of_infections(individual::Individual)

Returns an individual's number of infections (currently infected).

GEMS.number_of_vaccinations — Method

number_of_vaccinations(individual::Individual)

Returns the number of vaccinations.

GEMS.occupation — Method

occupation(individual::Individual)

Returns an individual's occupation class.

GEMS.office — Method

office(i::Individual, sim::Simulation)::Office

Returns the Office instance referenced in an individual.

GEMS.office_id — Method

office_id(individual::Individual)

Returns an individual's associated office's ID.

GEMS.onset_of_severeness! — Method

onset_of_severeness!(individual::Individual, tick::Int16)

Sets an individual's tick for the onset of severe symptoms, if it's a severe case.

GEMS.onset_of_severeness — Method

onset_of_severeness(individual::Individual)

Returns an individual's tick for the onset of severe symptoms, if it's a severe case.

GEMS.onset_of_symptoms! — Method

onset_of_symptoms!(individual::Individual, tick::Int16)

Sets an individual's tick for the onset of symptoms, if symptomatic.

GEMS.onset_of_symptoms — Method

onset_of_symptoms(individual::Individual)

Returns an individual's tick for the onset of symptoms, if symptomatic.

GEMS.pathogen_id — Method

pathogen_id(individual::Individual)

Returns an individual's pathogen (currently infected).

GEMS.presymptomatic! — Method

presymptomatic!(individual::Individual)

Marks the individual as asymptomatic, but infectious.

GEMS.presymptomatic — Method

presymptomatic(individual::Individual)

Returns if individual is asymptomatic.

GEMS.progress_disease! — Method

progress_disease!(indiv::Individual, tick::Int16, lookup_dict::Dict = SYMPTOM_CATEGORY_DICT)

Progresses the disease of the (infected) individual. The dictionary lookup_dict can be provided, if custom disease compartements were implemented. The key-value-pairs should align with the meaning of the parameter disease_state or symptom_category. This means, that if disease_state==1 means "Asymptomatic", that status[1]=="Asymptomatic".

This function doesn't kill agents as this is handled a level higher for the simulation to log deaths.

GEMS.quarantine_release_tick! — Method

quarantine_release_tick!(individual::Individual, tick::Int16)

Sets an individual's quarantine release tick.

GEMS.quarantine_release_tick — Method

quarantine_release_tick(individual::Individual)

Returns an individual's quarantine release tick.

GEMS.quarantine_status — Method

quarantine_status(individual::Individual)

Returns an individuals quarantine status.

GEMS.quarantine_tick! — Method

quarantine_tick!(individual::Individual, tick::Int16)

Sets an individual's quarantine tick.

GEMS.quarantine_tick — Method

quarantine_tick(individual::Individual)

Returns an individual's quarantine tick.

GEMS.recover! — Method

recover!(individual::Individual)

Recovers the individual from its infection.

GEMS.removed_tick! — Method

removed_tick!(individual::Individual, tick::Int16)

Sets an individual's removed tick (currently infected).

GEMS.removed_tick — Method

removed_tick(individual::Individual)

Returns an individual's removed tick (currently infected).

GEMS.reset! — Method

reset!(individual::Individual)

Resets all non-static values like the disease progression timing. The individual will get back into a state where it was never infected, vaccinated, tested, etc.

GEMS.sample_critical_death_rate — Method

sample_critical_death_rate(pathogen::Pathogen, indiv::Individual)

Returns a randomly drawn value from the death rate distribution for critical disease progressions.

GEMS.sample_hospitalization_rate — Method

sample_hospitalization_rate(pathogen::Pathogen, indiv::Individual)

Returns a randomly drawn value from the hospitalization rate distribution.

GEMS.sample_icu_rate — Method

sample_icu_rate(pathogen::Pathogen, indiv::Individual)

Returns a randomly drawn value from the ICU rate distribution.

GEMS.sample_individuals — Function

sample_individuals(individuals::Vector{Individual}, n::Int64)

Returns a subsample of a vector of Individuals of sample size n.

sample_individuals(setting::IndividualSetting, n::Int64)

Returns a subsample of the setting's Individuals of sample size n.

GEMS.sample_infectious_offset — Method

sample_infectious_offset(pathogen::Pathogen,  indiv::Individual)

Returns a randomly drawn (rounded) value from the infectious offset distribution.

GEMS.sample_length_of_stay — Method

sample_length_of_stay(pathogen::Pathogen,  indiv::Individual)

Returns a randomly drawn (rounded) value from the time_to_hopsitalization distribution for cases with severe disease progression.

GEMS.sample_mild_death_rate — Method

sample_mild_death_rate(pathogen::Pathogen, indiv::Individual)

Returns a randomly drawn value from the death rate distribution for cases with mild symptoms.

GEMS.sample_onset_of_severeness — Method

sample_onset_of_severeness(pathogen::Pathogen,  indiv::Individual)

Returns a randomly drawn (rounded) value from the onset_of_symptoms distribution.

GEMS.sample_onset_of_symptoms — Method

sample_onset_of_symptoms(pathogen::Pathogen,  indiv::Individual)

Returns a randomly drawn (rounded) value from the onset_of_symptoms distribution.

GEMS.sample_self_quarantine_rate — Method

sample_self_quarantine_rate(pathogen::Pathogen, indiv::Individual)

Returns a randomly drawn value from the ICU rate distribution.

GEMS.sample_severe_death_rate — Method

sample_severe_death_rate(pathogen::Pathogen, indiv::Individual)

Returns a randomly drawn value from the death rate distribution for severe disease progressions.

GEMS.sample_time_to_hospitalization — Method

sample_time_to_hospitalization(pathogen::Pathogen,  indiv::Individual)

Returns a randomly drawn (rounded) value from the time_to_hospitalization distribution for cases with severe disease progression.

GEMS.sample_time_to_icu — Method

sample_time_to_icu(pathogen::Pathogen,  indiv::Individual)

Returns a randomly drawn (rounded) value from the time_to_icu distribution for cases with critical disease progression.

GEMS.sample_time_to_recovery — Method

sample_time_to_recovery(pathogen::Pathogen,  indiv::Individual)

Returns a randomly drawn (rounded) value from the time_to_recovery distribution.

GEMS.sample_ventilation_rate — Method

sample_ventilation_rate(pathogen::Pathogen, indiv::Individual)

Returns a randomly drawn value from the Ventilation rate distribution.

Missing docstring.

Missing docstring for setting_id!(::Individual, ::DataType, ::Int32). Check Documenter's build log for details.

GEMS.setting_id — Method

setting_id(individual::Individual, type::DataType)

Returns the id of the setting of type associated with the individual. If the settingtype is unknown or the agent isn't part of a setting of that type, -1 will be returned.

GEMS.severe! — Method

severe!(individual::Individual)

Marks the individual's disease to be severe.

GEMS.severe — Method

severe(individual::Individual)

Returns wether the individual's disease is severe.

GEMS.sex — Method

sex(individual::Individual)

Return an individual's sex.

GEMS.social_factor! — Method

social_factor!(individual::Individual, val::Float32)

Overwrites the individual's social_factor attribute.

GEMS.social_factor — Method

social_factor(individual::Individual)

Returns an individual's social_factor value.

GEMS.symptom_category! — Method

symptom_category!(individual::Individual, category::Int8)

Sets an infected individuals symptom_category.

GEMS.symptom_category — Method

symptom_category(individual::Individual)

Returns an infected individual's symptom category.

GEMS.symptomatic! — Method

symptomatic!(individual::Individual)

Marks the individual as symptomatic.

GEMS.symptomatic — Method

symptomatic(individual::Individual)

Returns wether the individual is symptomatic.

GEMS.update_individual! — Method

update_individual!(indiv::Individual, tick::Int16, sim::Simulation)

Update the individual disease progression, handle its recovery and log its possible death. If the individual is not infected, this function will just return.

Parameters

indiv::Individual: Individual to update

tick::Int16: Current tick

sim::Simulation: Simulation object

GEMS.vaccination_tick — Method

vaccination_tick(individual::Individual)

Returns the time of last vaccination.

GEMS.vaccine_id — Method

vaccine_id(individual::Individual)

Returns the id of the vaccine the individual is vaccinated with.

GEMS.ventilate! — Method

ventilate!(individual::Individual)

Sets the individual to require ventilation.

GEMS.ventilated — Method

ventilated(individual::Individual)

Returns wether the individual needs ventilation.

GEMS.ventilation_tick! — Method

ventilation_tick!(individual::Individual, tick::Int16)

Sets an individual's tick for when it gets ventilated.

GEMS.ventilation_tick — Method

ventilation_tick(individual::Individual)

Returns an individual's tick when it gets ventilated.