Infections and Progressions

Overview Structs

GEMS.DiseaseProgression

Overview Functions

GEMS.death

GEMS.exposure

GEMS.hospital_admission

GEMS.hospital_discharge

GEMS.icu_admission

GEMS.icu_discharge

GEMS.infect!

GEMS.infectiousness_onset

GEMS.is_asymptomatic

GEMS.is_dead

GEMS.is_hospitalized

GEMS.is_icu

GEMS.is_infected

GEMS.is_infectious

GEMS.is_mild

GEMS.is_presymptomatic

GEMS.is_recovered

GEMS.is_severe

GEMS.is_symptomatic

GEMS.is_ventilated

GEMS.recovery

GEMS.severeness_offset

GEMS.severeness_onset

GEMS.spread_infection!

GEMS.symptom_onset

GEMS.try_to_infect!

GEMS.update_individual!

GEMS.ventilation_admission

GEMS.ventilation_discharge

Structs

GEMS.DiseaseProgression — Type

DiseaseProgression

A struct to represent the disease progression of an infectious disease.

Fields

pathogen_id::Int8: The ID of the pathogen associated with this disease progression.

exposure::Int16: Exposure (infection) tick of the disease.

infectiousness_onset::Int16: Tick when the individual becomes infectious.

symptom_onset::Int16: Tick when the individual develops symptoms.

severeness_onset::Int16: Tick when the individual develops severe symptoms.

hospital_admission::Int16: Tick when the individual is admitted to hospital.

icu_admission::Int16: Tick when the individual is admitted to ICU.

icu_discharge::Int16: Tick when the individual is discharged from ICU.

hospital_discharge::Int16: Tick when the individual is discharged from hospital.

recovery::Int16: Tick when the individual recovers from the disease.

death::Int16: Tick when the individual dies from the disease.

Constraints

exposure must be non-negative.

infectiousness_onset must be at least one tick after exposure.

symptom_onset is optional but must be at least one tick after exposure if set.

severeness_onset is optional, but if set, requires symptom_onset and cannot be prior to symptom_onset.

hospital_admission is optional, but if set, requires severeness_onset, cannot be prior to severeness_onset, and requires hospital_discharge.

icu_admission is optional, but if set, requires hospital_admission, cannot be prior to hospital_admission, and requires ICU_discharge.

icu_discharge is optional, but if set, requires ICU_admission, and cannot be prior to ICU_admission.

ventilation_admission is optional, but if set, requires hospital_admission, cannot be prior to hospital_admission, and requires ventilation_discharge.

ventilation_discharge is optional, but if set, requires ventilation_admission, and cannot be prior to ventilation_admission.

hospital_discharge is optional, but if set, requires hospital_admission, cannot be prior to hospital_admission, and cannot be prior to ICU_discharge if set.

severeness_offset is optional, but if set, requires severeness_onset, cannot be prior to severeness_onset, and cannot be prior to hospital_discharge if set.

recovery and death are mutually exclusive; at least one must be set.

recovery must be at least one tick after exposure, symptom_onset if set, severeness_onset if set, and cannot be prior to hospital_discharge.

death requires symptom_onset and cannot precede any other disease progression events.

Examples

The following example represents a mild symptomatic progression. An individual get infected at time 23, becomes infectious at 25, develops symptoms at 26, recovers at 31, and does not die.

julia> DiseaseProgression(exposure = 23, infectiousness_onset = 25, symptom_onset = 26, recovery  = 31)

Output:

DiseaseProgression(
  tick | event
    23 | exposure
    25 | infectiousness_onset
    26 | symptom_onset
    31 | recovery
)

Functions

Missing docstring.

Missing docstring for can_infect. Check Documenter's build log for details.

GEMS.death — Method

death(dp::DiseaseProgression)

Returns the tick when the individual with disease progression dp dies from the disease. If the individual recovers instead, this will return -1.

GEMS.exposure — Method

exposure(dp::DiseaseProgression)

Returns the exposure (infection) tick of the disease progression dp.

GEMS.hospital_admission — Method

hospital_admission(dp::DiseaseProgression)

Returns the tick when the individual with disease progression dp is admitted to hospital. If the individual is never admitted to hospital, this will return -1.

GEMS.hospital_discharge — Method

hospital_discharge(dp::DiseaseProgression)

Returns the tick when the individual with disease progression dp is discharged from hospital. If the individual is never admitted to hospital, this will return -1.

GEMS.icu_admission — Method

icu_admission(dp::DiseaseProgression)

Returns the tick when the individual with disease progression dp is admitted to ICU. If the individual is never admitted to ICU, this will return -1.

GEMS.icu_discharge — Method

icu_discharge(dp::DiseaseProgression)

Returns the tick when the individual with disease progression dp is discharged from ICU. If the individual is never admitted to ICU, this will return -1.

GEMS.infect! — Function

infect!(infectee::Individual,
    tick::Int16,
    pathogen::Pathogen;
    sim::Union{Simulation, Nothing} = nothing,
    rng::AbstractRNG = Random.default_rng(),
    infecter_id::Int32 = Int32(-1),
    setting_id::Int32 = Int32(-1),
    lon::Float32 = NaN32,
    lat::Float32 = NaN32,
    setting_type::Char = '?',
    ags::Int32 = Int32(-1),
    source_infection_id::Int32 = DEFAULT_INFECTION_ID)

Infect infectee with the specified pathogen and calculate time to infectiousness and time to recovery. Optional arguments infecter_id. setting_id, and setting_type can be passed for logging. It's not required to calulate the infection. The infection can only be logged, if Simulation object is passed (as this object holds the logger).

Parameters

infectee::Individual: Individual to infect

tick::Int16: Infection tick

pathogen::Pathogen: Pathogen to infect the individual with

sim::Union{Simulation, Nothing} = nothing (optional) = Simulation object (used to get logger)

rng::AbstractRNG = Random.default_rng() (optional): RNG to use for stochastic parts

infecter_id::Int32 = Int32(-1) (optional): Infecting individual

setting_id::Int32 = Int32(-1) (optional): ID of setting this infection happens in

lon::Float32 = NaN32 (optional): Longitude of the infection infection location (setting)

lat::Float32 = NaN32 (optional): Latitude of the infection infection location (setting)

setting_type::Char = '?' (optional): Setting type as char (e.g. "h" for Household)

ags::Int32 = Int32(-1) (optional): Amtlicher Gemeindeschlüssel (community identification number) of the region this infection happened in as Integer value

source_infection_id::Int32 = DEFAULT_INFECTION_ID (optional): Current infection ID of the infecting individual

Returns

Int32: New infection ID

infect!(infectee::Individual, sim::Simulation)

Infect infectee with the pathogen of the simulation at the current tick of the simulation. Mainly a convenience wrapper around infect! with less parameters. Used for example in test cases.

GEMS.infectiousness_onset — Method

infectiousness_onset(dp::DiseaseProgression)

Returns the tick when the individual with disease progression dp becomes infectious.

GEMS.is_asymptomatic — Method

is_asymptomatic(dp::DiseaseProgression, t::Int16)
isasymptomatic(dp::DiseaseProgression, t::Int16)
asymptomatic(dp::DiseaseProgression, t::Int16)

Returns true if the individual with disease progression dp is asymptomatic at tick t, otherwise false. Asymptomatic means the individual is infected, does currently not have symptoms and will not develop symptoms in the future.

GEMS.is_dead — Method

is_dead(dp::DiseaseProgression, t::Int16)
isdead(dp::DiseaseProgression, t::Int16)
dead(dp::DiseaseProgression, t::Int16)

Returns true if the individual with disease progression dp has died before tick t, otherwise false.

GEMS.is_hospitalized — Method

is_hospitalized(dp::DiseaseProgression, t::Int16)
ishospitalized(dp::DiseaseProgression, t::Int16)
hospitalized(dp::DiseaseProgression, t::Int16)

Returns true if the individual with disease progression dp is in hospital at tick t, otherwise false.

GEMS.is_icu — Method

is_icu(dp::DiseaseProgression, t::Int16)
isicu(dp::DiseaseProgression, t::Int16)
icu(dp::DiseaseProgression, t::Int16)

Returns true if the individual with disease progression dp is in the ICU at tick t, otherwise false.

GEMS.is_infected — Method

is_infected(dp::DiseaseProgression, t::Int16)
isinfected(dp::DiseaseProgression, t::Int16)
infected(dp::DiseaseProgression, t::Int16)

Returns true if the individual with disease progression dp is infected at tick t, otherwise false.

GEMS.is_infectious — Method

is_infectious(dp::DiseaseProgression, t::Int16)
isinfectious(dp::DiseaseProgression, t::Int16)
infectious(dp::DiseaseProgression, t::Int16)

Returns true if the individual with disease progression dp is infectious at tick t, otherwise false.

GEMS.is_mild — Method

is_mild(dp::DiseaseProgression, t::Int16)
ismild(dp::DiseaseProgression, t::Int16)
mild(dp::DiseaseProgression, t::Int16)

Returns true if the individual with disease progression dp has mild symptoms at tick t, otherwise false. A mild case is defined as symptomatic but not severe.

GEMS.is_presymptomatic — Method

is_presymptomatic(dp::DiseaseProgression, t::Int16)
ispresymptomatic(dp::DiseaseProgression, t::Int16)
presymptomatic(dp::DiseaseProgression, t::Int16)

Returns true if the individual with disease progression dp is presymptomatic at tick t, otherwise false. Presymptomatic means the individual is infected but has not yet developed symptoms (but will in the future).

GEMS.is_recovered — Method

is_recovered(dp::DiseaseProgression, t::Int16)
isrecovered(dp::DiseaseProgression, t::Int16)
recovered(dp::DiseaseProgression, t::Int16)

Returns true if the individual with disease progression dp has recovered before tick t, otherwise false.

GEMS.is_severe — Method

is_severe(dp::DiseaseProgression, t::Int16)
issevere(dp::DiseaseProgression, t::Int16)
severe(dp::DiseaseProgression, t::Int16)

Returns true if the individual with disease progression dp has severe symptoms at tick t, otherwise false.

GEMS.is_symptomatic — Method

is_symptomatic(dp::DiseaseProgression, t::Int16)
issymptomatic(dp::DiseaseProgression, t::Int16)
symptomatic(dp::DiseaseProgression, t::Int16)

Returns true if the individual with disease progression dp is symptomatic at tick t, otherwise false.

GEMS.is_ventilated — Method

is_ventilated(dp::DiseaseProgression, t::Int16)
isventilated(dp::DiseaseProgression, t::Int16)
ventilated(dp::DiseaseProgression, t::Int16)

Returns true if the individual with disease progression dp is on a ventilator at tick t, otherwise false.

GEMS.recovery — Method

recovery(dp::DiseaseProgression)

Returns the tick when the individual with disease progression dp recovers from the disease. If the individual dies instead, this will return -1.

GEMS.severeness_onset — Method

severeness_onset(dp::DiseaseProgression)

Returns the tick when the individual with disease progression dp develops severe symptoms. If the individual will not develop severe symptoms, this will return -1.

GEMS.severeness_offset — Method

severeness_offset(dp::DiseaseProgression)

Returns the tick when the individual with disease progression dp's severeness is reduced (i.e., moves from severe to mild symptoms). If the individual never has severe symptoms, this will return -1.

GEMS.spread_infection! — Method

spread_infection!(setting::Setting, sim::Simulation, pathogen::Pathogen)

Spreads the infection of pathogen inside the provided setting. This will simulate the infection dynamics at the time tick(sim) inside setting within the context of the simulation sim. This will also update all settings, the individual is part of, if the infection is successful.

Parameters

setting::Setting: Setting in which the pathogen shall be spreaded

sim::Simulation: Simulation object

pathogen::Pathogen: Pathogen to spread

GEMS.symptom_onset — Method

symptom_onset(dp::DiseaseProgression)

Returns the tick when the individual with disease progression dp develops symptoms. If the individual is asymptomatic, this will return -1.

GEMS.try_to_infect! — Function

try_to_infect!(infctr::Individual, infctd::Individual, sim::Simulation, pathogen::Pathogen, setting::Setting;
    source_infection_id::Int32 = DEFAULT_INFECTION_ID)

Tries to infect the infctd with the given pathogen transmitted by infctrat time tick(sim) with sim being the simulation. Success depends on whether the agent is alive, not already infected an whether an infection event was sampled using the provided distribution or probability. Returns true if infection was successful.

Parameters

infctr::Individual: Infecting individual

infctd::Individual: Individual to infect

sim::Simulation: Simulation object

pathogen::Pathogen: Pathogen to infect the individual with

setting::Setting: Setting this infection happens in

source_infection_id::Int32 = DEFAULT_INFECTION_ID (optional): Current infection ID of the infecting individual

Returns

Bool: True if infection was successful, false otherwise

GEMS.update_individual! — Method

update_individual!(indiv::Individual, tick::Int16, sim::Simulation)

Update the individual disease progression, handle its recovery and log its possible death. If the individual is not infected, this function will just return.

Parameters

indiv::Individual: Individual to update

tick::Int16: Current tick

sim::Simulation: Simulation object

GEMS.ventilation_admission — Method

ventilation_admission(dp::DiseaseProgression)

Returns the tick when the individual with disease progression dp is admitted to ventilation. If the individual is never admitted to ventilation, this will return -1.

GEMS.ventilation_discharge — Method

ventilation_discharge(dp::DiseaseProgression)

Returns the tick when the individual with disease progression dp is discharged from ventilation. If the individual is never admitted to ventilation, this will return -1.