Infections and Immunity

Overview Structs

GEMS.AgeDependentTransmissionRate

GEMS.ConstantTransmissionRate

Overview Functions

GEMS.infect!

GEMS.spread_infection!

GEMS.transmission_probability

GEMS.try_to_infect!

GEMS.update_individual!

Structs

GEMS.AgeDependentTransmissionRate — Type

AgeDependentTransmissionRate <: TransmissionFunction

A TransmissionFunction type that allows to define transmission probabilities for specific age groups. The age group corresponds to the infectee, i.e., the individual who may become infected, not the infecter.

Fields

age_groups::Vector{AgeGroup}: A vector of age groups.

age_transmission_rates::Vector{Real}: A vector of transmission rates corresponding to each age group.

Example

The code below instantiates an AgeDependentTransmissionRate with specific age groups and transmission rates.

adtr = AgeDependentTransmissionRate(
    age_groups = ["0-9", "10-19", "20-64", "65-"],
    transmission_rates = [0.1, 0.2, 0.3, 0.4]
)

GEMS.ConstantTransmissionRate — Type

ConstantTransmissionRate <: TransmissionFunction

A TransmissionFunction type that uses a constant transmission rate.

Missing docstring.

Missing docstring for TransmissionFunction. Check Documenter's build log for details.

Constructors

Missing docstring.

Missing docstring for AgeDependentTransmissionRate(;transmission_rate, ageGroups, ageTransmissions, distribution). Check Documenter's build log for details.

Functions

Missing docstring.

Missing docstring for disease_progression!(::Individual, ::Pathogen, ::Int16). Check Documenter's build log for details.

Missing docstring.

Missing docstring for estimate_disease_progression(::DiseaseProgressionStrat, ::Individual). Check Documenter's build log for details.

GEMS.infect! — Function

infect!(infectee::Individual, tick::Int16, pathogen::Pathogen;
    sim::Union{Simulation, Nothing} = nothing,
    rng::AbstractRNG = Random.default_rng(),
    infecter_id::Int32 = Int32(-1), setting_id::Int32 = Int32(-1), lon::Float32 = NaN32,
    lat::Float32 = NaN32, setting_type::Char = '?', ags::Int32 = Int32(-1),
    source_infection_id::Int32 = DEFAULT_INFECTION_ID)

Infect infectee with the specified pathogen and calculate time to infectiousness and time to recovery. Optional arguments infecter_id. setting_id, and setting_type can be passed for logging. It's not required to calulate the infection. The infection can only be logged, if Simulation object is passed (as this object holds the logger).

Parameters

infectee::Individual: Individual to infect

tick::Int16: Infection tick

pathogen::Pathogen: Pathogen to infect the individual with

sim::Simulation: Simulation object (used to get logger and current tick)

sim::Union{Simulation, Nothing} = nothing (optional) = Simulation object (used to get logger)

infecter_id::Int32 = Int32(-1) (optional): Infecting individual

setting_id::Int32 = Int32(-1) (optional): ID of setting this infection happens in

lon::Float32 = NaN32 (optional): Longitude of the infection infection location (setting)

lat::Float32 = NaN32 (optional): Latitude of the infection infection location (setting)

setting_type::Char = '?' (optional): Setting type as char (e.g. "h" for Household)

ags::Int32 = Int32(-1) (optional): Amtlicher Gemeindeschlüssel (community identification number) of the region this infection happened in as Integer value

source_infection_id::Int32 = DEFAULT_INFECTION_ID (optional): Current infection ID of the infecting individual

Returns

Int32: New infection ID

GEMS.spread_infection! — Method

spread_infection!(setting::Setting, sim::Simulation, pathogen::Pathogen)

Spreads the infection of pathogen inside the provided setting. This will simulate the infection dynamics at the time tick(sim) inside setting within the context of the simulation sim. This will also update all settings, the individual is part of, if the infection is successful.

Parameters

setting::Setting: Setting in which the pathogen shall be spreaded

sim::Simulation: Simulation object

pathogen::Pathogen: Pathogen to spread

GEMS.transmission_probability — Function

transmission_probability(transFunc::AgeDependentTransmissionRate, infecter::Individual, infectee::Individual, setting::Setting, tick::Int16, rng::AbstractRNG)::Float64

Calculates the transmission probability based on the age of the infectee using age-dependent transmission rates.

Parameters

transFunc::AgeDependentTransmissionRate: Transmission function struct

infecter::Individual: Infecting individual

infectee::Individual: Individual to infect

setting::Setting: Setting in which the infection happens

tick::Int16: Current tick

rng::AbstractRNG = Random.default_rng() (optional): RNG used for probability. Uses Random's default RNG as default.

Returns

Float64: Transmission probability p (0 <= p <= 1)

transmission_probability(transFunc::ConstantTransmissionRate, infecter::Individual, infected::Individual, setting::Setting, tick::Int16; rng::AbstractRNG = Random.default_rng())

Calculates the transmission probability for the ConstantTransmissionRate. Returns the transmission_rate for all individuals who have not been infected in the past. If the individual has already recovered, the function returns 0.0, assuming full indefinite natural immunity.

Parameters

transFunc::ConstantTransmissionRate: Transmission function struct

infecter::Individual: Infecting individual

infectee::Individual: Individual to infect

setting::Setting: Setting in which the infection happens

tick::Int16: Current tick

rng::AbstractRNG = Random.default_rng() (optional): RNG used for probability. Uses Random's default RNG as default.

Returns

Float64: Transmission probability p (0 <= p <= 1)

transmission_probability(transFunc::TransmissionFunction, infecter::Individual, infected::Individual, setting::Setting, tick::Int16; rng::AbstractRNG = Random.default_rng())

General function for TransmissionFunction struct. Should be overwritten for newly created structs, as it only serves to catch undefined transmission_probability functions.

GEMS.try_to_infect! — Function

try_to_infect!(infctr::Individual, infctd::Individual, sim::Simulation, pathogen::Pathogen, setting::Setting;
    source_infection_id::Int32 = DEFAULT_INFECTION_ID)

Tries to infect the infctd with the given pathogen transmitted by infctrat time tick(sim) with sim being the simulation. Success depends on whether the agent is alive, not already infected an whether an infection event was sampled using the provided distribution or probability. Returns true if infection was successful.

Parameters

infctr::Individual: Infecting individual

infctd::Individual: Individual to infect

sim::Simulation: Simulation object

pathogen::Pathogen: Pathogen to infect the individual with

setting::Setting: Setting this infection happens in

source_infection_id::Int32 = DEFAULT_INFECTION_ID (optional): Current infection ID of the infecting individual

Returns

Bool: True if infection was successful, false otherwise

GEMS.update_individual! — Method

update_individual!(indiv::Individual, tick::Int16, sim::Simulation)

Update the individual disease progression, handle its recovery and log its possible death. If the individual is not infected, this function will just return.

Parameters

indiv::Individual: Individual to update

tick::Int16: Current tick

sim::Simulation: Simulation object