Logger

Overview Structs

GEMS.EventLogger

GEMS.InfectionLogger

GEMS.Logger

GEMS.TickLogger

Overview Functions

Base.length

GEMS.lognow

GEMS.ticks

Structs

GEMS.Logger — Type

Supertype for all Loggers

Missing docstring.

Missing docstring for CustomLogger. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DeathLogger. Check Documenter's build log for details.

GEMS.EventLogger — Type

Supertype for all Loggers, which are logging certain events

GEMS.InfectionLogger — Type

InfectionLogger <: EventLogger

A logging structure specifically for infections. An infection event is given by all entries of the field-vectors at a given index. Data is thread-local to prevent lock contention.

Missing docstring.

Missing docstring for PoolTestLogger. Check Documenter's build log for details.

Missing docstring.

Missing docstring for QuarantineLogger. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SeroprevalenceLogger. Check Documenter's build log for details.

Missing docstring.

Missing docstring for StateLogger. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TestLogger. Check Documenter's build log for details.

GEMS.TickLogger — Type

Supertype for all Loggers, which are logging per tick

Functions

Missing docstring.

Missing docstring for dataframe(::CustomLogger). Check Documenter's build log for details.

Missing docstring.

Missing docstring for dataframe(::DeathLogger). Check Documenter's build log for details.

Missing docstring.

Missing docstring for dataframe(::InfectionLogger). Check Documenter's build log for details.

Missing docstring.

Missing docstring for dataframe(::PoolTestLogger). Check Documenter's build log for details.

Missing docstring.

Missing docstring for dataframe(::QuarantineLogger). Check Documenter's build log for details.

Missing docstring.

Missing docstring for dataframe(::SeroprevalenceLogger). Check Documenter's build log for details.

Missing docstring.

Missing docstring for dataframe(::StateLogger). Check Documenter's build log for details.

Missing docstring.

Missing docstring for dataframe(::TestLogger). Check Documenter's build log for details.

Missing docstring.

Missing docstring for dataframe(::VaccinationLogger). Check Documenter's build log for details.

Missing docstring.

Missing docstring for duplicate(::CustomLogger). Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_infections_between(::InfectionLogger, ::Int32, ::Int16, ::Int16). Check Documenter's build log for details.

Base.length — Function

length(eq::EventQueue)

Returns the number of Events in the EventQueue.

Missing docstring.

Missing docstring for log!. Check Documenter's build log for details.

GEMS.lognow — Function

lognow()

Returns the current time in HH:MM:SS format for logging purposes.

Missing docstring.

Missing docstring for save(::DeathLogger, ::AbstractString). Check Documenter's build log for details.

Missing docstring.

Missing docstring for save(::InfectionLogger, ::AbstractString). Check Documenter's build log for details.

Missing docstring.

Missing docstring for save(::PoolTestLogger, ::AbstractString). Check Documenter's build log for details.

Missing docstring.

Missing docstring for save(::TestLogger, ::AbstractString). Check Documenter's build log for details.

Missing docstring.

Missing docstring for save_JLD2. Check Documenter's build log for details.

GEMS.ticks — Method

ticks(logger::InfectionLogger)