Pathogens

Overview Structs

GEMS.Pathogen

Overview Functions

Structs

GEMS.Pathogen — Type

Pathogen <: Parameter

A type that holds all relevant information about a pathogen.

Fields

id::Int8: Unique identifier for the pathogen.

name::String: Name of the pathogen.

progressions::OrderedDict{DataType, ProgressionCategory}: An ordered dictionary that maps progression category types to their instances.

progression_assignment::ProgressionAssignmentFunction: A function that assigns a progression category to an individual. Must be a subtype of ProgressionAssignmentFunction.

transmission_function::TransmissionFunction: A function that calculates the transmission probability of the pathogen. Must be a subtype of TransmissionFunction.

Example

# Define disease progressions
dp_a = Asymptomatic(
    exposure_to_infectiousness = Poisson(2),
    symptom_onset_to_recovery = Poisson(5)
)
dp_s = Symptomatic(
    exposure_to_infectiousness = Poisson(3),
    infectiousness_to_symptom_onset = Poisson(1),
    symptom_onset_to_recovery = Poisson(7)
)

# Define progression assignment function
pa = RandomProgressionAssignment([Asymptomatic, Symptomatic])

# Define transmission function
tf = ConstantTransmissionRate(transmission_rate = 0.3)

# Create pathogen
pathogen = Pathogen(
    id = 1,
    name = "Covid19",
    progressions = [dp_a, dp_s],
    progression_assignment = pa,
    transmission_function = tf
)

Missing docstring.

Missing docstring for DiseaseProgressionStrat. Check Documenter's build log for details.

Functions

Missing docstring.

Missing docstring for critical_death_rate(::Pathogen). Check Documenter's build log for details.

Missing docstring.

Missing docstring for DiseaseProgressionStrat(). Check Documenter's build log for details.

Missing docstring.

Missing docstring for disease_progression_strat(::Pathogen). Check Documenter's build log for details.

Missing docstring.

Missing docstring for hospitalization_rate(::Pathogen). Check Documenter's build log for details.

Missing docstring.

Missing docstring for icu_rate(::Pathogen). Check Documenter's build log for details.

Missing docstring.

Missing docstring for id(::Pathogen). Check Documenter's build log for details.

Missing docstring.

Missing docstring for infection_rate(::Pathogen). Check Documenter's build log for details.

Missing docstring.

Missing docstring for infectious_offset(::Pathogen). Check Documenter's build log for details.

Missing docstring.

Missing docstring for length_of_stay(::Pathogen). Check Documenter's build log for details.

Missing docstring.

Missing docstring for mild_death_rate(::Pathogen). Check Documenter's build log for details.

Missing docstring.

Missing docstring for name(::Pathogen). Check Documenter's build log for details.

Missing docstring.

Missing docstring for onset_of_severeness(::Pathogen). Check Documenter's build log for details.

Missing docstring.

Missing docstring for onset_of_symptoms(::Pathogen). Check Documenter's build log for details.

Missing docstring.

Missing docstring for sample_critical_death_rate(::Pathogen, ::Individual). Check Documenter's build log for details.

Missing docstring.

Missing docstring for sample_hospitalization_rate(::Pathogen, ::Individual). Check Documenter's build log for details.

Missing docstring.

Missing docstring for sample_icu_rate(::Pathogen, ::Individual). Check Documenter's build log for details.

Missing docstring.

Missing docstring for sample_infectious_offset(::Pathogen, ::Individual). Check Documenter's build log for details.

Missing docstring.

Missing docstring for sample_length_of_stay(::Pathogen, ::Individual). Check Documenter's build log for details.

Missing docstring.

Missing docstring for sample_mild_death_rate(::Pathogen, ::Individual). Check Documenter's build log for details.

Missing docstring.

Missing docstring for sample_onset_of_severeness(::Pathogen, ::Individual). Check Documenter's build log for details.

Missing docstring.

Missing docstring for sample_onset_of_symptoms(::Pathogen, ::Individual). Check Documenter's build log for details.

Missing docstring.

Missing docstring for sample_self_quarantine_rate(::Pathogen, ::Individual). Check Documenter's build log for details.

Missing docstring.

Missing docstring for sample_severe_death_rate(::Pathogen, ::Individual). Check Documenter's build log for details.

Missing docstring.

Missing docstring for sample_time_to_hospitalization(::Pathogen, ::Individual). Check Documenter's build log for details.

Missing docstring.

Missing docstring for sample_time_to_icu(::Pathogen, ::Individual). Check Documenter's build log for details.

Missing docstring.

Missing docstring for sample_time_to_recovery(::Pathogen, ::Individual). Check Documenter's build log for details.

Missing docstring.

Missing docstring for sample_ventilation_rate(::Pathogen, ::Individual). Check Documenter's build log for details.

Missing docstring.

Missing docstring for severe_death_rate(::Pathogen). Check Documenter's build log for details.

Missing docstring.

Missing docstring for time_to_hospitalization(::Pathogen). Check Documenter's build log for details.

Missing docstring.

Missing docstring for time_to_icu(::Pathogen). Check Documenter's build log for details.

Missing docstring.

Missing docstring for time_to_recovery(::Pathogen). Check Documenter's build log for details.

Missing docstring.

Missing docstring for transmission_function!(::Pathogen, ::TransmissionFunction). Check Documenter's build log for details.

Missing docstring.

Missing docstring for transmission_function(::Pathogen). Check Documenter's build log for details.

Missing docstring.

Missing docstring for ventilation_rate(::Pathogen). Check Documenter's build log for details.