Pathogens

Overview Structs

GEMS.Pathogen

Overview Functions

Structs

GEMS.Pathogen — Type

Pathogen <: Parameter

A type that holds all relevant information about a pathogen.

Fields

id::Int8: Unique identifier for the pathogen.

name::String: Name of the pathogen.

progressions::OrderedDict{DataType, ProgressionCategory}: An ordered dictionary that maps progression category types to their instances.

progression_assignment::ProgressionAssignmentFunction: A function that assigns a progression category to an individual. Must be a subtype of ProgressionAssignmentFunction.

transmission_function::TransmissionFunction: A function that calculates the transmission probability of the pathogen. Must be a subtype of TransmissionFunction.

Example

# Define disease progressions
dp_a = Asymptomatic(
    exposure_to_infectiousness = Poisson(2),
    symptom_onset_to_recovery = Poisson(5)
)
dp_s = Symptomatic(
    exposure_to_infectiousness = Poisson(3),
    infectiousness_to_symptom_onset = Poisson(1),
    symptom_onset_to_recovery = Poisson(7)
)

# Define progression assignment function
pa = RandomProgressionAssignment([Asymptomatic, Symptomatic])

# Define transmission function
tf = ConstantTransmissionRate(transmission_rate = 0.3)

# Create pathogen
pathogen = Pathogen(
    id = 1,
    name = "Covid19",
    progressions = [dp_a, dp_s],
    progression_assignment = pa,
    transmission_function = tf
)

Functions