Pathogens & Immunity

Overview Structs

GEMS.AgeBasedProgressionAssignment

GEMS.AgeDependentTransmissionRate

GEMS.Asymptomatic

GEMS.ConstantTransmissionRate

GEMS.Hospitalized

GEMS.Pathogen

GEMS.Severe

Overview Functions

Pathogen Struct

GEMS.Pathogen — Type

Pathogen <: Parameter

A type that holds all relevant information about a pathogen.

Fields

id::Int8: Unique identifier for the pathogen.

name::String: Name of the pathogen.

progressions::OrderedDict{DataType, ProgressionCategory}: An ordered dictionary that maps progression category types to their instances.

progression_assignment::ProgressionAssignmentFunction: A function that assigns a progression category to an individual. Must be a subtype of ProgressionAssignmentFunction.

transmission_function::TransmissionFunction: A function that calculates the transmission probability of the pathogen. Must be a subtype of TransmissionFunction.

Example

# Define disease progressions
dp_a = Asymptomatic(
    exposure_to_infectiousness = Poisson(2),
    symptom_onset_to_recovery = Poisson(5)
)
dp_s = Symptomatic(
    exposure_to_infectiousness = Poisson(3),
    infectiousness_to_symptom_onset = Poisson(1),
    symptom_onset_to_recovery = Poisson(7)
)

# Define progression assignment function
pa = RandomProgressionAssignment([Asymptomatic, Symptomatic])

# Define transmission function
tf = ConstantTransmissionRate(transmission_rate = 0.3)

# Create pathogen
pathogen = Pathogen(
    id = 1,
    name = "Covid19",
    progressions = [dp_a, dp_s],
    progression_assignment = pa,
    transmission_function = tf
)

Progression Categories

Structs

GEMS.Asymptomatic — Type

Asymptomatic <: ProgressionCategory

A disease progression category where individuals do not show symptoms but can spread the disease.

IMPORTANT: The infectiousness onset must be at least 1 tick after exposure to avoid issues with immediate transmission. Therefore, the calculation for infectiousness_onset includes a +1 offset. The provided distributions should account for this offset to ensure realistic timing. Providing, for example a Poisson(2) distribution would result in an average of 3 ticks from exposure to infectiousness onset (Poisson(2) + 1).

Disease events

exposure -> infectiousness_onset -> recovery.

Parameters

exposure_to_infectiousness_onset::Union{Distribution, Real}: Time from exposure to becoming infectious.

infectiousness_onset_to_recovery::Union{Distribution, Real}: Time from becoming infectious to recovery.

Example

The code below instantiates an Asymptomatic progression category with specific distributions for the time intervals.

dp = Asymptomatic(
    exposure_to_infectiousness_onset = Poisson(3),
    infectiousness_onset_to_recovery = Poisson(7)
)

GEMS.Hospitalized — Type

Hospitalized <: ProgressionCategory

A disease progression category for individuals who develop severe symptoms requiring hospitalization but not ICU care.

IMPORTANT: The infectiousness onset must be at least 1 tick after exposure to avoid issues with immediate transmission. Therefore, the calculation for infectiousness_onset includes a +1 offset. The provided distributions should account for this offset to ensure realistic timing. Providing, for example a Poisson(2) distribution would result in an average of 3 ticks from exposure to infectiousness onset (Poisson(2) + 1).

Disease events

exposure -> infectiousness_onset -> symptom_onset -> severeness_onset -> hospital_admission -> hospital_discharge -> severeness_offset -> recovery.

Parameters

exposure_to_infectiousness_onset::Union{Distribution, Real}: Time from exposure to becoming infectious.

infectiousness_onset_to_symptom_onset::Union{Distribution, Real}: Time from becoming infectious to symptom onset.

symptom_onset_to_severeness_onset::Union{Distribution, Real}: Time from symptom onset to severeness onset.

severeness_onset_to_hospital_admission::Union{Distribution, Real}: Time from severeness onset to hospital admission.

hospital_admission_to_hospital_discharge::Union{Distribution, Real}: Time from hospital admission to hospital discharge.

hospital_discharge_to_severeness_offset::Union{Distribution, Real}: Time from hospital discharge to severeness offset.

severeness_offset_to_recovery::Union{Distribution, Real}: Time from severeness offset to recovery.

Example

The code below instantiates a Hospitalized progression category with specific distributions for the time intervals.

dp = Hospitalized(
    exposure_to_infectiousness_onset = Poisson(3),
    infectiousness_onset_to_symptom_onset = Poisson(1),
    symptom_onset_to_severeness_onset = Poisson(2),
    severeness_onset_to_hospital_admission = Poisson(1),
    hospital_admission_to_hospital_discharge = Poisson(10),
    hospital_discharge_to_severeness_offset = Poisson(2),
    severeness_offset_to_recovery = Poisson(1)
)

GEMS.Severe — Type

Severe <: ProgressionCategory

A disease progression category for individuals who develop severe symptoms. They will stay home during the severe stage of their illness but do not require hospitalization.

IMPORTANT: The infectiousness onset must be at least 1 tick after exposure to avoid issues with immediate transmission. Therefore, the calculation for infectiousness_onset includes a +1 offset. The provided distributions should account for this offset to ensure realistic timing. Providing, for example a Poisson(2) distribution would result in an average of 3 ticks from exposure to infectiousness onset (Poisson(2) + 1).

Disease events

exposure -> infectiousness_onset -> symptom_onset -> severeness_onset -> severeness_offset -> recovery.

Parameters

exposure_to_infectiousness_onset::Union{Distribution, Real}: Time from exposure to becoming infectious.

infectiousness_onset_to_symptom_onset::Union{Distribution, Real}: Time from becoming infectious to symptom onset.

symptom_onset_to_severeness_onset::Union{Distribution, Real}: Time from symptom onset to severeness onset.

severeness_onset_to_severeness_offset::Union{Distribution, Real}: Time from severeness onset to severeness offset.

severeness_offset_to_recovery::Union{Distribution, Real}: Time from severeness offset to recovery.

Example

The code below instantiates a Severe progression category with specific distributions for the time intervals.

dp = Severe(
    exposure_to_infectiousness_onset = Poisson(3),
    infectiousness_onset_to_symptom_onset = Poisson(1),
    symptom_onset_to_severeness_onset = Poisson(2),
    severeness_onset_to_severeness_offset = Poisson(3),
    severeness_offset_to_recovery = Poisson(7)
)

Functions

Progression Assignment

GEMS.AgeBasedProgressionAssignment — Type

AgeBasedProgressionAssignment <: ProgressionAssignmentFunction

An assignment function that assigns disease progression categories based on age groups and a stratification matrix.

Fields

age_groups::Vector{AgeGroup}: A vector of age groups.

progression_categories::Vector{DataType}: A vector of progression categories.

stratification_matrix::Matrix{Float64}: A matrix where each column corresponds to an age group and each row corresponds to a progression category.

Parameters for constructor

age_groups::Vector{String}: A vector of age group strings in the format "-X", "X-Y", or "Y-".

progression_categories::Vector{String}: A vector of progression category type names as strings.

stratification_matrix::Vector{Vector{T}} where T <: Real: A matrix (as a vector of vectors) where each column corresponds to an age group and each row corresponds to a progression category. Each column must sum to 1 and contain no negative values.

Example

The code below instantiates an AgeBasedProgressionAssignment with specific age groups, progression categories, and a stratification matrix.

aba = AgeBasedProgressionAssignment(
    age_groups = ["-14", "15-65", "66-"],
    progression_categories = ["Asymptomatic", "Symptomatic", "Hospitalized", "Critical"],
    stratification_matrix = [[0.4, 0.45, 0.1, 0.05],
                             [0.2, 0.6, 0.15, 0.05],
                             [0.1, 0.4, 0.3, 0.2]]
)

Tramission Function

GEMS.AgeDependentTransmissionRate — Type

AgeDependentTransmissionRate <: TransmissionFunction

A TransmissionFunction type that allows to define transmission probabilities for specific age groups. The age group corresponds to the infectee, i.e., the individual who may become infected, not the infecter.

Fields

age_groups::Vector{AgeGroup}: A vector of age groups.

age_transmission_rates::Vector{Real}: A vector of transmission rates corresponding to each age group.

Example

The code below instantiates an AgeDependentTransmissionRate with specific age groups and transmission rates.

adtr = AgeDependentTransmissionRate(
    age_groups = ["0-9", "10-19", "20-64", "65-"],
    transmission_rates = [0.1, 0.2, 0.3, 0.4]
)

GEMS.ConstantTransmissionRate — Type

ConstantTransmissionRate <: TransmissionFunction

A TransmissionFunction type that uses a constant transmission rate.

Functions