Simulation

alles aus setup folder configuration des sim objects

Overview Structs

GEMS.InfectedFraction

GEMS.NoneInfected

GEMS.Simulation

GEMS.StartCondition

GEMS.StopCriterion

GEMS.TimesUp

Overview Functions

GEMS.add_hospitalization_trigger!

GEMS.add_strategy!

GEMS.add_symptom_trigger!

GEMS.add_testtype!

GEMS.add_tick_trigger!

GEMS.ags

GEMS.configfile

GEMS.customlogger

GEMS.customlogger!

GEMS.customlogs

GEMS.deathlogger

GEMS.deaths

GEMS.evaluate

GEMS.event_queue

GEMS.fire_custom_loggers!

GEMS.fraction

GEMS.hospitalization_triggers

GEMS.incidence

GEMS.increment!

GEMS.infections

GEMS.info

GEMS.interval

GEMS.label

GEMS.limit

GEMS.parameters

GEMS.pathogen

GEMS.pathogen!

GEMS.pooltestlogger

GEMS.pooltests

GEMS.population

GEMS.populationDF

GEMS.populationfile

GEMS.quarantinelogger

GEMS.quarantines

GEMS.region_info

GEMS.remove_empty_settings!

GEMS.reset!

GEMS.run!

GEMS.settings

GEMS.settings

GEMS.should_fire

GEMS.start_condition

GEMS.step!

GEMS.stepmod

GEMS.stop_criterion

GEMS.strategies

GEMS.symptom_triggers

GEMS.testlogger

GEMS.tests

GEMS.testtypes

GEMS.tick

GEMS.tick_triggers

GEMS.tickunit

Structs

GEMS.InfectedFraction — Type

InfectedFraction <: StartCondition

A StartCondition that specifies a fraction of infected individuals (drawn at random).

Fields

fraction::Float64: A fraction of the whole population that has to be infected

pathogen::Pathogen: The pathogen with which the fraction has to be infected

Example

condition = InfectedFraction(fraction=0.05) # 5% of the population infected at random
sim = Simulation(start_condition = condition)

GEMS.NoneInfected — Type

NoneInfected <: StopCriterion

A StopCriterion that stops the simulation once no individual is infected.

GEMS.Simulation — Type

Simulation

A struct for the management of a single run, holding all necessary informations.

Initialization

Simulation(; kwargs...)
Simulation(params::Dict)

You can initialize a simulation without any parameters, which will then use the default configuration file. Providing any additional keyword arguments will override the respective configuration file parameter. If you provide a custom config file, the parameters in the config file will be used as defaults and only the provided keyword arguments will override them.

Here's a list of all available parameters:

Parameter	Type	Description

configfile	String	Path to the configuration file. If not provided, the default configuration will be used.

tickunit	Char	Time unit of one simulation step (tick). Must be one of 'h' (hours), 'd' (days), or 'w' (weeks).

start_date	Date	Start date of the simulation.

end_date	Date	End date of the simulation.

label	String	Label used for plot visualizations and aggregating simulations into batches.

population	String or Population	Path to a population file, a population identifier (e.g., 'DE'), or a Population object.

pop_size	Int	Size of the population to be created. Will be ignored if a population is provided.

avg_household_size	Float	Average household size for the population to be created. Will be ignored if a population is provided.

avg_office_size	Float	Average office size for the population to be created. Will be ignored if a population is provided.

avg_school_size	Float	Average school size for the population to be created. Will be ignored if a population is provided.

global_setting	Bool	Flag indicating whether to use the global setting.

settingsfile	String	Path to a settings file.

household_contacts	ContactSamplingMethod or Float	Method for sampling household contacts or a fixed value that will be regarded as the expected value of a Poisson distribution.

office_contacts	ContactSamplingMethod or Float	Method for sampling office contacts or a fixed value that will be regarded as the expected value of a Poisson distribution.

department_contacts	ContactSamplingMethod or Float	Method for sampling department contacts or a fixed value that will be regarded as the expected value of a Poisson distribution.

workplace_contacts	ContactSamplingMethod or Float	Method for sampling workplace contacts or a fixed value that will be regarded as the expected value of a Poisson distribution.

workplace_site_contacts	ContactSamplingMethod or Float	Method for sampling workplace site contacts or a fixed value that will be regarded as the expected value of a Poisson distribution.

school_class_contacts	ContactSamplingMethod or Float	Method for sampling school class contacts or a fixed value that will be regarded as the expected value of a Poisson distribution.

school_year_contacts	ContactSamplingMethod or Float	Method for sampling school year contacts or a fixed value that will be regarded as the expected value of a Poisson distribution.

school_contacts	ContactSamplingMethod or Float	Method for sampling school contacts or a fixed value that will be regarded as the expected value of a Poisson distribution.

school_complex_contacts	ContactSamplingMethod or Float	Method for sampling school complex contacts or a fixed value that will be regarded as the expected value of a Poisson distribution.

municipality_contacts	ContactSamplingMethod or Float	Method for sampling municipality contacts or a fixed value that will be regarded as the expected value of a Poisson distribution.

global_setting_contacts	ContactSamplingMethod or Float	Method for sampling global setting contacts or a fixed value that will be regarded as the expected value of a Poisson distribution. Requires global_setting to be true.

start_condition	StartCondition	A StartCondition object defining the initial situation of the simulation.

infected_fraction	Float	Fraction of the population to be initially infected. Will be ignored if a start_condition is provided.

stop_criterion	StopCriterion	A StopCriterion object defining the termination condition of the simulation.

pathogen	Pathogen	A Pathogen object defining the pathogen to be simulated.

transmission_function	TransmissionFunction	A TransmissionFunction object defining the transmission dynamics of the pathogen. Will be ignored if a pathogen is provided.

transmission_rate	Float	A fixed transmission rate that will be used to create a ConstantTransmissionRate transmission function. Will be ignored if a pathogen or transmission_function is provided.

stepmod	Function	A single-argument function that runs custom code on the simulation object in each tick.

Examples

# Initialize a simulation with default configuration
sim = Simulation()

# Initialize a simulation with a custom configuration file
sim = Simulation(configfile="path/to/configfile.toml")

# Initialize a simulation with custom parameters
sim = Simulation(
    tickunit='d',
    label="My Simulation",
    avg_household_size=5,
)

# Initialize a simulation with a predefined population model
sim = Simulation(population="SH") # Schleswig-Holstein, Germany

# Initialize a simulation with a custom population file
sim = Simulation(population="path/to/populationfile.csv")

# Initialize a simulation with a custom start condition
sim = Simulation(start_condition=PatientZero()) # starts with a single infected individual

# Initialize a simulation with a custom transmission rate
sim = Simulation(transmission_rate=0.1) # sets a constant per-contact transmission rate (chance) of 0.1

# Initialize a simulation with a parameter dictionary
params = Dict(
    :tickunit => 'd',
    :label => "My Simulation",
    :avg_household_size => 5,
)
sim = Simulation(params)

Internal Simulation Struct Fields

Data Sources

configfile::String: Path to config file

General

tick::Int16: Current tick/timestep

tickunit::Char: Time unit of one simulation step (tick)

startdate::Date: Start date of the simulation

enddate::Date: End date of the simulation

start_condition::StartCondition: Starting condition that sets the initial situation

stop_criterion::StopCriterion: Criterion that terminates a simulation run

label::String: Label for plot visualizations

Model

population::Population: Container to hold all present individuals

settings::SettingsContainer: All settings present in the simulation

pathogen::Pathogen: The pathogen of which infections are simulated

Logger

infectionlogger::InfectionLogger: A logger tracking all infections

deathlogger::DeathLogger: A logger specifically for the deaths of individuals

testlogger::TestLogger: A logger tracking all individual tests

pooltestlogger::PoolTestLogger: A logger tracking all pool tests

seroprevalencelogger::SeroprevalenceLogger: A logger tracking all seroprevalence tests

quarantinelogger::QuarantineLogger: A tracking cumulative quarantines per tick

customlogger::CustomLogger: A logger running custom methods on the Simulation object in each tick

Interventions

symptom_triggers::Vector{ITrigger}: List of all SymptomTriggers

tick_triggers::Vector{TickTrigger}: List of all TickTriggers

hospitalization_triggers::Vector{ITrigger}: List of all HospitalizationTriggers

event_queue::EventQueue: Event Queue to apply intervention measures

strategies::Vector{Strategy}: List of all registered intervention strategies

testtypes::Vector{AbstractTestType}: List of all TestTypes (e.g. Antigen- or PCR-Test)

Runtime Modifiers

stepmod::Function: Single-argment function that runs custom code on the simulation object in each tick

RNG

seed::Int64: Seed used to initialize the main RNG

rngs::Vector{AbstractRNG}: RNG instances for each thread

GEMS.StartCondition — Type

supertype for all start conditions

GEMS.StopCriterion — Type

supertype for all stop criteria

GEMS.TimesUp — Type

TimesUp <: StopCriterion

A StopCriterion that specifies a time limit.

Fields

limit::Int16: A time limit. When reached, the simulation should be terminated.

Constructors

Missing docstring.

Missing docstring for Simulation(; simargs...). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Simulation(::Dict{Symbol, <:Any}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Simulation(::String; simargs...). Check Documenter's build log for details.

Functions

GEMS.add_hospitalization_trigger! — Function

add_hospitalization_trigger!(simulation, trigger)

Adds a HospitalizationTrigger to the simulation.

GEMS.add_strategy! — Function

add_strategy!(simulation, strategy)

Adds an intervention Strategy to the simulation object. A strategy must be added to the simulation object to make it appear in the report. In order to execute a strategy during the simulation run, you must define a Trigger and link this strategy. Just adding it here will not execute the strategy.

GEMS.add_symptom_trigger! — Function

add_symptom_trigger!(simulation, trigger)

Adds a SymptomTrigger to the simulation.

GEMS.add_testtype! — Function

add_testtype!(simulation, testtype)

Adds a test type to the simulation.

GEMS.add_tick_trigger! — Function

add_tick_trigger!(simulation, trigger)

Adds a TickTrigger to the simulation.

GEMS.ags — Method

ags(patientzeros::PatientZeros)::Vector{Int64}

Returns the vector of ags where intial seeds should be planted.

GEMS.configfile — Method

configfile(simulation)

Returns configfile that was used to initialize simulation.

GEMS.customlogger! — Method

customlogger!(simulation, customlogger)

Sets the Simulation's CustomLogger.

GEMS.customlogger — Method

customlogger(simulation)

Returns the CustomLogger of the simulation.

GEMS.customlogs — Function

customlogs(simulation::Simulation)

Calls the dataframe() function on the internal simulation's CustomLogger.

GEMS.deaths — Method

deaths(simulation::Simulation)

Calls the dataframe() function on the internal simulation's DeathLogger.

GEMS.deathlogger — Method

deathlogger(simulation)

Returns the DeathLogger of the simulation.

GEMS.evaluate — Function

evaluate(simulation::Simulation, criterion::StopCriterion)

Evaluates whether the specified stop criterion is met for the simulation model. Return True if criterion was met. This is an 'abstract' function that must be implemented for concrete criterion types.

evaluate(simulation, criterion)

Returns true if none of the individuals are infected.

evaluate(simulation::Simulation, criterion::TimesUp)

Returns true if specified termination tick has been met.

GEMS.event_queue — Method

event_queue(simulation)

Returns the simulation's intervention event queue.

GEMS.fire_custom_loggers! — Method

fire_custom_loggers!(sim::Simulation)

Executes all custom functions that are stored in the CustomLogger on the Simulation object and stores them in the internal dataframe.

GEMS.fraction — Method

fraction(infectedFraction::InfectedFraction)

Returns fraction of individuals that shall be infected at the beginning using the InfectedFraction start condition.

GEMS.hospitalization_triggers — Method

hospitalization_triggers(simulation)

Returns the list of HospitalizationTriggers registered in the simulation.

GEMS.incidence — Function

incidence(simulation::Simulation, pathogen::Pathogen, base_size::Int = 100_000, duration::Int16 = Int16(7))

Returns the incidence at a particular pathogen and a point in time (current simulation tick). The duration defines a time-span for which the incidence is measured (default: 7 ticks). The base_size provides the population size reference (default: 100_000 individuals)

Parameters

simulation::Simulation: Simulation object

pathogen::Pathogen: Pathogen for which the incidence shall be calculated

base_size::Int = 100_000 (optional): Reference popuation size for incidence calculation

duration::Int16 = Int16(7) (optional): Reference duration (in ticks) for the incidence calculation

Returns

Float64: Incidence

incidence(simulation::Simulation, base_size::Int = 100_000, duration::Int16 = Int16(7))

Returns the incidence at a particular point in time (current simulation tick). NOTE: This will only work for the single-pathogen version. The duration defines a time-span for which the incidence is measured (default: 7 ticks). The base_size provides the population size reference (default: 100_000 individuals)

Parameters

simulation::Simulation: Simulation object

base_size::Int = 100_000 (optional): Reference popuation size for incidence calculation

duration::Int16 = Int16(7) (optional): Reference duration (in ticks) for the incidence calculation

Returns

Float64: Incidence

GEMS.increment! — Method

increment!(simulation)

Increments the current simulation's tick counter by 1.

GEMS.infections — Method

infections(simulation::Simulation)

Calls the dataframe() function on the internal simulation's InfectionLogger.

GEMS.info — Method

info(sim::Simulation)

Summary output for Simulation object configuration.

GEMS.interval — Function

interval(trigger::TickTrigger)

Returns the interval associated with a TickTrigger.

GEMS.label — Method

label(simulation::Simulation)

Returns simulation object's string label.

GEMS.limit — Function

limit(timesUp::TimesUp)

Returns time limit of a timesUp stop criterion.

GEMS.parameters — Function

parameters(d::Distribution)

Returns a dictionary containing the string() of the distribution, the mean and the std.

parameters(v::Vaccine)::Dict

Returns a dictionary containing the parameters of the vaccine.

GEMS.pathogen! — Method

pathogen!(simulation, pathogen)

Sets the pathogen of the simulation.

GEMS.pathogen — Function

pathogen(simulation)

Returns the pathogen of the simulation.

pathogen(infectedFraction::InfectedFraction)

Returns pathogen used to infect individuals at the beginning using the InfectedFraction start condition.

pathogen(patientzero::PatientZero)

Returns pathogen used to infect individuals at the beginning in this start condition.

pathogen(patientzeros::PatientZeros)

Returns pathogen used to infect individuals at the beginning in this start condition.

pathogen(regionalseeds::RegionalSeeds)

Returns pathogen used to infect individuals at the beginning using the RegionalSeeds start condition.

pathogen(tt::TestType)

Returns the TestType's associated pathogen.

pathogen(tt::SeroprevalenceTestType)

Returns the SeroprevalenceTestType's associated pathogen.

GEMS.pooltests — Function

pooltests(simulation::Simulation)

Calls the dataframe() function on the internal simulation's PoolTestLogger.

GEMS.pooltestlogger — Function

pooltestlogger(simulation)

Returns the PoolTestLogger of the simulation.

GEMS.population — Method

population(simulation)

Returns the population associated with the simulation run.

GEMS.populationDF — Method

populationDF(simulation::Simulation)

Calls the dataframe() function on the simulation's Population object.

GEMS.populationfile — Method

populationfile(simulation)

Returns populationfile that was used to initialize simulation.

GEMS.quarantinelogger — Method

quarantinelogger(simulation)

Returns the QuarantineLogger of the simulation.

GEMS.quarantines — Function

quarantines(simulation::Simulation)

Calls the dataframe() function on the internal simulation's QuarantineLogger.

GEMS.remove_empty_settings! — Method

remove_empty_settings!(sim::Simulation)

Removes all settings that have no individuals associated with them. In the end update all ids to ensure that they are still correspond to the entry in the settings container.

GEMS.reset! — Method

reset!(simulation)

Resets the current simulation's tick counter to 0.

GEMS.region_info — Method

region_info(sim::Simulation)

Returns a DataFrame containing information about the Municipalitys in the model with the following columns:

Name	Type	Description

ags	AGS	Amtlicher Gemeindeschlüssel (Community Identification Code)

pop_size	Int64	Number of individuals in that municipality

area	Float64	Area size of this municipality in km²

Note: This function will download the Germany shapefile, if it's not available locally, and return missing values for pop_size and area if the download cannot be completed.

GEMS.run! — Method

run!(simulation::Simulation; with_progressbar::Bool = true)

Takes and initializes Simulation object and calls the stepping function (step!) until the stop criterion is met.

Returns

Simulation: Simulation object

GEMS.settings — Method

settings(simulation::Simulation)

Returns a dictionary containing all settings, separated by setting type (key).

GEMS.settings — Method

settings(simulation::Simulation, settingtype::DataType)

Returns all settings of settingtype of the simulation.

GEMS.should_fire — Function

should_fire(trigger::TickTrigger, tick::Int16)

Evaluates whether a trigger should be fired at a given tick. Considers switch_tick and interval.

Returns

Bool: True, if the trigger should fire at this tick. False otherwise.

GEMS.start_condition — Function

start_condition(simulation)

Returns start condition associated with the simulation run.

start_condition(rd::ResultData)

Returns the StartCondition object the simulation was initialized with. Returns an empty dictionary if the data is not available in the input ResultData object.

GEMS.step! — Function

step!(simulation::Simulation)

Increments the simulation status by one tick and executes all events that shall be handled during this tick.

GEMS.stepmod — Function

stepmod(simulation::Simulation)

Returns the defined step mod.

GEMS.stop_criterion — Method

stop_criterion(simulation)

Returns stop criterion associated with the simulation run.

GEMS.strategies — Method

strategy(simulation)

Returns the intervention Strategys registered in the simulation.

GEMS.symptom_triggers — Method

symptom_triggers(simulation)

Returns the list of SymptomTriggers registered in the simulation.

GEMS.testlogger — Method

testlogger(simulation)

Returns the TestLogger of the simulation.

GEMS.tests — Method

tests(simulation::Simulation)

Calls the dataframe() function on the internal simulation's TestLogger.

GEMS.testtypes — Method

testtypes(simulation)

Returns the test types registered in the simulation.

GEMS.tick — Method

tick(simulation)

Returns current tick of the simulation run.

GEMS.tick_triggers — Method

tick_triggers(simulation)

Returns the list of TickTriggers registered in the simulation.

GEMS.tickunit — Method

tickunit(simulation)

Returns the unit of the ticks as a char like in date formats, i.e. 'd' means days, 'h' mean hours, etc.